% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_frequencies.R
\name{allele_frequencies}
\alias{allele_frequencies}
\title{Compute allele frequencies per markers and populations}
\usage{
allele_frequencies(data)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{stackr} \code{\link{tidy_genomic_data}}.}
}
\value{
A list with allele frequencies in a data frame in long and wide format,
and a matrix.
}
\description{
Compute allele frequencies per markers and populations.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

