% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignment_genodive}
\alias{assignment_genodive}
\title{Prepare and summarise the assignment results from GENODIVE}
\usage{
assignment_genodive(assignment.lmax, assignment.lhome, lmax.migrant.skip,
  lmax.number.migrant, lmax.skip, lhome.migrant.skip, lhome.number.migrant,
  lhome.skip, sites.levels, pop.labels, pop.levels, pop.id.start, pop.id.end,
  number.individuals, number.pop)
}
\arguments{
\item{assignment.lmax}{The file with likelihood ratio assignment results
from GENODIVE.}

\item{assignment.lhome}{The file with home likelihood assignment results
from GENODIVE.}

\item{lmax.migrant.skip}{The number of lines to skip before the migrant
info in the likelihood ratio assignment results. Usually = 11.}

\item{lmax.number.migrant}{The number of migrant detected
in the likelihood ratio assignment results.}

\item{lmax.skip}{The number of lines to skip before the individuals info
in the likelihood ratio assignment results.}

\item{lhome.migrant.skip}{The number of lines to skip before
the migrant info in the home likelihood assignment results. Usually = 11.}

\item{lhome.number.migrant}{The number of migrant detected
in the home likelihood assignment results.}

\item{lhome.skip}{The number of lines to skip before the individuals info
in the home likelihood assignment results.}

\item{sites.levels}{An optional character string with your sites names in
the same order as the pop levels.}

\item{pop.labels}{The pop label to identify your sampling sites.}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{number.individuals}{The number of individuals analysed.}

\item{number.pop}{The number of populations analysed.}
}
\description{
Import assignment results from GENODIVE.
Current version needs results from both statistics, the home likelihood
and likelihood ratio.
}

