% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_adjustment.R
\name{bias_adjustment}
\alias{bias_adjustment}
\title{Make a simple bias correction based on CRPS}
\usage{
bias_adjustment(y_true, y_pred, lambda = NULL)
}
\arguments{
\item{a}{data.frame with the following entries: 
\itemize{
  \item model (the model used to generate the correspondig predictions)
  \item geography (the region for which predictions are generated)
  \item date (the date of corresponding )
}}

\item{weights}{weights to give to the different models}

\item{type}{default is "crps"}
}
\value{
tibble with samples from the mixture model
}
\description{
given a vector of true values and a matrix of predicted values, 
finddata a bias correction term to correct for a mismatch between predictions
and true values

Caveat: not sure whether this is strictly better than just comparing the 
mean of the predictive samples against the true values
}
\examples{
 
S <- 50
T <- 100
y_true <- rnorm(T)
y_pred <- replicate(T, rnorm(S, mean = 1))

bias_adjustment(y_true, y_pred)

}
\references{
Missing
}
