% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/erase_genotypes.R
\name{blacklist_erase_genotype}
\alias{blacklist_erase_genotype}
\title{Blacklist erase genotypes.}
\usage{
blacklist_erase_genotype(tidy.vcf.file, allele.depth.threshold, gl.threshold,
  filename)
}
\arguments{
\item{tidy.vcf.file}{A data frame object or file (using ".tsv")
of class tidy VCF.}

\item{allele.depth.threshold}{Threshold number
of min depth of REF or ALT alleles.}

\item{gl.threshold}{Threshold number of mean genotype likelihood.}

\item{filename}{Name of the file written to the working directory.}
}
\value{
The function returns the individuals genotypes, by loci and
individuals, to erase.
}
\description{
This function creates a blacklist of loci and individual
genotypes useful for the function erase_genotype.
}
\details{
1. Keep the genotypes below the read depth and the
genotype likelihood threshold. 2. Keep the REF and ALT allele below the
coverage threshold. 3. Create a blacklist of genotypes to erase.
}

