% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stackr_workflow_dir.R
\name{build_stackr_workflow_dir}
\alias{build_stackr_workflow_dir}
\title{Build stacks workflow directories}
\usage{
build_stackr_workflow_dir(main.folder.name = "stacks_run", date = TRUE)
}
\arguments{
\item{main.folder.name}{Use this argument to customize the name of the main
folder for workflow. e.g. \code{main.folder.name = "stacks_lobster"}
Default: \code{main.folder.name = "stacks_run"}}

\item{date}{(logical) Should the current date be appended on the main folder
name? e.g. \code{stacks_lobster_20160630}
Default: \code{date = TRUE}}
}
\value{
A main folder containing different folders used during the stacks
workflow proposed here.
}
\description{
This prepares the different folder required to run the
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS} workflow
proposed in stackr.
}
\examples{
\dontrun{
# The simplest form of the function:
stackr::build_stackr_workflow_dir()
# that's it ! Now if you have your own main folder name:
build_stackr_workflow_dir(main.folder.name = "stacks_whiteshark", date = TRUE)
}
}
