% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_haplotypes.R
\name{change_pop_names}
\alias{change_pop_names}
\title{Transform into a factor the POP_ID column, change names and reorder the levels}
\usage{
change_pop_names(data, pop.levels = NULL, pop.labels = NULL)
}
\description{
Transform into a factor the POP_ID column, change names and reorder the levels
}
\keyword{internal}
