% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{combine_url}
\alias{combine_url}
\title{construct a query URL for a request, including checking special
operations}
\usage{
combine_url(base, id, special = NULL, special_ids = c(),
  special_no_ids = c())
}
\arguments{
\item{base}{base of query, such as "answers" or "questions"}

\item{id}{vector of IDs to search}

\item{special}{special parameter, which specifies the action (such as
retrieving an associated object with an ID)}

\item{special_ids}{vector of possible special parameters that require IDs}

\item{special_no_ids}{vector of possible special parameters that don't
require IDs}
}
\description{
construct a query URL for a request, including checking special
operations
}

