% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_array_for_stack_crps.R
\name{create_arrays}
\alias{create_arrays}
\title{Make array for weighting by CRPS}
\usage{
create_arrays(data)
}
\arguments{
\item{data}{a data.frame with the following entries: 
\itemize{
  \item model (the model used to generate the correspondig predictions)
  \item geography (the region for which predictions are generated)
  \item date (the date of corresponding )
}}
}
\value{
list with two elements, an array for the predictions and an array
for the true_values.
}
\description{
Takes a data.frame as input and ouputs a correctly sized array for the 
predictions and true values to be scores with the stan model
}
\examples{
 
data <- stackr::sample_prepared_data

create_arrays(data) 

}
\references{
Missing
}
