% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mixture_from_samples.R
\name{create_simple_mixture}
\alias{create_simple_mixture}
\title{Make simple mixture from samples}
\usage{
create_simple_mixture(data, weights = NULL, type = "crps")
}
\arguments{
\item{data}{a data.frame with the following entries: 
\itemize{
  \item model (the model used to generate the correspondig predictions)
  \item geography (the region for which predictions are generated)
  \item date (the date of corresponding )
}}

\item{weights}{weights to give to the different models}

\item{type}{default is "crps"}
}
\value{
tibble with samples from the mixture model
}
\description{
Takes a data.frame as input and ouputs a correctly sized array for the 
predictions and true values to be scores with the stan model
}
\examples{
 
data <- stackr::sample_prepared_data

weights <- stackr::stack_crps(data)

create_sampled_mixture(data, weights = weights)

}
\references{
Missing
}
