% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking_weights_crps.R
\name{crps_weights}
\alias{crps_weights}
\title{Obtain CRPS stacking weights}
\usage{
crps_weights(data, lambda = NULL, gamma = NULL, dirichlet_alpha = 1.001)
}
\arguments{
\item{data}{a data.frame with the following entries: 
\itemize{
  \item y_obs the true observed value
  \item y_pred a predicted value corresponding to the true value in y_obs
  \item model (the model used to generate the correspondig predictions)
  \item geography (the regions for which predictions are generated). If 
  geography is missing, it will be assumed there are no different regions.
  \item date (the date of the corresponding prediction / true value)
}}

\item{lambda}{weights given to timepoints. If \code{lamba} is \code{NULL}, 
the default gives more weight to recent time points with 
lambda[t] = 2 - (1 - t / T)^2. 
\code{lambda = "equal"} uses equal weights}

\item{gamma}{weights given to regions. If \code{gamma} is \code{NULL} the 
default is equal weights for the regions. Weights are mapped to regions
alphabetically, so make sure that the the weights correspond to the 
regions in alphabetical order.}

\item{dirichlet_alpha}{prior for the weights}
}
\value{
returns a vector with the model weights
}
\description{
given true values and predictive samples from different models, 
`crps_weights` returns the stacking weights which produce the ensemble
that minimises the Continuos Ranked Probability Score (CRPS).
}
\examples{

\dontrun{
splitdate <- as.Date("2020-04-01") 

traindata <- stackr::sample_prepared_data \%>\%
  dplyr::filter(date <= splitdate)

testdata <- stackr::sample_prepared_data \%>\%
  dplyr::filter(date > splitdate)

weights <- crps_weights(traindata)

test_mixture <- mixture_from_sample(testdata, 
                                    weights = weights)

rbind(testdata, 
      test_mixture) \%>\%
  group_by(model, forecast_date) \%>\%
  dplyr::mutate(crps = scoringutils::crps(unique(y_obs), 
                                          t(as.vector(y_pred)))) \%>\%
  group_by(model) \%>\%
  dplyr::summarise(crps = mean(crps))
}

}
\references{
Strictly Proper Scoring Rules, Prediction,and Estimation,
Tilmann Gneiting and Adrian E. Raftery, 2007, Journal of the American
Statistical Association, Volume 102, 2007 - Issue 477

Using Stacking to Average Bayesian Predictive Distributions, 
Yuling Yao , Aki Vehtari, Daniel Simpson, and Andrew Gelman, 2018, 
Bayesian Analysis 13, Number 3, pp. 917–1003
}
