% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pruning.R
\name{data_pruning}
\alias{data_pruning}
\title{Data set pruning}
\arguments{
\item{data}{Options include the VCF (1) or an haplotype files (2) created in STACKS 
(\code{data = "batch_1.vcf"} and \code{data = "batch_1.haplotypes.tsv"}, 
respectively) or a data frame (3) with tab separating the 
genotypes in columns (\code{data = "data.tsv"}). 
The 1st column is the \code{POP_ID}, 2nd colum 
the \code{INDIVIDUALS} and the remaining columns are the markers IDs
containing genotypes in the format: 3 digits per allele
and no space between alleles (e.g. 235240 : allele1 = 235 and allele2 = 240).
Missing genotypes are coded \code{0} or \code{000000}. 
Note that the \code{POP_ID} column can be any hierarchical grouping. 
See the argument \code{strata} for other means of controlling grouping.
The last option for data input is a PLINK file in 
\code{tped/tfam} format (e.g. \code{data =  "data.assignment.tped"}). 
The first 2 columns of the \code{tfam} file will be used for the 
\code{strata} argument below, unless a new one is provided. 
Columns 1, 3 and 4 of the \code{tped} are discarded. The remaining columns 
correspond to the genotype in the format \code{01/04} 
where \code{A = 01, C = 02, G = 03 and T = 04}. For \code{A/T} format, use 
PLINK or bash to convert.
Use VCFTOOLS \url{http://vcftools.sourceforge.net/} with \code{--plink-tped} 
to convert very large VCF file. For \code{.ped} file conversion to 
\code{.tped} use PLINK \url{http://pngu.mgh.harvard.edu/~purcell/plink/} 
with \code{--recode transpose}.}

\item{whitelist.markers}{(optional) A whitelist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter by chromosome and/or locus and/or by snp.
The whitelist is in the working directory (e.g. "whitelist.txt").
de novo CHROM column with 'un' need to be changed to 1. 
Default \code{NULL} for no whitelist of markers. In the VCF, the column ID is
the LOCUS identification.}

\item{monomorphic.out}{(optional) For PLINK file, should the monomorphic 
markers present in the dataset be filtered out ? 
Default: \code{monomorphic.out = TRUE}.}

\item{blacklist.genotype}{(optional) Useful to erase genotype with below 
average quality, e.g. genotype with more than 2 alleles in diploid individual likely corresponding to
sequencing errors or genotypes with poor genotype likelihood or coverage. 
The blacklist has a minimum of 2 column headers (markers and individuals). 
Markers can be 1 column (CHROM or LOCUS or POS), 
a combination of 2 (e.g. CHROM and POS or CHROM and LOCUS or LOCUS and POS) or 
all 3 (CHROM, LOCUS, POS) The markers columns must be designated: CHROM (character
or integer) and/or LOCUS (integer) and/or POS (integer). The id column designated
INDIVIDUALS (character) columns header. The blacklist must be in the working 
directory (e.g. "blacklist.genotype.txt"). For de novo VCF, CHROM column 
with 'un' need to be changed to 1. Default \code{NULL} for no blacklist of 
genotypes to erase.}

\item{snp.ld}{(optional) For VCF file only. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 3 options: \code{"random"} selection, \code{"first"} or
\code{"last"} SNP on the same short read/haplotype. 
Default: \code{snp.ld = NULL}.
Note that for other file type, use stackr package for haplotype file and 
create a whitelist, for plink and data frames, use PLINK linkage 
disequilibrium based SNP pruning option.}

\item{common.markers}{(optional) Logical. Default: \code{common.markers = TRUE}, 
will only keep markers in common (genotyped) between all the populations.}

\item{maf.thresholds}{(string, double, optional) String with 
local/populations and global/overall maf thresholds, respectively.
Default: \code{maf.thresholds = NULL}. 
e.g. \code{maf.thresholds = c(0.05, 0.1)} for a local maf threshold 
of 0.05 and a global threshold of 0.1. Available for VCF, PLINK and data frame 
files. Use stackr for haplotypes files.}

\item{maf.pop.num.threshold}{(integer, optional) When maf thresholds are used,
this argument is for the number of pop required to pass the maf thresholds
to keep the locus. Default: \code{maf.pop.num.threshold = 1}}

\item{maf.approach}{(character, optional). By \code{maf.approach = "SNP"} or 
by \code{maf.approach = "haplotype"}.
The function will consider the SNP or ID/LOCUS/haplotype/read MAF statistics 
to filter the markers.
Default is \code{maf.approach = "SNP"}. The \code{haplotype} approach is 
restricted to VCF file.}

\item{maf.operator}{(character, optional) \code{maf.operator = "AND"} or default \code{maf.operator = "OR"}.
When filtering over LOCUS or SNP, do you want the local \code{"AND"}
global MAF to pass the thresholds, or ... you want the local \code{"OR"}
global MAF to pass the thresholds, to keep the marker?}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the working directory
(e.g. "blacklist.txt").}

\item{filename}{(optional) The name of the file written to the directory.
Use the extension ".txt" at the end. Default \code{assignment_data.txt}.
The number of markers used will be appended to the name of the file.}

\item{pop.levels}{(optional) A character string with your populations ordered.}

\item{pop.labels}{(optional) A character string for your populations labels.
If you need to rename sampling sites in \code{pop.levels} or combined sites/pop
into a different names, here is the place.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample. Your individuals are identified 
in this form : SPECIES-POPULATION-MATURITY-YEAR-ID = CHI-QUE-ADU-2014-020,
then, \code{pop.id.start} = 5. If you didn't name your individuals
with the pop id in it, use the \code{strata} argument.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample. Your individuals are identified 
in this form : SPECIES-POPULATION-MATURITY-YEAR-ID = CHI-QUE-ADU-2014-020,
then, \code{pop.id.end} = 7. If you didn't name your individuals
with the pop id in it, use the \code{strata} argument.}

\item{strata}{(optional) A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}. Default: \code{strata = NULL}. With a 
data frame of genotypes the strata is the INDIVIDUALS and POP_ID columns, with
PLINK files, the \code{tfam} first 2 columns are used. 
If a \code{strata} file is specified, the strata file will have
precedence. The \code{STRATA} column can be any hierarchical grouping.}

\item{pop.select}{(string) Conduct the assignment analysis on a
selected list of populations. Default = \code{NULL} for no selection and keep
all population.
e.g. \code{pop.select = "QUE"} to select QUE population samples.
\code{pop.select = c("QUE", "ONT")} to select QUE and ONT population samples.}
}
\value{
Depending on arguments selected, several files are written to the your
working directory or \code{folder}
}
\description{
This function enable to prune the data set with different 
argument (see below) in order to be prep for subsequent analysis.
Various input files are offered. Individuals, populations and
markers can be filtered and/or selected in several ways using blacklist,
whitelist and other arguments.
}
\details{
You need to have either the \code{pop.id.start} and \code{pop.id.end}
or the \code{strata} argument, to identify your populations.
}
\examples{
\dontrun{
data_pruning(data = "plink.tped",
pop.levels = c("pop_1", "pop_2", "pop_3", "pop_4", "pop_5", "pop_6", "pop_7"),
monomorphic.out <- TRUE,
maf.thresholds <- c(0.05, 0.1),
maf.pop.num.threshold <- 1,
maf.approach <- "SNP",
maf.operator <- "OR",
common.markers <- TRUE,
pop.select <- c("pop_1", "pop_2", "pop_3")
) 
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.

Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.

Purcell S, Neale B, Todd-Brown K, Thomas L, Ferreira MAR, 
Bender D, et al. 
PLINK: a tool set for whole-genome association and population-based linkage 
analyses. 
American Journal of Human Genetics. 2007; 81: 559–575. doi:10.1086/519795
}

