% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_biallelic_markers.R
\name{detect_biallelic_markers}
\alias{detect_biallelic_markers}
\title{Detect biallelic data}
\usage{
detect_biallelic_markers(data, verbose = FALSE)
}
\arguments{
\item{data}{A tidy genomic data set in the working directory tidy formats.
\emph{How to get a tidy data frame ?}
Look for \pkg{stackr} \code{\link{tidy_genomic_data}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty 
during execution. 
Default: \code{verbose = FALSE}.}
}
\value{
A logical character string (TRUE/FALSE). That answer the question if
the data set is biallelic or not.
}
\description{
Detect if markers in tidy dataset are biallelic.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

