% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_genomic_format.R
\name{detect_genomic_format}
\alias{detect_genomic_format}
\title{Used internally in stackr to detect the file format}
\usage{
detect_genomic_format(data)
}
\arguments{
\item{data}{A genomic data set in the global environment}
}
\description{
Detect file format of genomic data set.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

