% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_mixed_genomes.R
\name{detect_mixed_genomes}
\alias{detect_mixed_genomes}
\title{Detect mixed genomes}
\usage{
detect_mixed_genomes(data, ind.heterozygosity.threshold = NULL)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{stackr} \code{\link{tidy_genomic_data}}.}

\item{ind.heterozygosity.threshold}{(double, optional)
Blacklist individuals based on observed heterozygosity of individuals 
(averaged across markers).
The value for the threshold is a proportion (0 to 1), where 1 is similar to the
default, \code{ind.heterozygosity.threshold = NULL}, and turn off the filter
(the function will only output the plots and table of heterozygosity).
Individuals with mean heterozygosity higher (>) than the threshold
will be blacklisted.}
}
\value{
The function returns inside the global environment a list with
4 objects:

\enumerate{
\item the individual's heterozigosity:\code{$individual.heterozigosity}
\item the blacklisted individuals based on the individual's heterozigosity:\code{$blacklist.ind.het}
\item the boxplot of individual heterozygosity:\code{$individual.heterozygosity.boxplot}
\item the manhattan plot of individual heterozygosity:\code{$individual.heterozygosity.manhattan.plot}
}


In the working directory, output is conditional to \code{interactive.filter} argument
}
\description{
Highlight outliers individual's heterozygosity for a quick
diagnostic of mixed samples.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\seealso{
\link{plot_density_distribution_het}
}

