% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MAF_diagnostic.R
\name{diagnostic_maf}
\alias{diagnostic_maf}
\title{MAF diagnostic}
\usage{
diagnostic_maf(data, group.rank, filename)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv")
of class sumstats or tidy VCF.}

\item{group.rank}{The number of group to class the MAF (Number)}

\item{filename}{Name of the file written to the working directory (optional).}
}
\description{
Minor Allele Frequency diagnostic, help choose a filter threshold.
}
\details{
Highly recommended to look at the distribution of MAF
\link{figure_density_distribution_maf}.
}
\seealso{
\link{filter_maf}
}

