% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{erase_genotypes}
\alias{erase_genotypes}
\title{Erase genotypes that didn't pass min coverage threshold
This function modify the 'batch_x.haplotypes.tsv' file
with a blacklist of loci and individuals.}
\usage{
erase_genotypes(haplotype.file, blacklist.genotypes, filename)
}
\arguments{
\item{haplotype.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{blacklist.genotypes}{A blacklist of loci and genotypes containing at least 2 columns header 'LOCUS' and 'SAMPLES'.}
}
\description{
Erase genotypes that didn't pass min coverage threshold
This function modify the 'batch_x.haplotypes.tsv' file
with a blacklist of loci and individuals.
}

