% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_catalog_sample.R
\name{extract_catalog_sql_ids}
\alias{extract_catalog_sql_ids}
\title{Extract sample SQL IDs from STACKS catalog file}
\usage{
extract_catalog_sql_ids(
  x = "06_ustacks_2_gstacks/catalog.tags.tsv.gz",
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{x}{(character, path). The path to the \code{catalog.tags.tsv.gz} file.
Default: \code{x = "catalog.tags.tsv.gz"}.}

\item{parallel.core}{(integer) Enable parallel execution with the number of threads.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
The function returns a tibble with the SQL IDs.
}
\description{
This function reads the output of
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/comp/cstacks.php}{cstacks} tags
file to generate a tibble of sample SQL ids that were used to generate the catalog.
}
\examples{
\dontrun{
ids <- stackr::extract_catalog_sql_id(x = "06_ustacks_2_gstacks/catalog.tags.tsv.gz")
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/cstacks.php}{cstacks}
\code{\link{run_cstacks}}
}
