% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignment.R
\name{figure_assignment}
\alias{figure_assignment}
\title{Cleveland dot plot figure of assignement results.}
\usage{
figure_assignment(assignment.summary)
}
\arguments{
\item{assignment.summary}{The assignment summary file created
with assignment_genodive function.}
}
\description{
GGPLOT2 Cleveland dot plot figure of assignment results.
}

