% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignment.R
\name{figure_assignment_plot}
\alias{figure_assignment_plot}
\title{Assignment plot of genotype likelihood.}
\usage{
figure_assignment_plot(data, l.skip, sites.levels, pop.labels, pop.levels,
  pop.id.start, pop.id.end, number.individuals, number.pop, POPA, POPB)
}
\arguments{
\item{data}{The assignment results from GENODIVE, home likelihood or likelihood ratio.}

\item{l.skip}{The number of lines to skip before the individuals info.}

\item{sites.levels}{An optional character string with your sites names in
the same order as the pop levels.}

\item{pop.labels}{The pop label to identify your sampling sites.}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{number.individuals}{The number of individuals analysed.}

\item{number.pop}{The number of populations analysed.}

\item{POPA}{First population to compare.}

\item{POPB}{Second population to compare (with A).}
}
\description{
under construction
}

