% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignment.R
\name{figure_assignment_stacked_bar}
\alias{figure_assignment_stacked_bar}
\title{Classic stacked bar plot figure of assignement results.}
\usage{
figure_assignment_stacked_bar(assignment.summary)
}
\arguments{
\item{assignment.summary}{The assignment summary file created
with assignment_genodive function.}
}
\description{
GGPLOT2 stacked bar plot figure of assignment results.
}

