% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{figure_box_plot_coverage}
\alias{figure_box_plot_coverage}
\title{Figure box plot of coverage summary statistics.}
\usage{
figure_box_plot_coverage(data)
}
\arguments{
\item{data}{Coverage summary file.}
}
\description{
Create box plots of coverage summary statistics.
Use the coverage summary file created with coverage_summary function.
}

