% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{figure_box_plot_genotype_likelihood}
\alias{figure_box_plot_genotype_likelihood}
\title{Figure box plot of genotype likelihood summary statistics}
\usage{
figure_box_plot_genotype_likelihood(data)
}
\arguments{
\item{data}{genotype likelihood summary file.}
}
\description{
Create box plots of genotype likelihood summary statistics.
Use the genotype likelihood summary file created
with genotype_likelihood_summary function.
}

