% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{figure_density_distribution_maf}
\alias{figure_density_distribution_maf}
\title{Figure density distribution of minor allele frequency (MAF)
summary statistics.}
\usage{
figure_density_distribution_maf(data, maf.group, aes.colour, adjust.bin,
  x.title)
}
\arguments{
\item{data}{sumstats or tidy vcf files.}

\item{maf.group}{The GGPLOT2 aes (e.g. aes(x = FREQ_ALT, na.rm = F)).}

\item{aes.colour}{GGPLOT2 aesthetics colour,
e.g. aes(y = ..scaled.., color = GROUP).}

\item{adjust.bin}{Adjust GGPLOT2 bin size (0 to 1).}

\item{x.title}{Title of the x-axis.}
}
\description{
Create density distribution of MAF summary statistics.
}

