% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{figure_snp_number_loci}
\alias{figure_snp_number_loci}
\title{Figure of the distribution of SNP per locus before and after filters.}
\usage{
figure_snp_number_loci(before.filter.data, after.filter.data)
}
\arguments{
\item{before.filter.data}{Data set before filter.}

\item{after.filter.data}{Data set after filter.}
}
\description{
Figure of the distribution of SNP per locus before and after filters.
}

