% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genotype_likelihood_test.R
\name{filter_genotype_likelihood_interactive}
\alias{filter_genotype_likelihood_interactive}
\title{Genotype likelihood interactive filter}
\arguments{
\item{tidy.vcf}{A tidy vcf object or file (using ".tsv"), 
created with read_stacks_vcf.}

\item{pop.levels}{(optional) A character string with your populations ordered.
See example below.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to 
be interactive. With default: \code{interactive.filter == TRUE}, the user is 
asked to see figures of distribution before making decisions for filtering 
with the genotype likelihood.}

\item{approach}{Character. By \code{approach = "SNP"} or 
by \code{approach = "haplotype"}. 
The function will consider the SNP or haplotype GL statistics to filter the marker. 
Default: \code{approach = "haplotype"}.}

\item{gl.ind.threshold}{(Integer, optional if interactive session) 
Threshold number of individual's genotype likelihood.}

\item{gl.mean.threshold}{(Integer, optional if interactive session) 
Threshold number of mean genotype likelihood.}

\item{gl.min.threshold}{(Integer, optional if interactive session) 
Threshold number of min genotype likelihood.}

\item{gl.diff.threshold}{(Integer, optional if interactive session)
Threshold number of diff genotype likelihood,
the difference between the max and min GL over a loci/read/haplotype.}

\item{pop.threshold}{(optional if interactive session)
A threshold number: proportion, percentage
or fixed number e.g. 0.50, 50 or 5.}

\item{percent}{(logical, optional if interactive session). 
Is the pop.threshold argument a percentage? TRUE/FALSE.
This argument is necessary to distinguish percentage from integer 
for population threshold, (e.g. 5 percent or 5 populations).}

\item{filename}{(optional) Name of the filtered data set, 
written to the working directory.}
}
\value{
With \code{interactive.filter = FALSE}, the function returns a 
filtered tidy vcf data frame inside the global environment. 
In the working directory 5 files are written:
1. filters_parameters.tsv is updated or created.
The filter parameters, values and returned results 
(markers numbers or blacklist genotypes) are inside that file. 
2. blacklist.genotypes.gl.tsv, 3. blacklist.markers.gl.tsv, 
4. whitelist.markers.gl.tsv and 
5. vcf.tidy.paralogs.id.gl.tsv (the same filtered tidy vcf data frame).

With \code{interactive.filter = TRUE}, a list with 20 objects is created. 
The information range from summary tables to plots. The objects names are 
found by using \code{names(list.name)}. The object can be isolated in separate
object outside the list by following the example below.
}
\description{
Filter genotypes and markers basedgenotype likelihood
using a tidy VCF file.
}
\details{
There is 4 steps in the interactive version:  

1. gl_individuals_populations: the user is asked to inspect 
the genotype likelihood at the individuals and populations levels"), 

2. blacklist_genotypes: the option is given to blacklist individual genotypes
based on low quality GL, 

3. gl_markers: the user is asked to inspecting the genotype likelihood 
at the marker level and 

4. blacklist_markers: the option is given to blacklist markers genotypes 
based on low quality GL found at the marker level.

Using the haplotype approach: The summary statistics are averaged
and nested: SNP -> individuals -> population -> loci. e.g. the mean GL is the average
genotype likelihood for all individuals of pop x for loci x.
The gl.diff.threshold is the difference between the max and min GL found for 
a locus. This argument is given to help your spot big difference in GL for 
a loci. Markers with small differences have more stability in the estimate.
}
\examples{
\dontrun{
If I have 10 populations, the 3 examples below will give the same output results:

beluga.vcf.tidy.gl <- filter_genotype_likelihood_interactive(
tidy.vcf = beluga.vcf.tidy, 
interactive.filter = TRUE,
approach = "haplotype", 
gl.ind.threshold = 0,
gl.mean.threshold = 10, 
gl.min.threshold = 0, 
gl.diff.threshold = 200, 
pop.threshold = 50, 
percent = TRUE)

beluga.vcf.tidy.gl <- filter_genotype_likelihood_interactive(
tidy.vcf = beluga.vcf.tidy, 
interactive.filter = FALSE,
approach = "haplotype",
gl.ind.threshold = 0,
gl.mean.threshold = 10, 
gl.min.threshold = 0, 
gl.diff.threshold = 200, 
pop.threshold = 0.5, 
percent = FALSE)

beluga.vcf.tidy.gl <- filter_genotype_likelihood_interactive(
tidy.vcf = beluga.vcf.tidy, 
approach = "haplotype", 
gl.ind.threshold = 0,
gl.mean.threshold = 10, 
gl.min.threshold = 0, 
gl.diff.threshold = 100, 
pop.threshold = 5, 
percent = FALSE)

If interactive.filter = TRUE, a list is created and to view the filtered tidy vcf:
tidy.data <- beluga.vcf.tidy.gl$tidy.vcf.filtered.gl

Inside the same list, to isolate the blacklist.genotypes:
bg <- beluga.vcf.tidy.gl$blacklist.genotypes
}
}

