% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_het.R
\name{filter_het}
\alias{filter_het}
\title{Heterozygosity filter}
\usage{
filter_het(data, approach = "haplotype", het.threshold, het.diff.threshold,
  pop.threshold, percent, filename)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv")
of class sumstats or a tidy VCF summarised.}

\item{approach}{Character. By \code{"SNP"} or by \code{"haplotype"}. 
The function will consider the SNP or haplotype MAF statistics to filter the marker. 
Default by \code{"haplotype"}.}

\item{het.threshold}{Number (0 - 0.5)}

\item{het.diff.threshold}{Number (0 - 1).}

\item{pop.threshold}{Fixed number of pop required to keep the locus.}

\item{percent}{Is the threshold a percentage ? TRUE or FALSE.}

\item{filename}{Name of the file written to the working directory (optional).}
}
\description{
Observed heterozygosity filter from a sumstats prepared file
or a tidy VCF file.
}
\seealso{
\link{plot_density_distribution_het}
}

