% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_blacklist.R
\name{blacklist_loci_filter}
\alias{blacklist_loci_filter}
\title{filter_loci_blacklist}
\usage{
blacklist_loci_filter(data, blacklist, type)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv").}

\item{blacklist}{The blacklist is a data frame or file (using ".tsv")
with one column named `LOCUS`.}

\item{type}{A description of the type of blacklist filtering used
(e.g. "paralog filter").}
}
\description{
This function use a blacklist of loci to filter
a dataset (sumstats, tidy vcf, haplotype file).
}

