% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_quick.R
\name{filter_quick}
\alias{filter_quick}
\title{Combination of all filters with a tidy vcf}
\usage{
filter_quick(vcf.tidy, pop.id.start, pop.id.end, pop.levels,
  read.depth.threshold, allele.depth.threshold, allele.imbalance.threshold,
  allele.min.depth.threshold, read.depth.max.threshold, gl.mean.threshold,
  gl.min.threshold, gl.diff.threshold, gl.pop.threshold, gl.percent,
  ind.threshold, threshold.fixed, pop.threshold, pop.percent,
  local.maf.threshold, global.maf.threshold, maf.pop.threshold, het.threshold,
  het.diff.threshold, het.pop.threshold, het.percent, fis.min.threshold,
  fis.max.threshold, fis.diff.threshold, fis.pop.threshold, fis.percent,
  max.snp.number, filename)
}
\arguments{
\item{vcf.tidy}{Working on it.}

\item{pop.id.start}{Working on it.}

\item{pop.id.end}{Working on it.}

\item{pop.levels}{Working on it.}

\item{read.depth.threshold}{Working on it.}

\item{allele.depth.threshold}{Working on it.}

\item{allele.imbalance.threshold}{Working on it.}

\item{allele.min.depth.threshold}{Working on it.}

\item{read.depth.max.threshold}{Working on it.}

\item{gl.mean.threshold}{Working on it.}

\item{gl.min.threshold}{Working on it.}

\item{gl.diff.threshold}{Working on it.}

\item{gl.pop.threshold}{Working on it.}

\item{gl.percent}{Working on it.}

\item{ind.threshold}{Working on it.}

\item{threshold.fixed}{Working on it.}

\item{pop.threshold}{Working on it.}

\item{pop.percent}{Working on it.}

\item{local.maf.threshold}{Working on it.}

\item{global.maf.threshold}{Working on it.}

\item{maf.pop.threshold}{Working on it.}

\item{het.threshold}{Working on it.}

\item{het.diff.threshold}{Working on it.}

\item{het.pop.threshold}{Working on it.}

\item{het.percent}{Working on it.}

\item{fis.min.threshold}{Working on it.}

\item{fis.max.threshold}{Working on it.}

\item{fis.diff.threshold}{Working on it.}

\item{fis.pop.threshold}{Working on it.}

\item{fis.percent}{Working on it.}

\item{max.snp.number}{Working on it.}

\item{filename}{Name of the whitelist file written to the directory.}
}
\value{
A list with 3 dataframes (whitelist, blacklist.genotypes, tidy.vcf.filtered)
}
\description{
under construction...
}

