% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicate_id.R
\name{find_duplicate_id}
\alias{find_duplicate_id}
\title{Compute pairwise genome similarity to highligh potential duplicate 
individuals}
\arguments{
\item{data}{A file in the working directory or object in the global environment 
in wide or long (tidy) formats. \code{\link{tidy_genomic_data}} can
transform numerous genomic data format in tidy data frames.}

\item{parallel.core}{(optional) The number of core for parallel computation 
of pairwise Fst. 
If not selected \code{detectCores()-1} is used as default.}
}
\value{
A list with 2 objects: 
\code{$duplicate.info.summary}: a data frame with the summary of the similarity analysis.
\code{$similarity.plot}: boxplot of the similarity analysis.
}
\description{
The function computes pairwise genome similarity 
to highligh potential duplicate individuals.
}
\examples{
\dontrun{
dup <- find_duplicate_id(data = "wombat_tidy.tsv", parallel.core = 8)

# To view the data frame:
dup.df <- dup$duplicate.info.summary

# To view the boxplot:
bp <- dup$similarity.plot

# Based on the look of the distribution using the boxplot, I choose to filter 
# the resuls with a threshold of 90\%:
dup.filtered <-  \%>\% filter(PROP_IDENTICAL > 0.90)

# Get the list of duplicates id
dup.list.names <- data.frame(INDIVIDUALS = unique(c(dup.filtered$ID1, dup.filtered$ID2)))
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

