% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genotypes_summary.R
\name{genotypes_summary}
\alias{genotypes_summary}
\title{Summary of \code{batch_x.genotypes.txt} and
\code{batch_1.markers.tsv} files.}
\usage{
genotypes_summary(genotypes, markers, filter.monomorphic = TRUE,
  filter.missing.band = TRUE, filter.mean.log.likelihood = FALSE,
  B = FALSE, filter.GOF = FALSE, filter.GOF.p.value = FALSE,
  ind.genotyped = 1, filename = FALSE)
}
\arguments{
\item{genotypes}{The \code{batch_x.genotypes.txt} created by STACKS genotypes
module.}

\item{markers}{The \code{batch_1.markers.tsv} created by STACKS genotypes
module.}

\item{filter.monomorphic}{(optional) Should monomorphic loci be filtered out.
Default = \code{TRUE}.}

\item{filter.missing.band}{(optional) Should loci with missing
band be filtered out. Default = \code{TRUE}.}

\item{filter.mean.log.likelihood}{(optional) Should a mean log likelihood
filter be applied to the loci.
Default = \code{FALSE}. e.g. filter.mean.log.likelihood = -10.}

\item{B}{(optional) Should the the segregation distortion p-value
be computed with a Monte Carlo test with \code{B} replicates.
For more details, see package \code{stats} function
 \code{chisq.test}. Default = \code{FALSE}.}

\item{filter.GOF}{(optional) Filer the goodness-of-fit for segregation
distortion. Default = \code{FALSE}.}

\item{filter.GOF.p.value}{(optional) Filer the goodness-of-fit p-value of
GOF segregation distortion. Default = \code{FALSE}.}

\item{ind.genotyped}{Number. Filter the number of individual
progeny required to keep the marker.}

\item{filename}{(optional) The name of the file written in the directory.}
}
\value{
The function returns a list with the
genotypes summary \code{$geno.sum}, joinmap markers \code{$joinmap.sum}
and onemap markers \code{$onemap.sum} summary (use $ to access each
components).
}
\description{
Useful summary of \code{batch_x.genotypes.txt} and
 \code{batch_1.markers.tsv} files produced by STACKS genotypes module use
 for linkage mapping.
}
\details{
SIGNIFICANCE results pvalue (goodness-of-fit pvalue):
<= 0.0001 = ****, <= 0.001 & > 0.0001 = ***,
<= 0.01 & > 0.001 = **, <= 0.05 & > 0.01 = *.
}
\examples{
\dontrun{
work in progress...
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}

