% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo2colony.R
\name{haplo2colony}
\alias{haplo2colony}
\title{From a \code{batch_x.haplotypes.tsv} STACKS file
export \code{COLONY} input files}
\usage{
haplo2colony(haplotypes.file, blacklist.id = NULL, whitelist.loci = NULL,
  sample.markers = "all", pop.id.start, pop.id.end, pop.select = "all",
  allele.freq = FALSE, inbreeding = 0, mating.sys.males = 0,
  mating.sys.females = 0, clone = 0, run.length = 2, analysis = 1,
  allelic.dropout = 0, error.rate = 0.02, print.all.colony.opt = FALSE,
  imputations = FALSE, imputations.group = "populations", num.tree = 100,
  iteration.rf = 10, split.number = 100, verbose = FALSE,
  parallel.core = 2, filename = "colony")
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the directory
 (e.g. "blacklist.txt").}

\item{whitelist.loci}{(optional) A whitelist of loci and 
a column header 'LOCUS'.
The whitelist is in the directory (e.g. "whitelist.txt").}

\item{sample.markers}{(number) Should COLONY output contain a random 
subsample of your markers. 
Default= \code{"all"}.
e.g. \code{sample.markers = 500} to use only 500 randomly chosen markers.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.select}{(string) Should the COLONY output be on all populations
or a select group. Default = \code{"all"}. 
e.g. \code{pop.select = "QUE"} to select QUE population sample. 
\code{pop.select = c("QUE", "ONT")} to select QUE and ONT population sample.}

\item{allele.freq}{(optional string) Population string chosen to estimate
the allele frequency for each locus. Default = \code{FALSE}. 
e.g. \code{allele.freq = "QUE"} or \code{allele.freq = c("QUE", "ONT")} 
or \code{allele.freq = "overall"}.}

\item{inbreeding}{(boolean) 0/1 no inbreeding/inbreeding. Default = \code{0}}

\item{mating.sys.males}{(boolean) Mating system in males.
0/1 polygyny/monogyny. Default = \code{0}.}

\item{mating.sys.females}{(boolean) Mating system in females.
0/1 polygyny/monogyny. Default = \code{0}.}

\item{clone}{(boolean) Should clones and duplicated individuals be inferred.
0/1, yes/no. Default = \code{0}.}

\item{run.length}{(integer) Length of run. 1 (short), 2 (medium), 3 (long),
4 (very long). Default = \code{2}.}

\item{analysis}{(integer) Analysis method. 
0 (Pairwise-Likelihood Score), 1 (Full Likelihood),
2 (combined Pairwise-Likelihood Score and Full Likelihood).
Default = \code{1}.}

\item{allelic.dropout}{Locus allelic dropout rate. Default = \code{0} See Colony manual.}

\item{error.rate}{Locus error rate. Default = \code{0.02}, see COLONY manual.}

\item{print.all.colony.opt}{(logical) Should all COLONY options be printed in the file.
This require manual curation, for the file to work directly with COLONY. 
Default = \code{FALSE}.}

\item{imputations}{Should a map-independent imputations of markers be
computed. Available choices are: (1) \code{FALSE} for no imputation.
(2) \code{"max"} to use the most frequent category for imputations.
 (3) \code{"rf"} using Random Forest algorithm. Default = \code{FALSE}.}

\item{imputations.group}{\code{"global"} or \code{"populations"}.
Should the imputations be computed globally or by populations. If you choose
global, turn the verbose to \code{TRUE}, to see progress.
Default = \code{"populations"}.}

\item{num.tree}{The number of trees to grow in Random Forest. Default is 100.}

\item{iteration.rf}{The number of iterations of missing data algorithm 
in Random Forest. Default is 10.}

\item{split.number}{Non-negative integer value used to specify 
random splitting in Random Forest. Default is 100.}

\item{verbose}{Logical. Should trace output be enabled on each iteration 
in Random Forest ? Default is \code{FALSE}.}

\item{parallel.core}{(optional) The number of core for OpenMP 
shared-memory parallel programming of Random Forest imputations. 
For more info on how to install the
OpenMP version see \code{\link[randomForestSRC]{randomForestSRC-package}}.
If not selected \code{detectCores()-1} is used as default.}

\item{filename}{Name of the acronym for filenaming in the working directory.}
}
\value{
When no imputation is selected ....
When imputation is selected a list with 2 objects is returned
and accessed with \code{$no.imputation} or \code{$imputed}.
}
\description{
This function can first filter the haplotypes file 
with a whitelist of loci and a blacklist of individuals (optional). 
Then it will convert the file to the required \code{COLONY} input files.
Map-independent imputation using Random Forest or the most frequent category
is also available as an option.
}
\details{
It is highly recommended to read the user guide distributed with
COLONY to find out the details for input and output of the software. 
Not all options are provided here. 
However, all required options to run COLONY will be printed 
in the file written in your working directory. Change the values accordingly.
The imputations using Random Forest requires more time to compute 
and can take several minutes and hours depending on the size of 
the dataset and polymorphism of the species used. e.g. with a 
low polymorphic taxa, and a data set containing 30\% missing data, 
5 000 haplotypes loci and 500 individuals will require 15 min.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011) 
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences. 
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013) 
Stacks: an analysis tool set for population genomics. 
Molecular Ecology, 22, 3124-3140.

Ishwaran H. and Kogalur U.B. (2015). Random Forests for Survival,
 Regression and Classification (RF-SRC), R package version 1.6.1.

Ishwaran H. and Kogalur U.B. (2007). Random survival forests
for R. R News 7(2), 25-31.

Ishwaran H., Kogalur U.B., Blackstone E.H. and Lauer M.S. (2008).
Random survival forests. Ann. Appl. Statist. 2(3), 841--860.

Jones OR, Wang J (2010) COLONY: a program for parentage and 
sibship inference from multilocus genotype data. 
Molecular Ecology Resources, 10, 551–555.

Wang J (2012) Computationally Efficient Sibship and 
Parentage Assignment from Multilocus Marker Data. Genetics, 191, 183–194.
}
\seealso{
COLONY is available on Jinliang Wang www 
\url{http://www.zsl.org/science/software/colony}
\code{randomForestSRC} is available on 
CRAN \url{http://cran.r-project.org/web/packages/randomForestSRC/} 
and github \url{https://github.com/ehrlinger/randomForestSRC}
}

