% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo2gsi_sim.R
\name{haplo2gsi_sim}
\alias{haplo2gsi_sim}
\title{Use the batch_x.haplotypes.tsv file to write a gsi_sim input file.}
\usage{
haplo2gsi_sim(haplotypes.file, whitelist.loci = NULL, erase.paralogs = NULL,
  sample.markers = "all", iterations = 10, blacklist.id = NULL,
  gsi_sim.filename = "gsi_sim_data.txt", pop.levels, pop.labels, pop.id.start,
  pop.id.end, baseline, mixture, imputations = FALSE,
  imputations.group = "populations", num.tree = 100, iteration.rf = 10,
  split.number = 100, verbose = FALSE, parallel.core = 2)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{whitelist.loci}{(optional) A whitelist of loci and 
a column header 'LOCUS'.
The whitelist is in the directory (e.g. "whitelist.txt").}

\item{erase.paralogs}{(logical) \code{FALSE} Loci with more than 2 alleles 
(paralogs and/or sequencing errors) will be removed. \code{TRUE} genotypes 
with more than 2 alleles (paralogs and/or sequencing errors) will be erased.}

\item{sample.markers}{(number) Should the output contain a random 
subsample of your markers. Default= \code{"all"}.
e.g. \code{sample.markers = 500} to use only 500 randomly chosen markers.}

\item{iterations}{The number of iterations for marker resampling.
Default is 10.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the directory
 (e.g. "blacklist.txt").}

\item{gsi_sim.filename}{The name of the file written to the directory.
Use the extension ".txt" at the end. Default \code{gsi_sim_data.txt}.
The number of markers used will be appended to the name of the file.}

\item{pop.levels}{A character string with your populations ordered.}

\item{pop.labels}{An optional character string with new populations names.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{baseline}{(optional) A character string with your baseline id. 
From the \code{pop.id.start} and \code{pop.id.end} you isolate 
the baseline and mixture group. Here you need to give the id for 
baseline e.g. \code{c("QUE-ADU", "ONT-ADU")}.}

\item{mixture}{(optional) But required if bseline was selected. A character 
string with your mixture id. e.g. \code{c("QUE-JUV", "ONT-JUV")}.}

\item{imputations}{Should a map-independent imputations of markers be
computed. Available choices are: (1) \code{FALSE} for no imputation.
(2) \code{"max"} to use the most frequent category for imputations.
 (3) \code{"rf"} using Random Forest algorithm. Default = \code{FALSE}.}

\item{imputations.group}{\code{"global"} or \code{"populations"}.
Should the imputations be computed globally or by populations. If you choose
global, turn the verbose to \code{TRUE}, to see progress.
Default = \code{"populations"}.}

\item{num.tree}{The number of trees to grow in Random Forest. Default is 100.}

\item{iteration.rf}{The number of iterations of missing data algorithm 
in Random Forest. Default is 10.}

\item{split.number}{Non-negative integer value used to specify 
random splitting in Random Forest. Default is 100.}

\item{verbose}{Logical. Should trace output be enabled on each iteration 
in Random Forest ? Default is \code{FALSE}.}

\item{parallel.core}{(optional) The number of core for OpenMP shared-memory parallel
programming of Random Forest imputations. For more info on how to install the
OpenMP version see \code{\link[randomForestSRC]{randomForestSRC-package}}.
If not selected \code{detectCores()-1} is used as default.}
}
\value{
When no imputation is selected a gsi_sim file is saved to the 
working directory. When imputation is selected 2 gsi_sim files are saved to
the working directory.
}
\description{
\code{gsi_sim} is a tool for doing and simulating genetic stock identification. 
The \code{haplo2gsi_sim} function can first filter the haplotypes file 
with a whitelist of loci and a blacklist of individuals. 
Map-independent imputation using Random Forest or the most frequent category
is also available as an option.
}
\details{
The imputations using Random Forest requires more time to compute 
and can take several
minutes and hours depending on the size of the dataset and polymorphism of
the species used. e.g. with a low polymorphic taxa, and a data set 
containing 30\% missing data, 5 000 haplotypes loci and 500 individuals 
will require 15 min.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Anderson, Eric C., Robin S. Waples, and Steven T. Kalinowski. (2008) 
An improved method for predicting the accuracy of genetic stock identification.
Canadian Journal of Fisheries and Aquatic Sciences 65, 7:1475-1486.

Anderson, E. C. (2010) Assessing the power of informative subsets of 
loci for population assignment: standard methods are upwardly biased.
Molecular ecology resources 10, 4:701-710.

Catchen JM, Amores A, Hohenlohe PA et al. (2011) 
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences. 
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013) 
Stacks: an analysis tool set for population genomics. 
Molecular Ecology, 22, 3124-3140.

Ishwaran H. and Kogalur U.B. (2015). Random Forests for Survival,
 Regression and Classification (RF-SRC), R package version 1.6.1.

Ishwaran H. and Kogalur U.B. (2007). Random survival forests
for R. R News 7(2), 25-31.

Ishwaran H., Kogalur U.B., Blackstone E.H. and Lauer M.S. (2008).
Random survival forests. Ann. Appl. Statist. 2(3), 841--860.
}

