% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haplo_imputation.R
\name{haplo_imputation}
\alias{haplo_imputation}
\title{Convert between batch_x.haplotypes.tsv and \code{gtypes} objects}
\usage{
haplo_imputation(haplotypes.file, whitelist.loci = NULL,
  blacklist.id = NULL, pop.levels, pop.id.start, pop.id.end)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{whitelist.loci}{(optional) A whitelist of loci and
a column header 'LOCUS'.
The whitelist is in the directory (e.g. "whitelist.txt").}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the directory
 (e.g. "blacklist.txt").}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}
}
\value{
a \code{gtypes} objects.
}
\description{
This function can first filter the haplotypes file
with a whitelist of loci
and a blacklist of individuals (optional). Then it will convert the file
to a strataG \code{gtypes} objects.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
strataG is available on CRAN and https://github.com/EricArcher/
}

