% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R, R/individual_filter.R
\name{individual_filter}
\alias{individual_filter}
\title{Individual filter}
\usage{
individual_filter(data, population.map, pop.id.start, pop.id.end, pop.levels,
  ind.threshold, threshold.fixed)

individual_filter(data, population.map, pop.id.start, pop.id.end, pop.levels,
  ind.threshold, threshold.fixed)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv")
of class sumstats or tidy VCF.}

\item{population.map}{The population map or individuals listed in one column.
No headers.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{ind.threshold}{The individual threshold, proportion, percentage or
number e.g. 0.70, 70, 15.}

\item{threshold.fixed}{Is the threshold fixed ? TRUE or FALSE.}

\item{data}{A data frame object or file (using ".tsv")
of class sumstats or tidy VCF.}

\item{population.map}{The population map or individuals listed in one column.
No headers.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{ind.threshold}{The individual threshold, proportion, percentage or
number e.g. 0.70, 70, 15.}

\item{threshold.fixed}{Is the threshold fixed ? TRUE or FALSE.}
}
\description{
Filter individuals number, percentage or fix threshold
in a sumstats prepared file or a tidy VCF file.

Filter individuals number, percentage or fix threshold
in a sumstats prepared file or a tidy VCF file.
}

