% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_common_markers.R
\name{keep_common_markers}
\alias{keep_common_markers}
\title{Keep markers in common between all populations}
\usage{
keep_common_markers(data, verbose = FALSE)
}
\arguments{
\item{data}{A tidy genomic data set in the working directory tidy formats.
\emph{How to get a tidy data frame ?}
Look for \pkg{stackr} \code{\link{tidy_genomic_data}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty 
during execution. 
Default: \code{verbose = FALSE}.}
}
\value{
The filtered input data set with markers found in all the populations
present in the data set.
}
\description{
Keep markers in common between all populations.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

