% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_sample_file}
\alias{list_sample_file}
\title{list_sample_file}
\usage{
list_sample_file(f, full.path = FALSE, recursive = FALSE, paired.end = FALSE)
}
\description{
List sample file in folder
}
\keyword{internal}
