% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopensemble-package.R
\docType{package}
\name{lopensemble-package}
\alias{lopensemble-package}
\alias{lopensemble}
\alias{Using}
\alias{Stacking}
\alias{to}
\alias{Average}
\alias{Bayesian}
\alias{Predictive}
\alias{Distributions,}
\alias{Yuling}
\alias{Yao}
\alias{,}
\alias{Aki}
\alias{Vehtari,}
\alias{Daniel}
\alias{Simpson,}
\alias{and}
\alias{Andrew}
\alias{Gelman,}
\alias{2018,}
\alias{Analysis}
\alias{13,}
\alias{Number}
\alias{3,}
\alias{pp.}
\alias{917–1003}
\alias{Strictly}
\alias{Proper}
\alias{Scoring}
\alias{Rules,}
\alias{Prediction,and}
\alias{Estimation,}
\alias{Tilmann}
\alias{Gneiting}
\alias{Adrian}
\alias{E.}
\alias{Raftery,}
\alias{2007,}
\alias{Journal}
\alias{of}
\alias{the}
\alias{American}
\alias{Statistical}
\alias{Association,}
\alias{Volume}
\alias{102,}
\alias{2007}
\alias{-}
\alias{Issue}
\alias{477}
\title{The 'lopensemble' package.}
\description{
The lopensemble package provides an easy way to stack models based on the
  Continuous Ranked Probability Score (CRPS) based on predictive samples.
  Given some training data with true observed values as well as predictions
  generated from different models, lopensemble finds the optimal weights to
  form an ensemble from these models. Using these weights, lopensemble can
  then provide samples from the optimal model mixture by drawing from the
  predictice samples of those models in the correct proportion. This gives a
  true mixture model solely based on predictive samples and is in this regard
  superior to other ensembling techniques like Bayesian Model Averaging.

Currently stacking using CRPS is supported. Weights are generated using the
  \link{crps_weights} function. With these weights and predictive samples,
  \link{mixture_from_samples} can be used to obtain predictive samples from
  the optimal mixture model.
}
\author{
\strong{Maintainer}: Nikos Bosse \email{nikosbosse@gmail.com} (\href{https://orcid.org/0000-0002-7750-5280}{ORCID})

Authors:
\itemize{
  \item Yuling Yao \email{yy2619@columbia.edu}
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID})
}

}
\keyword{internal}
