% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_tools.R
\name{make_tfam}
\alias{make_tfam}
\title{Write a PLINK tfam file from a population map.}
\usage{
make_tfam(population.map, pop.id.start, pop.id.end, pop.levels, filename)
}
\arguments{
\item{population.map}{The population map or individuals listed in one column. No headers.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.levels}{(optional) Character string with your populations ordered.}

\item{filename}{The name of the file written in the directory.}
}
\description{
Make a tfam file for PLINK. Useful in ADMIXTURE for bootstrapping.
}
\examples{
\dontrun{
make_tfam(
population.map = "population_map_turle.txt", 
pop.id.start = 1, 
pop.id.end = 3, 
pop.levels = c("GAP", "ISA", "CRU"),
filename="turtle.tfam")
}
}
\references{
Purcell S, Neale B, Todd-Brown K et al. (2007) 
PLINK: a tool set for whole-genome association and population-based linkage 
analyses. American Journal of Human Genetics, 81, 559–575.
}

