% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackr_parallel.R
\docType{data}
\name{PORT}
\alias{PORT}
\title{mclapply_progress_bar}
\format{An object of class \code{numeric} of length 1.}
\usage{
PORT
}
\arguments{
\item{X}{see \pkg{parallel} \code{\link{mclapply}}}

\item{FUN}{see \pkg{parallel} \code{\link{mclapply}}}

\item{...}{see \pkg{parallel} \code{\link{mclapply}}}

\item{mc.cores}{see \pkg{parallel} \code{\link{mclapply}}}
}
\description{
progress bar
}
\keyword{internal}

