% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tsv2bam.R
\name{merge_bam}
\alias{merge_bam}
\title{merge_bam}
\usage{
merge_bam(x, new.folder, cmd.path, output.folder, parallel.core)
}
\description{
Function that merge an iteration of BAM files
}
\keyword{internal}
