% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_fq}
\alias{merge_fq}
\title{merge_fq}
\usage{
merge_fq(
  x,
  output.dir = NULL,
  append.filename = "-R.fq",
  parallel.core = parallel::detectCores() - 1
)
}
\description{
Merge FQ files (single-end only)
}
\keyword{internal}
