% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tsv2bam.R
\name{merge_parallel}
\alias{merge_parallel}
\title{merge_parallel}
\usage{
merge_parallel(cmd.path, output.folder, parallel.core, batch.id)
}
\description{
Merge bam in parallel. This is the complete function to merge
in parallel BAM files
}
\keyword{internal}
