% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missing_info.R
\name{missing_genotype_filtered}
\alias{missing_genotype_filtered}
\title{Missing genotype AFTER filter function.
This function filters the 'batch_x.haplotypes.tsv' file
with a whitelist of loci.}
\usage{
missing_genotype_filtered(haplotype.file, whitelist, pop.levels, pop.id.start,
  pop.id.end, blacklist.genotypes)
}
\arguments{
\item{haplotype.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{whitelist}{A whitelist of loci containing a column header 'LOCUS'.}

\item{pop.levels}{A character string with your populations ordered.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}
}
\description{
Missing genotype AFTER filter function.
This function filters the 'batch_x.haplotypes.tsv' file
with a whitelist of loci.
}

