% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_visualization.R
\name{missing_visualization}
\alias{missing_visualization}
\title{Visualize missing genotypes in genomic data set.}
\usage{
missing_visualization(data, vcf.metadata = FALSE, strata = NULL,
  strata.select = "POP_ID", distance.method = "euclidean",
  ind.missing.geno.threshold = c(10, 20, 30, 40, 50, 60, 70),
  pop.levels = NULL, pop.labels = NULL, pop.select = NULL,
  blacklist.id = NULL, blacklist.genotype = NULL,
  whitelist.markers = NULL, monomorphic.out = TRUE, max.marker = NULL,
  snp.ld = NULL, common.markers = FALSE, filename = NULL)
}
\arguments{
\item{data}{11 options: VCF (SNPs or Haplotypes,
to make the vcf population ready, see details below),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop,
and a data frame in long/tidy or wide format.
\emph{See details} of \code{\link{tidy_genomic_data}}.}

\item{vcf.metadata}{(optional, logical or string) For VCF files only.
With \code{logical, TRUE/FALSE}, \code{vcf.metadata = FALSE}, only the genotype
information is kept (GT field). With \code{vcf.metadata = TRUE},
all the metadata contained in the \code{FORMAT} field will be kept in
the tidy data file. stackr is currently keeping/cleaning only these metadata:
\code{"DP", "AD", "GL", "PL", "GQ", "GOF", "NR", "NV"}.
If you need different one, submit a request.
With \code{string}, explicitely ask for a particular FORMAT field you want
to keep (except the GT field that is always imported).
e.g. \code{vcf.metadata = "PL"} or \code{vcf.metadata = c("DP", "GL")}.
Default: \code{vcf.metadata = FALSE}.}

\item{strata}{(optional/required) Required for VCF and haplotypes files,
optional for the other file formats supported.

The strata file is a tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}. With a
data frame of genotypes the strata is the INDIVIDUALS and POP_ID columns, with
PLINK files, the \code{tfam} first 2 columns are used.
If a \code{strata} file is specified, the strata file will have
precedence. The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[stackr]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks `population map file`, make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
Default: \code{strata = NULL}.}

\item{strata.select}{(optional, character) Use this argument to select the column
from the strata file to generate the PCoA-IBM plot. More than 1 column you
want to visualize, use a string of character 
e.g. \code{strata.select = c("POP_ID", "LANES", "SEQUENCER", "WATERSHED")} to test
4 grouping columns inside the \code{strata} file.
Default: \code{strata.select = "POP_ID"}}

\item{distance.method}{(character) The distance measure to be used. 
This must be one of "euclidean", "maximum", "manhattan", "canberra", 
"binary" or "minkowski". The function uses \code{\link[stats]{dist}}. 
Default: \code{distance.method = "euclidean"}}

\item{ind.missing.geno.threshold}{(string) Percentage of missing genotype 
allowed per individuals. 
Default:\code{ind.missing.geno.threshold = c(10,20,30,40,50,60,70)}}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.levels = c("NEW", "NEW", "ALB")}.
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated,
there is also the \code{strata} argument that can do the same thing,
see below.
White spaces in population names are replaced by underscore.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is an object in your
global environment or a file in the working directory
(e.g. "blacklist.txt").
Default: \code{blacklist.id = NULL}.}

\item{blacklist.genotype}{(optional) Useful to erase genotype with below
average quality, e.g. genotype with more than 2 alleles in diploid likely
sequencing errors or genotypes with poor genotype likelihood or coverage.
The blacklist has a minimum of 2 column headers (markers and individuals).
Markers can be 1 column (CHROM or LOCUS or POS),
a combination of 2 (e.g. CHROM and POS or CHROM and LOCUS or LOCUS and POS) or
all 3 (CHROM, LOCUS, POS). The markers columns must be designated: CHROM (character
or integer) and/or LOCUS (integer) and/or POS (integer). The id column designated
INDIVIDUALS (character) columns header. The blacklist must be in the working
directory (e.g. "blacklist.genotype.txt"). For de novo VCF, CHROM column
with 'un' need to be changed to 1.
Default: \code{blacklist.genotype = NULL} for no blacklist of
genotypes to erase.}

\item{whitelist.markers}{(optional) A whitelist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter by chromosome and/or locus and/or by snp.
The whitelist is in the working directory (e.g. "whitelist.txt").
de novo CHROM column with 'un' need to be changed to 1.
In the VCF, the column ID is the LOCUS identification.
Default \code{whitelist.markers = NULL} for no whitelist of markers.}

\item{monomorphic.out}{(optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{monomorphic.out = TRUE}.}

\item{max.marker}{(integer, optional) For large PLINK files,
useful to subsample marker number. e.g. if the data set
contains 200 000 markers and \code{max.marker = 10000}, 10000 markers are
subsampled randomly from the 200000 markers. If you need specific markers,
use \code{whitelist.markers} argument.
Default: \code{max.marker = NULL}.}

\item{snp.ld}{(optional) \strong{For VCF file only}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 3 options: \code{"random"} selection, \code{"first"} or
\code{"last"} SNP on the same short read/haplotype. For long distance linkage
disequilibrium pruning, see details below.
Default: \code{snp.ld = NULL}.}

\item{common.markers}{(optional) Logical. Default: \code{common.markers = TRUE},
will only keep markers in common (genotyped) between all the populations.}

\item{filename}{(optional) Name of the tidy data set, 
written to the working directory.}
}
\value{
A list is created with several objects: the tidy data, 
the principal coordinates 
with eigenvalues of the PCoA, the identity-by-missingness plot, several 
summary tables and plots of missing information
per individuals, populations and markers. Blacklisted id are also included. 
Whitelists of markers with different level of missingness are also created.
A heatmap showing the missing values in black and genotypes in grey provide a
general overview of the missing data.
}
\description{
Use this function to visualize pattern of missing data.
 \itemize{
   \item \strong{Imput file:} various file format are supported 
   (see \code{data} argument below).
   \item \strong{Filters:} genotypes, markers, individuals and populations can be 
  filtered and/or selected in several ways using blacklist,
  whitelist and other arguments.
   \item \strong{IBM-PCoA:} conduct identity-by-missingness analyses using 
   Principal Coordinates Analysis, PCoA (also called Multidimensional Scaling, MDS).
   \item \strong{FH measure vs missingness}: missingness at the individual level
   is contrasted against FH, a new measure of IBDg 
   (Keller et al., 2011; Kardos et al., 2015; Hedrick & Garcia-Dorado, 2016)
   FH is based on the excess in the observed number of homozygous
   genotypes within an individual relative to the mean number of homozygous
   genotypes expected under random mating.
   IBDg is a proxy measure of the realized proportion of the genome
   that is identical by descent.
   Within this function, we're using a modified version of the measure
   described in (Keller et al., 2011; Kardos et al., 2015).
   The new measure is population-wise and tailored for RADseq data
   (see \code{\link[stackr]{ibdg_fh}} for details).
   \item \strong{Figures and Tables:} figures and summary tables of 
   missing information at the marker, individual and population level are
   generated.
   \item \strong{Blacklist:} create blacklist of individuals based on 
   desired thresholds of missing genotypes.
   \item \strong{Tidy data:} if the filename argument is used, the 
   function also output the data in a tidy format.
}
}
\examples{
\dontrun{
Using a  VCF file, the simplest for of the function:
ibm.koala <- missing_visualization(
data = "batch_1.vcf", 
strata = "population.map.strata.tsv"
)
# To see what's inside the list
names(ibm.koala) 
# To view the heatmap:
ibm.koala$heatmap
# To view the IBM analysis plot:
ibm.koala$ibm_plot
}
}
\references{
Legendre, P. and Legendre, L. (1998) Numerical Ecology, 
2nd English edition. Amsterdam: Elsevier Science BV.

Keller MC, Visscher PM, Goddard ME (2011)
Quantification of inbreeding due to distant ancestors and its detection
 using dense single nucleotide polymorphism data. Genetics, 189, 237–249.

Kardos M, Luikart G, Allendorf FW (2015)
Measuring individual inbreeding in the age of genomics: marker-based
measures are better than pedigrees. Heredity, 115, 63–72.

Hedrick PW, Garcia-Dorado A. (2016)
Understanding Inbreeding Depression, Purging, and Genetic Rescue.
Trends in Ecology and Evolution. 2016;31: 940-952.
doi:10.1016/j.tree.2016.09.005
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
