% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions_raw.R
\name{mixture_ensemble_raw}
\alias{mixture_ensemble_raw}
\title{Raw version of the mixture model from sample function}
\usage{
mixture_ensemble_raw(
  individual_draws,
  weight,
  random_seed = 1,
  permutation = TRUE
)
}
\arguments{
\item{data}{a data.frame with the following entries: 
\itemize{
  \item model (the model used to generate the correspondig predictions)
  \item geography (the region for which predictions are generated)
  \item date (the date of corresponding )
}}

\item{weights}{weights to give to the different models}

\item{type}{default is "crps"}
}
\value{
tibble with samples from the mixture model
}
\description{
Takes a data.frame as input and ouputs a correctly sized array for the 
predictions and true values to be scores with the stan model
}
\examples{


}
\references{
Missing
}
