% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mixture_from_samples.R
\name{mixture_from_sample}
\alias{mixture_from_sample}
\title{Make mixture model from predictive samples}
\usage{
mixture_from_sample(data, weights = NULL)
}
\arguments{
\item{data}{a data.frame with the following entries: 
\itemize{
  \item y_obs the true observed value
  \item y_pred a predicted value corresponding to the true value in y_obs
  \item model (the model used to generate the correspondig predictions)
  \item geography (the regions for which predictions are generated). If 
  geography is missing, it will be assumed there are no different regions.
  \item date (the date of the corresponding prediction / true value)
}}

\item{weights}{stacking weights used to combine the original model to a 
mixture model}
}
\value{
data.frame with samples from the mixture model
}
\description{
The function takes a data.frame with predictive samples generated from 
different models as well as weights corresponding to these models as input. 
It then returns predictive samples from a mixture model
generated by stacking the original models using these weights.
}
\details{

}
\examples{

\dontrun{
data <- stackr::sample_prepared_data

weights <- stackr::crps_weights(data)

create_sampled_mixture(data, weights = weights)
}

}
\references{
Using Stacking to Average Bayesian Predictive Distributions, 
Yuling Yao , Aki Vehtari, Daniel Simpson, and Andrew Gelman, 2018, 
Bayesian Analysis 13, Number 3, pp. 917–1003
}
