% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mixture_from_samples.R
\name{mixture_from_samples}
\alias{mixture_from_samples}
\title{Make mixture model from predictive samples}
\usage{
mixture_from_samples(data, weights = NULL)
}
\arguments{
\item{data}{a data.frame with the following entries: 
\itemize{
  \item y_obs, the true observed values (optional)
  \item y_pred, predicted values corresponding to the true values in y_obs
  \item model, the name of the model used to generate the correspondig 
  predictions
  \item geography (optional), the regions for which predictions are 
  generated. If geography is missing, it will be assumed there are no 
  geographical differenes to take into account. Internally, regions will
  be ordered alphabetically 
  \item date (the date of the corresponding prediction / true value). Also
  works with numbers to indicate timesteps
}}

\item{weights}{stacking weights used to combine the original model to a 
mixture model}
}
\value{
data.frame with samples from the mixture model. The following 
columns are returned: 
\itemize{
  \item y_obs, the true observed values, if they were given as input
  \item y_pred, predicted values corresponding to the true values in y_obs
  \item model, the name of the model used to generate the correspondig 
  predictions
  \item geography (optional), the regions for which predictions are 
  generated. If geography is missing, it will be assumed there are no 
  geographical differenes to take into account. Internally, regions will
  be ordered alphabetically 
  \item date (the date of the corresponding prediction / true value). Also
  works with numbers to indicate timesteps
}
}
\description{
The function takes a data.frame with predictive samples generated from 
different models as well as weights corresponding to these models as input. 
It then returns predictive samples from a mixture model
generated by stacking the original models using these weights.
}
\examples{

\dontrun{
data <- stackr::example_data

weights <- stackr::crps_weights(data)

stackr::mixture_from_samples(data, weights = weights)
}

}
\references{
Using Stacking to Average Bayesian Predictive Distributions, 
Yuling Yao , Aki Vehtari, Daniel Simpson, and Andrew Gelman, 2018, 
Bayesian Analysis 13, Number 3, pp. 917–1003
}
