% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_sample.R
\name{normalize_samples}
\alias{normalize_samples}
\title{Normalize samples}
\usage{
normalize_samples(project.info = NULL, path.samples, sample.reads = 1e+06,
  number.samples = 3, random.seed = NULL,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{project.info}{(character, path, optional) When using the stackr pipeline,
a project info file is created. This file provides all the info and stats
generated by stacks and stackr.
If no \code{project.info} file is provided, the function will have to look
at the number of reads in the fastq files and this takes longer.
Default: \code{project.info = NULL}.}

\item{path.samples}{(character, path) Give the path to the samples to normalize.}

\item{sample.reads}{(integer) The number of reads to pick randomly.
Default: \code{sample.reads = 1000000}.}

\item{number.samples}{(interger) The number of samples to generate from the sample(s) to
normalize.
Default: \code{number.samples = 5}.}

\item{random.seed}{(integer, optional) For reproducibility, set an integer
that will be used inside function that requires randomness. With default,
a random number is generated and printed in the appropriate output.
Default: \code{random.seed = NULL}.}

\item{parallel.core}{(optional) The number of core for parallel
programming.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
For stacks specific output see \href{http://catchenlab.life.illinois.edu/stacks/comp/process_radtags.php}{process_radtags}.
}
\description{
Normalize samples fastq files before de novo assembly or alignment.
}
\details{
\strong{REPLICATES ?}

You have replicates? Awesome. stackr makes it easy to keep track of replicates.
Use the same name for individual replicates. They will have different barcodes,
and can potentially be on different lanes. No problem. stackr will combine
fastq file at the end, keeping original replicates intact. However, stackr
will be appending integers (e.g. STU-QUE-ADU-2017-0001-1, STU-QUE-ADU-2017-0001-2)
at the end of the names you've chosen). Combined replicates
will have -R at the end (e.g STU-QUE-ADU-2017-0001-R for the combination of the 2 replicates.)
}
\examples{
\dontrun{
# library(stackr)
source("http://bioconductor.org/biocLite.R")
biocLite("ShortRead")
# Using OpenMP threads
nthreads <- .Call(ShortRead:::.set_omp_threads, 1L)
on.exit(.Call(ShortRead:::.set_omp_threads, nthreads))
# using defaults:
normalize_samples(path.samples = "~/corals")

# customizing the function:
normalize_samples(
   project.info = "project.info.corals.tsv",
   path.samples = "~/corals",
   sample.reads = 2000000,
   number.samples = 5,
   random.seed = NULL,
   parallel.core = 16)
}
}
