% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{nucleotide_number_position}
\alias{nucleotide_number_position}
\title{Figure of the distribution of SNP nucleotide position alond the read.}
\usage{
nucleotide_number_position(before.filter.data, gl.blacklist, maf.blacklist,
  het.blacklist, fis.blacklist, all.filters.blacklist)
}
\arguments{
\item{before.filter.data}{Data set before filter.}

\item{gl.blacklist}{Blacklist to show the impact of the filter on marker
density distribution.}
}
\description{
Figure of the distribution of SNP nucleotide position alond the read.
}

