% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicate_genomes.R
\name{pairwise_genome_similarity}
\alias{pairwise_genome_similarity}
\title{Pairwise genome similarity}
\usage{
pairwise_genome_similarity(list.pair, id.pairwise = NULL, input.prep = NULL,
  ...)
}
\description{
genome method
}
\keyword{internal}
