% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{paralogs}
\alias{paralogs}
\title{Create a list of putative paralogs.}
\usage{
paralogs(haplotype.file, pop.id.start, pop.id.end, pop.levels)
}
\arguments{
\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}
}
\description{
Output putative paralogs loci.
}

