% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{phistats_summary}
\alias{phistats_summary}
\title{Import and summarise the batch_x.phistats.tsv file.}
\usage{
phistats_summary(data, skip.lines)
}
\arguments{
\item{data}{The 'batch_x.phistats.tsv' created by STACKS.}

\item{skip.line}{The number of line without the header
to start reading the data.}
}
\description{
Import and summarise the batch_x.phistats.tsv file.
Necessary preparation for density distribution and box plot figures.
}

