% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_ibm.R
\name{plink_ibm}
\alias{plink_ibm}
\title{Identity-by-missingness (IBM) analysis.}
\usage{
plink_ibm(vcf.file, whitelist = NULL, denovo = TRUE, blacklist.id = NULL,
  strata, strata.select)
}
\arguments{
\item{vcf.file}{The VCF file created by STACKS.}

\item{whitelist}{(optional) A whitelist of markers to keep, 
for VCFTools 2 columns are required: 'CHROM' and 'POS'. 
Default: \code{whitelist = NULL}.}

\item{denovo}{(logical) Default: \code{denovo = TRUE}. For de novo analysis 
the vcf file created by stacks as a \code{un} in the \code{CHROM} column. 
We need to change that internally for the function to work in VCFTools and 
PLINK.}

\item{blacklist.id}{(optional) A blacklist with individual ID. NO 
column header. The blacklist is in the directory (e.g. "blacklist.id.txt").
Default: \code{blacklist.id = NULL}}

\item{strata}{File with column header 'INDIVIDUALS' and any other 
columns containing e.g. population id, sequencer id, lanes, etc. 
Any info that might impact missing data. 
Willl be used in \code{strata.select} argument.}

\item{strata.select}{Control the colour in ggplot2 plot. 
e.g. use \code{"POP_ID")} to see the strata column \code{"POP_ID"} 
or \code{"LANES"}, if you have a column in your strata file that starts 
with LANES (for sequencing lane...).}
}
\value{
The function returns a list with the identity-by-missingness results 
($ibm) and a MDS plot (MultiDimensional Scaling) of ibm results ($plot)
In the working directory:
the modified vcf file, if de novo was selected, plink tfiles and ibm results.
}
\description{
Diagnose your missing genotypes pattern with PLINK's 
identity-by-missingness (IBM) analysis.
}
\note{
\code{plink_ibm} assumes that the command line version of 
VCFtools \url{https://vcftools.github.io/index.html} and
PLINK version 1.90 \url{https://www.cog-genomics.org/plink2} are 
properly installed and available on the command line, so it is executable from 
any directory (more info on how to do this, here 
\url{http://gbs-cloud-tutorial.readthedocs.org/en/latest/03_computer_setup.html?highlight=bash_profile#save-time}.
On MacOSX, the easiest way is to put plink binary and VCFtools binaries
in the folder \code{/usr/local/bin}. For a copy/paste tutorial on how to
install VCFtools and PLINK follow the steps here \url{http://gbs-cloud-tutorial.readthedocs.org/en/latest/07_start_from_new_image.html#install-gbs-software}
}
\examples{
\dontrun{
With raw VCF without filtering the loci:
ibm <- plink_ibm(
vcf.file = "batch_1.vcf", 
denovo = TRUE, 
strata = "population.map.txt", 
strata.select = "POP_ID"
)

With filtered data:
ibm <- plink_ibm(
vcf.file = "batch_1.vcf", 
whitelist = "whitelist.vcf.txt", 
denovo = TRUE, 
strata = "population.map.txt", 
strata.select = "POP_ID"
)

If the object for the function is 'ibm' then:
to access the figure
ibm.plot <- ibm$plot
The ibm results of plink in a dataframe with your strata:
ibm.results <- ibm$ibm
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Danecek P, Auton A, Abecasis G et al. (2011) 
The variant call format and VCFtools. Bioinformatics, 27, 2156–2158.

Purcell S, Neale B, Todd-Brown K et al. (2007) 
PLINK: a tool set for whole-genome association and population-based linkage 
analyses. American Journal of Human Genetics, 81, 559–575.

Catchen JM, Amores A, Hohenlohe PA et al. (2011) 
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences. 
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013) 
Stacks: an analysis tool set for population genomics. 
Molecular Ecology, 22, 3124-3140.
}

