% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignment.R
\name{plot_assignment_stacked_bar}
\alias{plot_assignment_stacked_bar}
\title{Classic stacked bar plot figure of assignement results.}
\usage{
plot_assignment_stacked_bar(assignment.summary, pop.levels)
}
\arguments{
\item{assignment.summary}{The assignment summary file created
with assignment_genodive function.}

\item{pop.levels}{An optional character string with your populations ordered.}
}
\description{
GGPLOT2 stacked bar plot figure of assignment results.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Meirmans PG, Van Tienderen PH (2004) genotype and genodive:
two programs for the analysis of genetic diversity of asexual organisms.
Molecular Ecology Notes, 4, 792-794.
}

