% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_snp_position_read}
\alias{plot_snp_position_read}
\title{Figure of the distribution of SNP nucleotide position alond the read}
\usage{
plot_snp_position_read(data, aes.colour, y.title)
}
\arguments{
\item{data}{Data for the figure.}

\item{aes.colour}{GGPLOT2 aesthetic.}

\item{y.title}{Title of the Y-axis.}
}
\description{
Figure of the distribution of SNP nucleotide position alond the read
}

