% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/population_filter.R
\name{population_filter}
\alias{population_filter}
\title{Population filter}
\usage{
population_filter(data, pop.threshold, percent, filename)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv").
The data is a sumstats prepared file,
a tidy VCF file or any of the last two file format modified by other filters.}

\item{pop.threshold}{A threshold number: proportion, percentage
or fixed number e.g. 0.70, 70 or 15.}

\item{percent}{Is the threshold a percentage ? TRUE or FALSE.}
}
\description{
Filter loci based on population threshold with loci: number,
percentage or fix threshold.
}

