% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_map.R
\name{population_map}
\alias{population_map}
\title{Read and prepare a population map.}
\usage{
population_map(data, pop.id.start, pop.id.end, pop.levels)
}
\arguments{
\item{data}{The population map or individuals listed in one column. No headers.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.levels}{An optional character string with your populations ordered.}
}
\description{
Read the same population map used in STACKS populations module.
}

