% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_regression_average.R
\name{qra}
\alias{qra}
\title{Quantile Regression Average
Calculates a quantile regression average for forecasts.}
\usage{
qra(
  forecasts,
  data,
  target_date,
  min_date,
  max_date,
  pool,
  intervals,
  per_centile_weights = FALSE,
  enforce_normalisation = TRUE
)
}
\arguments{
\item{forecasts}{data frame with forecasts}

\item{data}{data frame with data}

\item{target_date}{the date for which to create a QRA; by default, will use
the latest \code{creation_date} in \code{forecasts}}

\item{min_date}{the minimum creation date for past forecasts to be included}

\item{max_date}{the maximum creation date for past forecasts to be included}

\item{pool}{any columns to pool as a list of character vectors (e.g.,
horizon, geography_scale, etc.); by default, will not pool across anything}

\item{intervals}{Numeric - which central intervals to consider; by default will
consider the maximum spanning set. Options are determined by data but will be between
0 and 1.}

\item{per_centile_weights}{if TRUE, separate weights are calculated for
different centiles}

\item{enforce_normalisation}{if TRUE, normalisation (weights sum to 1) is enforced}
}
\value{
a list of `ensemble`, a data frame similar to the input forecast,
but with \code{model} set to "Quantile regression average" and the values
set to the weighted averages; and `weights`, a data frame giving the weights
}
\description{
Quantile Regression Average
Calculates a quantile regression average for forecasts.
}
\references{
Nowotarski, J., Weron, R. Computing electricity spot price prediction 
intervals using quantile regression and forecast averaging. Comput Stat 30, 
791–803 (2015). https://doi.org/10.1007/s00180-014-0523-0
}
\author{
Sebastian Funk
}
