% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_depth_plot.R
\name{read_depth_plot}
\alias{read_depth_plot}
\title{Generate a figure with the read depth groups}
\usage{
read_depth_plot(
  fq.file,
  min.coverage.fig = 7L,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{fq.file}{(character, path). The path to the individual fastq file to check.
Default: \code{fq.file = "my-sample.fq.gz"}.}

\item{min.coverage.fig}{(character, path). Minimum coverage used to draw the
color on the figure.
Default: \code{min.coverage.fig = 7L}.}

\item{parallel.core}{(integer) Enable parallel execution with the number of threads.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
The function returns the read depth groups plot.
}
\description{
This function reads the fastq file of an individual and generate
a figure of read coverage groups.
}
\details{
4 read coverage groups are shown:
\enumerate{
\item distinct reads with low coverage (in red): these reads are likely
sequencing errors or uninformative polymorphisms (shared only by a few samples).
\item disting reads for a target coverage (in green):
  \itemize{
  \item Usually represent around 80% of the reads in the FQ file.
  \item It’s a safe coverage range to start exploring your data (open for discussion).
  \item Lower threshold (default = 7): you can’t escape it, it’s your tolerance
to call heterozygote a true heterozygote. You want a minimum coverage for
both the reference and the alternative allele. Yes, you can use population
information to lower this threshold or use some fancy bayesian algorithm.
  \item Higher threshold: is a lot more open for discussion, here it’s the lower
limit of another group (the orange, see below for description). Minus 1 bp.
}
\item distinct reads with high coverage > 1 read depth (in yellow): those are
legitimate locus with high coverage.
\item distinct unique reads with high coverage (in orange): those loci are
usually paralogs, transposable elements, etc.
}
}
\examples{
\dontrun{
require(vroom)
check.reads.depth.groups <- read_depth_plot(fq.file = "my-sample.fq.gz")
}
}
\references{
Ilut, D., Nydam, M., Hare, M. (2014).
Defining Loci in Restriction-Based Reduced Representation Genomic Data from Non
model Species:
Sources of Bias and Diagnostics for Optimal Clustering BioMed Research
International  2014.
https://dx.doi.org/10.1155/2014/675158
}
