% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ustacks.R
\name{read_stacks_ustacks_log}
\alias{read_stacks_ustacks_log}
\title{Read stacks ustacks log}
\usage{
read_stacks_ustacks_log(log.file = NULL, ustacks.folder = NULL)
}
\description{
Read stacks ustacks log
}
\keyword{internal}
