% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stacks_vcf.R
\name{read_stacks_vcf}
\alias{read_stacks_vcf}
\title{Read a VCF file produced by STACKS and transform in tidy format}
\usage{
read_stacks_vcf(vcf.file, pop.id.start, pop.id.end, pop.levels, pop.labels,
  whitelist, blacklist, blacklist.id, filename)
}
\arguments{
\item{vcf.file}{The VCF file created by STACKS.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.levels}{An character string with your populations ordered.}

\item{pop.labels}{An optional character string with new populations names.}

\item{whitelist}{(optional) A whitelist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter by CHROM and/or locus and/or by snp.
The whitelist is in the working directory (e.g. "whitelist.txt").
de novo CHROM column with 'un' need to be changed to 1. 
Default \code{NULL} for no whitelist of markers.}

\item{blacklist}{(optional) A blacklist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter out by CHROM and/or locus and/or by snp.
The blacklist is in the working directory (e.g. "blacklist.loci.txt").
de novo CHROM column with 'un' need to be changed to 1. 
Default \code{NULL} for no blacklist of markers.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the working directory
 (e.g. "blacklist.id.30.txt").}

\item{filename}{(optional) The name of the file written in the directory.}
}
\description{
Import a VCF file created by STACKS and mofify to a tidy format.
}
\details{
Your individuals are identified in this form : 
CHI-QUE-ADU-2014-020, (SPE-POP-MAT-YEA-ID). Then, \code{pop.id.start} = 5
and \code{pop.id.end} = 7. With 5 populations: ALK, JAP, NOR, QUE, LAB, 
the \code{pop.levels} option could look like this: 
c("JAP", "ALK", "QUE", "LAB", "NOR"). Useful for hierarchical clustering.
Allele depth was introduced in STACKS v.1.28, so this field won't 
be available to prior version.
}
\examples{
\dontrun{
vcf.tidy <- read_stacks_vcf(
vcf.file = "batch_1.vcf", 
pop.id.start = 5, 
pop.id.end = 7, 
pop.levels = c("JAP", "ALK", "QUE", "LAB", "NOR"), 
whitelist = "whitelist.loci.txt", 
blacklist = "blacklist.loci.paralogs.txt", 
filename = "vcf.tidy.paralogs.tsv")
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.

Catchen JM, Amores A, Hohenlohe PA et al. (2011) 
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences. 
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013) 
Stacks: an analysis tool set for population genomics. 
Molecular Ecology, 22, 3124-3140.
}

