% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_length_distribution.R
\name{reads_length_distribution}
\alias{reads_length_distribution}
\title{Generate the read length distribution of an individual fastq file}
\usage{
reads_length_distribution(fq.file, parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{fq.file}{(character, path). The path to the individual fastq file to check.
Default: \code{fq.file = "my-sample.fq.gz"}.}

\item{parallel.core}{(integer) Enable parallel execution with the number of threads.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
The function returns a plot and a tibble with potential reads length
thresholds and associated number of reads.
}
\description{
This function reads the fastq file of an individual and
generate the read length distribution to help decide the threshold to cut the
reads to a specific length.
}
\details{
coming soon, just try it in the meantime...
}
\examples{
\dontrun{
require(vroom)
reads.length.info <- stackr::reads_length_distribution(
  fq.file = "my-sample.fq.gz")
}
}
