% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_reads.R
\name{reads_stats}
\alias{reads_stats}
\title{reads_stats}
\usage{
reads_stats(
  fq.files,
  read.depth.plot = TRUE,
  min.coverage.fig = 7L,
  output.dir = NULL,
  parallel.core = parallel::detectCores() - 1,
  p,
  verbose = TRUE
)
}
\description{
Main function to summarise the reads
}
\keyword{internal}
