% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_radproc.R
\name{run_radproc}
\alias{run_radproc}
\title{Run RADproc}
\usage{
run_radproc(
  file.type = "gzfastq",
  f = "04_process_radtags",
  o = "06_ustacks_2_gstacks",
  a = FALSE,
  M = 2,
  m = 3,
  n = 2,
  parallel.core = parallel::detectCores() - 1,
  x = 3,
  S = 2,
  D = 7,
  cmd.path = "/usr/local/bin/RADProc"
)
}
\arguments{
\item{file.type}{(character) Input file Type.
Supported types: fasta, fastq, gzfasta, or gzfastq.
Default: \code{file.type = "gzfastq"}.}

\item{f}{(path) Input file path. Usually,
the stacks process_radtags output folder.
Default: \code{f = "04_process_radtags"}.}

\item{o}{(path) Output path to write results.
Default: \code{o = "06_ustacks_2_gstacks"}.}

\item{a}{(logical) Enable parameter sweep mode.
Default: \code{a = FALSE}.
Please use \code{run_ustacks} with mismatch testing.}

\item{M}{Maximum distance (in nucleotides) allowed between stacks to form
network.
Default: \code{M = 2}.}

\item{m}{Minimum depth of coverage.
Default: \code{m = 3}.}

\item{n}{Maximum distance (in nucleotides) allowed between catalog loci to merge
Default: \code{n = 2}.}

\item{parallel.core}{(integer) Enable parallel execution with num_threads
threads.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{x}{(integer) Maximum number stacks per locus.
Default: \code{x = 3}.}

\item{S}{(percentage) Minimum sample percentage.
Default: \code{S = 30}.}

\item{D}{(percentage) Minimum average coverage depth.
Default: \code{D = 7}.}

\item{cmd.path}{(character, path) Provide the FULL path to RADProc
program. See details on how to install RADProc in
\href{http://thierrygosselin.github.io/stackr/articles/stackr.html#ustackscstacks-alternatives}{stackr vignette}.
Default: \code{cmd.path = "/usr/local/bin/RADProc"}.}
}
\value{
Returns 3 files per samples: \code{.snps.tsv}, \code{.tags.tsv},
\code{.alleles.tsv}. Also returns 3 catalog files: \code{catalog.snps.tsv,
catalog.tags.tsv, catalog.alleles.tsv}.
}
\description{
Runs \href{https://github.com/beiko-lab/RADProc}{RADproc}.
The approach **replaces ustacks and cstacks** steps. Read the paper
for more information.
Read \href{http://thierrygosselin.github.io/stackr/articles/stackr.html}{stackr} vignette.
}
\examples{
\dontrun{
# The simplest form of the function:
u <- stackr::run_radproc() # that's it !
}
}
\references{
Ravindran, P., Bentzen, P., Bradbury, I., Beiko, R. (2019).
RADProc: A computationally efficient de novo locus assembler for population
studies using RADseq data.
Molecular Ecology Resources 19(1), 272-282.
https://dx.doi.org/10.1111/1755-0998.12954
}
\seealso{
\href{https://github.com/beiko-lab/RADProc}{RADproc}.#'
}
