% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rxstacks.R
\name{run_rxstacks}
\alias{run_rxstacks}
\title{Run STACKS rxstacks module}
\usage{
run_rxstacks(P = "06_ustacks_cstacks_sstacks",
  o = "07_rxstacks_cstacks_sstacks_populations", strata = NULL,
  t = parallel::detectCores() - 1, b = "guess", lnl_filter = TRUE,
  lnl_lim = NULL, lnl_dist = TRUE, conf_filter = TRUE, conf_lim = 0.75,
  prune_haplo = TRUE, max_haplo = NULL, model_type = "snp", alpha = 0.1,
  bound_low = 0, bound_high = 1, h = FALSE, verbose = FALSE)
}
\arguments{
\item{P}{(Character) Path to the directory containing STACKS files.
Default: \code{P = "06_ustacks_cstacks_sstacks"}.
Inside the folder \code{06_ustacks_cstacks_sstacks}, you should have:
\itemize{
  \item \strong{4 files for each samples:} The sample name is the prefix of
  the files ending with:
\code{.alleles.tsv.gz, .models.tsv.gz, .snps.tsv.gz, .tags.tsv.gz}.
Those files are created in the
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks}
and
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks}
modules.
}}

\item{o}{(Character) Output path to write results.
Default: \code{o = "/07_rxstacks_cstacks_sstacks_populations"}.}

\item{strata}{(optional) Path to a strata file to get
the rxstacks results by population. With default, the results is presented
overall samples.
The strata file is a tab delimited file with 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
The \code{STRATA} column can be any hierarchical grouping.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
Default: \code{strata = NULL}.}

\item{t}{(Integer) enable parallel execution with num_threads threads.
Default: \code{t = parallel::detectCores() - 1}}

\item{b}{(Integer) Database/batch ID of the input catalog to consider.
Advice: don't modify the default.
Default: \code{b = "guess"}.}

\item{lnl_filter}{(Logical) Filter catalog loci based on the mean log likelihood of
the catalog locus in the population.
Default: \code{lnl_filter = TRUE}.}

\item{lnl_lim}{(Double) Minimum log likelihood required to keep a catalog locus.
Default: \code{lnl_lim = NULL}.}

\item{lnl_dist}{(Logical) Print distribution of mean log likelihoods for
catalog loci AFTER the corrections. If you need the log likelihoods BEFORE the
correction use the output from \code{\link{run_sstacks}}, or if not generated,
use the function \code{\link{summary_catalog_log_lik}}.
Default: \code{lnl_dist = TRUE}.}

\item{conf_filter}{(Logical) Filter confounded loci.
Default: \code{conf_filter = TRUE}.}

\item{conf_lim}{(Double) Proportion of loci in population that must be
confounded relative to the catalog locus.
Default: \code{conf_lim = 0.75}.}

\item{prune_haplo}{(Logical) Prune out non-biological haplotypes unlikely to
occur in the population.
Default: \code{prune_haplo = TRUE}.}

\item{max_haplo}{(Integer) Only consider haplotypes for pruning if they occur
in fewer than \code{max_haplo_cnt samples}.
Default: \code{max_haplo = NULL}.}

\item{model_type}{Either 'snp' (default), 'bounded', or 'fixed'.
Default: \code{model_type = "snp"}.}

\item{alpha}{For the SNP or Bounded SNP model,
Chi square significance level required to call
a heterozygote or homozygote, either 0.1, 0.05.
Default: \code{alpha = 0.1}.}

\item{bound_low}{For the bounded SNP model, lower bound for epsilon,
the error rate, between 0 and 1.0.
Default: \code{bound_low = 0}.}

\item{bound_high}{For the bounded SNP model, upper bound for epsilon,
the error rate, between 0 and 1.0.
Default: \code{bound_high = 1}.}

\item{h}{Display this help messsage.
Default: \code{h = FALSE}}

\item{verbose}{Extended logging, including coordinates of all changed
nucleotides (forces single-threaded execution).
Advice: don't modify the default.
Default: \code{verbose = FALSE}}
}
\value{
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}
returns a \code{.matches.tsv.gz file for each sample}
}
\description{
Run \href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}
module inside R!
}
\details{
\strong{Computer or server problem during the rxstacks ? Power outage? No problem:}
Restart the function as it was, the function will pick up at the sample it
was correcting before the problem.
}
\examples{
\dontrun{
# First: get the log-likelihood distribution of catalog loci from
# \\code{\\link{run_sstacks}} output.
# If not run during \\code{\\link{run_sstacks}},
use the function \\code{\\link{summary_catalog_log.lik}}.
# Run rxstacks with appropriate filter thresholds:
rxstacks.tiger <- stackr::run_rxstacks(
lnl_filter = TRUE,
lnl_lim = -10,
lnl_dist = FALSE,
model_type = "bounded",
bound_low = 0, bound_high = 0.15,
strata = "population_map_tiger.txt")
# get all sorts of tables and figures:
rxstacks.tiger$rxstacks.correction.ind
rxstacks.tiger$rxstacks.correction.overall
rxstacks.tiger$rxstacks.correction.overall.proportion
rxstacks.tiger$rxstacks.correction.filled.bar.plot
rxstacks.tiger$rxstacks.correction.bar.plot
# Get the distribution of log-likelihood of catalog loci after correction:
rxstacks.tiger$log.likelihood.fig
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}
}
