% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sstacks.R
\name{run_sstacks}
\alias{run_sstacks}
\title{Run STACKS sstacks module}
\usage{
run_sstacks(b = "guess", P = "06_ustacks_cstacks_sstacks", M = NULL,
  sample.list = NULL, catalog.prefix = "guess", g = FALSE,
  p = parallel::detectCores() - 1, o = "06_ustacks_cstacks_sstacks",
  x = FALSE, gapped = TRUE, lnl_dist = TRUE, v = FALSE, h = FALSE)
}
\arguments{
\item{b}{(Integer) Database/batch ID of the input catalog to consider.
Advice: don't modify the default.
Default: \code{b = "guess"}.}

\item{P}{(character) Path to the directory containing STACKS files.
Contrary to \href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks},
this argument can be use with \code{sample.list} (the \code{s} in \href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks})
Default: \code{P = "06_ustacks_cstacks_sstacks"}.}

\item{M}{(character) Path to a population map file from which to take sample names (this argument won't work if \code{sample.list}).
Advice: don't modify the default (please see details).
Default: \code{M = NULL}.}

\item{sample.list}{This is for the \code{s} option in
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks}.
\code{s: Filename prefix from which to load sample loci}.
Here, you have 2 choices: 1. you leave empty and let the function use the
default:\code{sample.list = NULL} which will scan for the files in the
\code{P} folder given above. 2. you supply a character string of the
samples. This could come from the \code{INDIVIDUALS_REP} column of the
project info file, e.g. \code{sample.list = project.info$INDIVIDUALS_REP}.
(Please see details).}

\item{catalog.prefix}{This is for the \code{c} option in
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks}.
\code{c: TSV file from which to load the catalog loci.}
With the default, the function will guess the catalog prefix.
Using \code{catalog.prefix = NULL} tells the function you want to use \code{P/M} mode (see details).
Default: \code{catalog.prefix = "guess"}.}

\item{g}{Base matching on genomic location, not sequence identity.
Default: \code{g = FALSE}}

\item{p}{(Integer) Enable parallel execution with num_threads threads.
Default: \code{p = parallel::detectCores() - 1}}

\item{o}{output path to write results.
Default: \code{o = "06_ustacks_cstacks_sstacks"}}

\item{x}{Don't verify haplotype of matching locus.
Default: \code{x = FALSE}}

\item{gapped}{Gapped assembly options: do you want to preform
gapped alignments between stacks.
Default: \code{gapped = TRUE}}

\item{lnl_dist}{(Logical) Print distribution of mean log likelihoods for catalog loci.
This argument is different from \href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks},
but will save you time and help you for the next step,
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}.
Use the log-likelihood distribution to choose the appropriate filter thresholds for
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}.
If this is the second time you run the function (after rxstacks), you get the
the distribution AFTER the rxstacks corrections.
Default: \code{lnl_dist = TRUE}.}

\item{v}{Print program version.
Default: \code{v = FALSE}}

\item{h}{Display this help messsage.
Default: \code{h = FALSE}}
}
\value{
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks}
returns a \code{.matches.tsv.gz file for each sample}.
If \code{lnl_dist = TRUE}, the function will also return a
summary of catalog loci log-likelihood.
}
\description{
Run \href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks}
module inside R!
Inside the \code{P} folder (where the \href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks} and the \href{http://catchenlab.life.illinois.edu/stacks/comp/cstacks.php}{cstacks} files are),
you should have:
\itemize{
  \item \strong{3 Catalog files:} the files created in
\href{http://catchenlab.life.illinois.edu/stacks/comp/cstacks.php}{cstacks}
and usually looking like this:
\code{batch_1.catalog.alleles.tsv.gz,
batch_1.catalog.snps.tsv.gz,
batch_1.catalog.tags.tsv.gz}
  \item \strong{4 files for each samples:} The sample name is the prefix of
  the files ending with:
\code{.alleles.tsv.gz, .models.tsv.gz, .snps.tsv.gz, .tags.tsv.gz}.
Those files are created in the
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks}
module.
}

The same files should be present if this function is ran after
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}, note
that you still need to generate a new catalog after \href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}
before running \href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks}.
}
\details{
\strong{Computer or server problem during the sstacks ?}
Just launch again the same command, the function will start again, but only
with the unmatched samples!



\strong{Some argument can't be use together:}

Don't use these arguments together: \code{M} and \code{sample.list/catalog.prefix}.
You can't add samples to an existing catalog using a population map.
If you want to match new samples to an existing catalog, very easy,
just add the samples in the folder and let \code{run_sstacks} function find the samples that were not matched to the catalog.
}
\examples{
\dontrun{
# The simplest form of the function when using the stackr workflow:
run_sstacks()
# that's it ! Now if you have your own workflow folders, etc. Enter them like this:
sstacks <- run_sstacks (P = "/my/input/path", p = 32, b = 2, catalog.prefix = "batch_2",
sample.list = c("ind1", "ind2", "..."), o = "/my/output/path", g = FALSE,
x = FALSE, gapped = FALSE)
# Get the fig of log-likelihood of catalog loci to help decide for the best
# rxstacks thresholds:
sstacks$log.likelihood.fig
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks}
}
