% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tsv2bam.R
\name{run_tsv2bam}
\alias{run_tsv2bam}
\title{Run STACKS tsv2bam and merges BAM files}
\usage{
run_tsv2bam(P = "06_ustacks_2_gstacks",
  M = "02_project_info/population.map.tsv2bam.tsv", R = NULL,
  t = parallel::detectCores() - 1,
  cmd.path = "/usr/local/bin/samtools", h = FALSE)
}
\arguments{
\item{P}{(path, character) Path to the directory containing STACKS files.
Default: \code{P = "06_ustacks_2_gstacks"}.
Inside the folder, you should have:
\itemize{
  \item \strong{the catalog files:} starting with \code{batch_} and ending with
  \code{.alleles.tsv.gz, .snps.tsv.gz, .tags.tsv.gz};
  \item \strong{3 files for each samples:} The sample name is the prefix for
  the files ending with:
\code{.alleles.tsv.gz, .snps.tsv.gz, .tags.tsv.gz}.
Those files are created in the
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks},
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks} and
\href{http://catchenlab.life.illinois.edu/stacks/comp/cxstacks.php}{cxstacks}
modules.
}}

\item{M}{(character, path) Path to a population map file.
Note that the \code{-s} option is not used inside \strong{stackr}.
Default: \code{M = "02_project_info/population.map.tsv2bam.tsv"}.}

\item{R}{(path, character) Directory where to find the paired-end reads files
(in fastq/fasta/bam (gz) format).}

\item{t}{(integer) Enable parallel execution with the number of threads.
Default: \code{t = parallel::detectCores() - 1}}

\item{cmd.path}{(character, path) Provide the FULL path to SAMtools
program. See details on how to install SAMtools.
Default: \code{cmd.path = "/usr/local/bin/samtools"}.}

\item{h}{Display this help messsage.
Default: \code{h = FALSE}}
}
\value{
\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{tsv2bam}
returns a set of \code{.matches.bam} files.

The function \code{run_tsv2bam} returns a list with the number of individuals, the batch ID number,
a summary data frame and a plot containing:
\enumerate{
\item INDIVIDUALS: the sample id
\item ALL_LOCUS: the total number of locus for the individual (shown in subplot A)
\item LOCUS: the number of locus with a one-to-one relationship (shown in subplot B)
with the catalog
\item MATCH_PERCENT: the percentage of locus with a one-to-one relationship
with the catalog (shown in subplot C)

Addtionally, the function returns a catalog.bam file, generated
by merging all the individual BAM files in parallel.
}
}
\description{
Runs \href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{tsv2bam}
module and additionnally, this function will also generate a summary of
stacks tsv2bam and will merge in parallel BAM sample files into a unique
BAM catalog file using SAMtools or Sambamba.
\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{tsv2bam}
converts the data (single-end or paired-end) from being organized by sample
into being organized by locus. This allows downstream improvements
(e.g. Bayesian SNP calling).
}
\details{
\strong{Install SAMtools}
\href{link to detailed instructions on how to install SAMtools}{http://gbs-cloud-tutorial.readthedocs.io/en/latest/07_start_from_new_image.html#install-gbs-radseq-software-time-required-30min}
}
\examples{
\dontrun{
# The simplest form of the function:
bam.sum <- stackr::run_tsv2bam() # that's it !
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.

Li H., Handsaker B., Wysoker A., Fennell T., Ruan J., Homer N.,
Marth G., Abecasis G., Durbin R. and 1000 Genome Project Data Processing
Subgroup (2009) The Sequence alignment/map (SAM) format and SAMtools.
Bioinformatics, 25, 2078-9.

Li H A statistical framework for SNP calling,
mutation discovery, association mapping and population genetical parameter
estimation from sequencing data. Bioinformatics. 2011 Nov 1;27(21):2987-93.

A. Tarasov, A. J. Vilella, E. Cuppen, I. J. Nijman, and P. Prins.
Sambamba: fast processing of NGS alignment formats. Bioinformatics, 2015.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}

\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{stacks Version 2.0Beta6}

\href{http://www.htslib.org}{SAMtools}

\href{http://lomereiter.github.io/sambamba/}{Sambamba}
}
