% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ustacks.R
\name{run_ustacks}
\alias{run_ustacks}
\title{Run STACKS ustacks module}
\usage{
run_ustacks(mismatch.testing = FALSE, sample.list = NULL,
  project.info = NULL, f = "04_process_radtags",
  o = "06_ustacks_cstacks_sstacks", m = 3, M = 2, N = M + 2,
  t = "guess", R = FALSE, H = TRUE, p = parallel::detectCores() - 1,
  h = FALSE, d = TRUE, keep_high_cov = FALSE, max_locus_stacks = 3,
  k_len = NULL, gapped = TRUE, max_gaps = 2, min_aln_len = 0.8,
  model_type = "snp", alpha = 0.05, bound_low = 0, bound_high = 0.2,
  bc_err_freq = NULL)
}
\arguments{
\item{mismatch.testing}{(logical). Default: \code{mismatch.testing = FALSE}.}

\item{sample.list}{The default use all the samples in \code{f},
usually, the stacks process_radtags output folder, here defaulted to
\code{f = "04_process_radtags"}, see \code{f} argument below.
When using \code{mismatch.testing = TRUE}, only one sample is allowed inside the folder specified in \code{f}
(e.g. choose one with a mean number of read, or MB, file size).
Default: \code{sample.list = NULL}.
Power outage? no problem, see details below.}

\item{project.info}{When using the stackr pipeline,
a project info file is created. This file will be modified inside this function.
The file is in the working directory (given the path or in the global environment).
If no \code{project.info} file is provided, the function first look in the
working directory for file(s) with "project.info" in it's name. If several files are found,
the latest one created is used.
Default: \code{project.info = NULL}.
Power outage? no problem, see details below.}

\item{f}{Input file path. Usually,
the stacks process_radtags output folder.
Default: \code{f = "04_process_radtags"}.}

\item{o}{Output path to write results.
Default: \code{o = "06_ustacks_cstacks_sstacks"}.}

\item{m}{Minimum depth of coverage required to create a stack.
Default: \code{m = 3}.}

\item{M}{Maximum distance (in nucleotides) allowed between stacks.
Default: \code{M = 2}.}

\item{N}{Maximum distance allowed to align secondary reads to primary stacks.
Default: \code{N = M + 2}.}

\item{t}{Input file type.
Supported types: fasta, fastq, gzfasta, gzfastq, fq.gz, fastq.gz.
Default: \code{t = "guess"}.}

\item{R}{Retain unused reads. Default: \code{R = FALSE}.}

\item{H}{Disable calling haplotypes from secondary reads.
Default: \code{H = TRUE}.}

\item{p}{Enable parallel execution with num_threads threads.
Default: \code{p = parallel::detectCores() - 1}.}

\item{h}{Display this help messsage. Default: \code{h = FALSE}.}

\item{d}{Enable the Deleveraging algorithm, used for resolving over merged tags.
Default: \code{d = TRUE}.}

\item{keep_high_cov}{Disable the algorithm that removes highly-repetitive stacks and nearby errors.
Default: \code{keep_high_cov = FALSE}.}

\item{max_locus_stacks}{Maximum number of stacks at a single de novo locus.
Default: \code{max_locus_stacks = 3}.}

\item{k_len}{Specify k-mer size for matching between alleles and loci.
Default: \code{k_len = NULL}.}

\item{gapped}{Perform gapped alignments between stacks.
Default: \code{gapped = TRUE}.}

\item{max_gaps}{Number of gaps allowed between stacks before merging.
Default: \code{max_gaps = 2}.}

\item{min_aln_len}{Minimum length of aligned sequence in a gapped alignment.
Default: \code{min_aln_len = 0.8}.}

\item{model_type}{Either 'snp' (default), 'bounded', or 'fixed'.
Default: \code{model_type = "snp"}.}

\item{alpha}{For the SNP or Bounded SNP model,
Chi square significance level required to call
a heterozygote or homozygote, either 0.1, 0.05.
Default: \code{alpha = 0.05}.}

\item{bound_low}{For the bounded SNP model, lower bound for epsilon,
the error rate, between 0 and 1.0.
Default: \code{bound_low = 0}.}

\item{bound_high}{For the bounded SNP model, upper bound for epsilon,
the error rate, between 0 and 1.0.
Default: \code{bound_high = 0.2}.}

\item{bc_err_freq}{For the fixed model, specify the barcode error frequency, between 0 and 1.0.
Default: \code{bc_err_freq = NULL}.}
}
\value{
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks}
returns 4 files per samples: \code{.snps.tsv.gz}, \code{.tags.tsv.gz},
\code{.alleles.tsv.gz},  \code{.models.tsv.gz}. In the global environment,
the function returns a project info file (updated if one was provided) and
a summary of ustacks for each samples.
}
\description{
Run \href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks}
module inside R!
Inside the folder \code{04_process_radtags}, you should have all the
individual's fastq files. Those files are created in the
\href{http://catchenlab.life.illinois.edu/stacks/comp/process_radtags.php}{process_radtags}
module.
}
\details{
\code{-i} the unique integer ID to identify the sample (SQL ID),
is taken from the project info file. If no project info file is provided,
the id is created sequentially from the sample files. This id will be written
in the project info file (if no file is found or given, a new file is created in the working directory).

\strong{Power outage? No problem:}

Restart the function as it was. After the re-start the project info file created automatically during the previous run
will be used. This ensure: i) that the unique SQL ids are not duplicated and
ii) that ustacks can start at the sample is was assembling before the outage.
}
\examples{
\dontrun{
# to test different mismatches (M values from 1 to 5) on a sample with an average read number:
mismatch <- stackr::run_ustacks(
mismatch.testing = TRUE,
f = “mismatch_testing_folder_path”,
M = 1:5,
p = 12)
# the other arguments: defaults
# Check out the list to figure out the best threshold to use on all samples
# A summary table is also written in the output folder.

# To run ustacks on all samples, using defaults:
ustacks.lobster <- stackr::run_ustacks()
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks}
}
