% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ustacks.R
\name{run_ustacks_one_sample}
\alias{run_ustacks_one_sample}
\title{run_ustacks_one_sample}
\usage{
run_ustacks_one_sample(mismatch.testing = FALSE, sample.list = NULL,
  project.info = NULL, f = "04_process_radtags",
  o = "06_ustacks_cstacks_sstacks", m = 3, M = 2, N = M + 2,
  t = "guess", R = FALSE, H = TRUE, p = parallel::detectCores() -
  1, h = FALSE, d = TRUE, keep.high.cov = FALSE,
  high.cov.thres = 3, max.locus.stacks = 3, k.len = NULL,
  max.gaps = 2, min.aln.len = 0.8, disable.gapped = FALSE,
  model.type = "snp", alpha = 0.05, bound.low = 0,
  bound.high = 0.2, bc.err.freq = NULL)
}
\description{
run ustacks for 1 sample
}
\keyword{internal}
