% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tsv2bam.R
\name{split_bam_list}
\alias{split_bam_list}
\title{split_bam_list}
\usage{
split_bam_list(x, parallel.core = parallel::detectCores() - 1)
}
\description{
Split a list of bam files for parallel merging
}
\keyword{internal}
