% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stack_answers.R
\name{stack_answers}
\alias{stack_answers}
\title{Query answers from the Stack Exchange API}
\usage{
stack_answers(id = NULL, special = NULL, ...)
}
\arguments{
\item{id}{A vector containing one or more answer IDs, or none to query
all answers}

\item{special}{If \code{"comments"}, return the comments on the answers
rather than the answers themselves}

\item{...}{Additional arguments to API (see below)}
}
\value{
A \code{data.frame} of answers.

An answer object is documented in full here:

https://api.stackexchange.com/docs/types/answer

It comes with the following columns:

\itemize{
  \item{answer_id}
  \item{community_owned_date}
  \item{creation_date}
  \item{is_accepted}
  \item{last_activity_date}
  \item{last_edit_date}
  \item{locked_date}
  \item{owner}
  \item{protected_date}
  \item{question_id}
  \item{score}
}
}
\description{
Query for a list of answers, or information related to a one or more
specific answers.
}
\details{
API querying methods allow the following additional options:
\describe{
  \item{site}{ID of Stack Exchange site (by default, Stack Overflow)}
  \item{sort}{field to sort by}
  \item{order}{whether the \code{sort} field should be arranged in descending
("desc") or ascending ("asc") order}
  \item{min}{Minimum value of the \code{sort} field}
  \item{max}{Maximum value of the \code{sort} field}
  \item{fromdate}{Starting date}
  \item{todate}{Ending date}
  \item{page}{Which page to start from}
  \item{pagesize}{Size of each page to extract (max 100)}
  \item{num_pages}{Number of pages to extract}
}
}

