% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stack_badges.R
\name{stack_badges}
\alias{stack_badges}
\title{Query badges from the Stack Exchange API}
\usage{
stack_badges(id = NULL, special = NULL, ...)
}
\arguments{
\item{id}{A vector containing one or more IDs of badges, or none to
retrieve all badges in alphabetical order}

\item{special}{One of \code{c("name", "tags", "recipients")}, to
retrive only non-tag badges, only tag badges, or to retrieve recipients
of the given badges}

\item{...}{Additional arguments to API (see below)}
}
\value{
A \code{data.frame} of badges (TODO)
}
\description{
Query for badges from a particular site
}
\details{
API querying methods allow the following additional options:
\describe{
  \item{site}{ID of Stack Exchange site (by default, Stack Overflow)}
  \item{sort}{field to sort by}
  \item{order}{whether the \code{sort} field should be arranged in descending
("desc") or ascending ("asc") order}
  \item{min}{Minimum value of the \code{sort} field}
  \item{max}{Maximum value of the \code{sort} field}
  \item{fromdate}{Starting date}
  \item{todate}{Ending date}
  \item{page}{Which page to start from}
  \item{pagesize}{Size of each page to extract (max 100)}
  \item{num_pages}{Number of pages to extract}
}
}

