% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stack_comments.R
\name{stack_comments}
\alias{stack_comments}
\title{Query comments from the Stack Exchange API}
\usage{
stack_comments(id = NULL, ...)
}
\arguments{
\item{id}{A vector containing one or more IDs, or none to query
all comments}

\item{...}{Additional arguments to API (see below)}
}
\value{
A \code{data.frame} of questions or answers (TODO)
}
\description{
Query for comments, either based on IDs or on other filters.
}
\details{
API querying methods allow the following additional options:
\describe{
  \item{site}{ID of Stack Exchange site (by default, Stack Overflow)}
  \item{sort}{field to sort by}
  \item{order}{whether the \code{sort} field should be arranged in descending
("desc") or ascending ("asc") order}
  \item{min}{Minimum value of the \code{sort} field}
  \item{max}{Maximum value of the \code{sort} field}
  \item{fromdate}{Starting date}
  \item{todate}{Ending date}
  \item{page}{Which page to start from}
  \item{pagesize}{Size of each page to extract (max 100)}
  \item{num_pages}{Number of pages to extract}
}
}

