% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking_weights_crps.R
\name{stack_crps}
\alias{stack_crps}
\title{Model weighting by CRPS}
\usage{
stack_crps(data, lambda = NULL, gamma = NULL, dirichlet_alpha = 1.001)
}
\arguments{
\item{data}{a data.frame with the following entries: 
\itemize{
  \item model (the model used to generate the correspondig predictions)
  \item geography (the region for which predictions are generated)
  \item date (the date of corresponding )
}}

\item{lambda}{weigh timepoints differently}

\item{gamma}{weigh countries differently}

\item{dirichlet_alpha}{prior for the weights}
}
\value{
returns a vector with the model weights
}
\description{
Model weights for stacking by CRPS. MISSING.
}
\examples{



splitdate <- as.Date("2020-04-01") 

traindata <- stackr::sample_prepared_data \%>\%
  dplyr::filter(date <= splitdate)

testdata <- stackr::sample_prepared_data \%>\%
  dplyr::filter(date > splitdate)

weights <- stackr::stack_crps(traindata)

test_mixture <- mixture_from_sample(testdata, 
                                    weights = weights)

rbind(testdata, 
      test_mixture) \%>\%
  group_by(model, forecast_date) \%>\%
  dplyr::mutate(crps = scoringutils::crps(unique(y_obs), 
                                          t(as.vector(y_pred)))) \%>\%
  group_by(model) \%>\%
  dplyr::summarise(crps = mean(crps))


#Missing
}
\references{
Missing
}
