% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{stack_GET}
\alias{stack_GET}
\title{Make a GET request to the Stack Exchange API}
\usage{
stack_GET(path, site = "stackoverflow", page = 1, num_pages = 1, ...)
}
\arguments{
\item{path}{the query path, such as "answers/" or "users/{id}"}

\item{site}{site to query; by default Stack Overflow}

\item{page}{which page to start from}

\item{num_pages}{number of consecutive pages to query; by default 1}

\item{...}{additional parameters to the method}
}
\description{
Make a GET request to the Stack Exchange API
}

