% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stack_info.R
\name{stack_info}
\alias{stack_info}
\title{Query general info for a Stack Exchange site}
\usage{
stack_info(site = "stackoverflow", ...)
}
\arguments{
\item{site}{Stack Exchange site to query (default Stack Overflow)}

\item{...}{Additional API arguments (not used)}
}
\value{
A one-row \code{data.frame} containing statistics
about the site.
}
\description{
Query for general information from a site, such as the number of users,
answers, and questions, and statistics for activity per minute
}

