% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{stack_parse}
\alias{stack_parse}
\title{Parse the results of a Stack Exchange API query into a data.frame.}
\usage{
stack_parse(req)
}
\arguments{
\item{req}{a request from httr::GET}
}
\description{
The additional metadata, such as "has_more", "quota_max", and
"quota_remaining" is included in a list as the attribute "metadata".
}

