% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stack_posts.R
\name{stack_posts}
\alias{stack_posts}
\title{Query posts from the Stack Exchange API}
\usage{
stack_posts(id = NULL, special = NULL, ...)
}
\arguments{
\item{id}{A vector containing one or more IDs, or none to query
all questions and answers}

\item{special}{One of \code{c("comments", "revisions", "suggested-edits")}, to
return that information associated with the specified answers rather than the
answers themselves}

\item{...}{Additional arguments to API (see below)}
}
\value{
A \code{data.frame} of questions or answers (TODO)
}
\description{
Query for posts, either questions or answers, based on IDs.
}
\details{
API querying methods allow the following additional options:
\describe{
  \item{site}{ID of Stack Exchange site (by default, Stack Overflow)}
  \item{sort}{field to sort by}
  \item{order}{whether the \code{sort} field should be arranged in descending
("desc") or ascending ("asc") order}
  \item{min}{Minimum value of the \code{sort} field}
  \item{max}{Maximum value of the \code{sort} field}
  \item{fromdate}{Starting date}
  \item{todate}{Ending date}
  \item{page}{Which page to start from}
  \item{pagesize}{Size of each page to extract (max 100)}
  \item{num_pages}{Number of pages to extract}
}
}

