% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stack_privileges.R
\name{stack_privileges}
\alias{stack_privileges}
\title{Query privileges for a Stack Exchange site}
\usage{
stack_privileges(site = "stackoverflow", ...)
}
\arguments{
\item{site}{Stack Exchange site to query (default Stack Overflow)}

\item{...}{Additional API arguments (not used)}
}
\value{
A one-row \code{data.frame} containing statistics
about the site.
}
\description{
Query for reputation-based privileges from a Stack Exchange site.
}

