% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stack_search.R
\name{stack_search}
\alias{stack_search}
\title{Search a Stack Exchange site via the API}
\usage{
stack_search(intitle, tagged, nottagged, q, accepted, answers, body, closed,
  migrated, notice, title, user, url, views, wiki, ...)
}
\arguments{
\item{intitle}{string to use for searching title}

\item{tagged}{vector of tags that must be included in questions}

\item{nottagged}{vector of tags to be excluded from questions}

\item{q}{free form text parameter that matches questions based on Stack
Exchange's (undocumented) algorithm}

\item{accepted}{boolean to filter for accepted or unaccepted questions}

\item{answers}{minimum number of answers}

\item{body}{text that must appear in the body of questions}

\item{closed}{boolean to filter for open or closed questions}

\item{migrated}{boolean to filter for questions that were, or were not,
migrated to another site}

\item{notice}{boolean to filter for questions with a post notice}

\item{title}{text which must appear in a title; redundant with intitle above}

\item{user}{the id of the user to filter for}

\item{url}{url that must be included in the question}

\item{views}{the minimum number of views a question can have}

\item{wiki}{boolean to filter for questions that are (or are not) community
wiki}

\item{...}{Additional arguments to API (see below)}
}
\value{
A \code{data.frame} of questions.

A question object is documented in full here:

https://api.stackexchange.com/docs/types/question

It comes with the following columns:

\itemize{
  \item{accepted_answer_id}
  \item{answer_count}
  \item{bounty_amount}
  \item{bounty_closes_date}
  \item{closed_date}
  \item{closed_reason}
  \item{community_owned_date}
  \item{creation_date}
  \item{is_answered}
  \item{last_activity_date}
  \item{last_edit_date}
  \item{link}
  \item{locked_date}
  \item{migrated_from}
  \item{migrated_to}
  \item{owner}
  \item{protected_date}
  \item{question_id}
  \item{score}
  \item{tags}
  \item{title}
  \item{view_count}
}
}
\description{
Access the Stack Exchange search functionality.
}
\details{
API querying methods allow the following additional options:
\describe{
  \item{site}{ID of Stack Exchange site (by default, Stack Overflow)}
  \item{sort}{field to sort by}
  \item{order}{whether the \code{sort} field should be arranged in descending
("desc") or ascending ("asc") order}
  \item{min}{Minimum value of the \code{sort} field}
  \item{max}{Maximum value of the \code{sort} field}
  \item{fromdate}{Starting date}
  \item{todate}{Ending date}
  \item{page}{Which page to start from}
  \item{pagesize}{Size of each page to extract (max 100)}
  \item{num_pages}{Number of pages to extract}
}
}

