% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stack_sites.R
\name{stack_sites}
\alias{stack_sites}
\title{Query sites from the Stack Exchange API}
\usage{
stack_sites(...)
}
\arguments{
\item{...}{Extra options are \code{page} and \code{pagesize}}
}
\value{
A \code{data.frame} of sites
}
\description{
Query for a list of sites in the network
}

