% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stack_tags.R
\name{stack_tags}
\alias{stack_tags}
\title{Query tags from the Stack Exchange API}
\usage{
stack_tags(name = NULL, special = NULL, ...)
}
\arguments{
\item{name}{A vector containing one or more names of tags, or none to query
from all tags}

\item{special}{One of \code{c("faq", "related", "synonyms", "wikis")}, to
extract information about a particular tag, or one of
\code{c("moderator-only", "required", "synonyms")}, to request a list of a
particular kind of tags, or all synonyms on the site.}

\item{...}{Additional arguments to API (see below)}
}
\value{
A \code{data.frame} of questions or answers (TODO)
}
\description{
Query for tags, either based on names or on other filters.
}
\details{
API querying methods allow the following additional options:
\describe{
  \item{site}{ID of Stack Exchange site (by default, Stack Overflow)}
  \item{sort}{field to sort by}
  \item{order}{whether the \code{sort} field should be arranged in descending
("desc") or ascending ("asc") order}
  \item{min}{Minimum value of the \code{sort} field}
  \item{max}{Maximum value of the \code{sort} field}
  \item{fromdate}{Starting date}
  \item{todate}{Ending date}
  \item{page}{Which page to start from}
  \item{pagesize}{Size of each page to extract (max 100)}
  \item{num_pages}{Number of pages to extract}
}
}

