% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stack_users.R
\name{stack_users}
\alias{stack_users}
\title{Query information about users from the Stack Exchange API}
\usage{
stack_users(id = NULL, special = NULL, ...)
}
\arguments{
\item{id}{A vector containing one or more user IDs}

\item{special}{One of many options to describe what information to
extract from each user}

\item{...}{Additional arguments to API (see below)}
}
\value{
A \code{data.frame} containing each returned user

A \code{data.frame} of users
}
\description{
Query for a list of users, or information related to one or more
specific users.
}
\details{
API querying methods allow the following additional options:
\describe{
  \item{site}{ID of Stack Exchange site (by default, Stack Overflow)}
  \item{sort}{field to sort by}
  \item{order}{whether the \code{sort} field should be arranged in descending
("desc") or ascending ("asc") order}
  \item{min}{Minimum value of the \code{sort} field}
  \item{max}{Maximum value of the \code{sort} field}
  \item{fromdate}{Starting date}
  \item{todate}{Ending date}
  \item{page}{Which page to start from}
  \item{pagesize}{Size of each page to extract (max 100)}
  \item{num_pages}{Number of pages to extract}
}
}

