% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackr.R
\name{stackr}
\alias{stackr}
\title{Pass Caught Errors to a Stackoverflow Query}
\usage{
stackr(expr, browse = TRUE)
}
\arguments{
\item{expr}{an object to be evaluated.}

\item{browse}{open web browser? Default is TRUE}
}
\description{
Pass Caught Errors to a Stackoverflow Query
}
\examples{
\dontrun{
stackr({
 1+1
 is.na(NA)
 plot(1,1)
 for(i in 1:3) print(b)
})
}
}
