% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackr_imputations_module.R
\name{stackr_imputer}
\alias{stackr_imputer}
\title{stackr_imputer}
\usage{
stackr_imputer(data, num.tree = 100, pred.mean.matching = 0,
  random.seed = NULL, parallel.core = parallel::detectCores() - 1,
  hierarchical.levels = "populations", marker.list = marker.list,
  verbose = verbose)
}
\description{
imputations using Ranger package and predictive mean matching
}
\keyword{internal}
