% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_sample.R
\name{stackr_normalize}
\alias{stackr_normalize}
\title{stackr_normalize}
\usage{
stackr_normalize(fastq.files, project.info, number.samples, sample.reads,
  parallel.core = parallel::detectCores() - 1)
}
\description{
main normalize function
}
\keyword{internal}
