% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_coverage}
\alias{summary_coverage}
\title{Coverage summary}
\usage{
summary_coverage(tidy.vcf, pop.levels, filename)
}
\arguments{
\item{tidy.vcf}{A tidy vcf object or file (using ".tsv"), 
created with read_stacks_vcf.}

\item{pop.levels}{Character string defining your ordered populations.}

\item{filename}{Name of the file saved to the working directory.}
}
\value{
A list with 2 tables: the long format of loci and populations
coverage statistics and the short format by populations.
The short-format is more user-friendly and
is written to the working directory.
}
\description{
This function create a table summary of the important
coverage statistics from the tidy vcf created with read_stacks_vcf.
}
\details{
The tables contains summary statistics (mean, median, min, max)
of read, ref and alt allele coverage. To access 
the two tables, use $. The table that summarise by populations was created
using average nested: loci -> individuals -> populations.
The long format is used for creating figures.
}

