% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cstacks.R
\name{summary_cstacks}
\alias{summary_cstacks}
\title{Summarize STACKS cstacks log file generated by stackr}
\usage{
summary_cstacks(cstacks.log, verbose = FALSE)
}
\arguments{
\item{cstacks.log}{(path, character).
The cstacks log file generated by \code{\link{run_cstacks}}.}

\item{verbose}{(logical, optional) Make the function a little more chatty during
execution.
Default: \code{verbose = FALSE}.}
}
\value{
When the function is run separately it returns an object in the
global environment and a file in \code{08_stacks_results} folder, if it exists and
if not, the file is written in the working directory.
The summary is a tibble:
\enumerate{
\item INDIVIDUALS: the sample id
\item LOCUS_CATALOG_START: the number of locus in the catalog at the start when
processing the sample.
\item LOCI_MATCH_TO_CATALOG: the number of loci for the sample matching loci
in the catalog.
\item LOCI_MATCH_TO_CATALOG_GAPPED: the number of loci for the sample matching loci
in the catalog after allowing for gapped alignment.
\item LOCI_ADDED: the number of loci added to the catalog after processing the
sample.
\item LOCI_LINKED: the number of loci that matched more than one catalog
locus (linked loci).
\item LOCUS_CATALOG_END: the number of locus in the catalog at the end of
processing the sample.
\item MISMATHES: the number of mismatches allowed between loci of different samples.
}
}
\description{
This function reads the log file of stackr \code{\link{run_cstacks}}
and summarise the information.

The information shown is particularly helpful when choosing the best thresholds.
This function is run automatically inside \code{\link{run_cstacks}}, but
it can be run on it's own.
}
\examples{
\dontrun{
cstacks.summary <- stackr::summary_cstacks(
cstacks.log = "09_log_files/cstacks_20191101@1108.log",
verbose = TRUE
)
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/cstacks.php}{cstacks}
\code{\link{run_cstacks}}
}
