% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_genotype_likelihood}
\alias{summary_genotype_likelihood}
\title{Genotype likelihood summary}
\usage{
summary_genotype_likelihood(tidy.vcf.file, pop.levels, filename)
}
\arguments{
\item{tidy.vcf.file}{The tidy VCF file created with read_stacks_vcf.}

\item{pop.levels}{Character string defining your ordered populations.}

\item{filename}{The name of the file written in the directory.}
}
\value{
A list with 2 tables: the long format of loci ($gl.summary.long)
and populations genotype likelihood statistics
and the short format by populations ($gl.summary.pop).
The short-format is more user-friendly and
is written to the working directory.
}
\description{
This function create 2 tables summary of the important
genotype likelihood statistics from the tidy vcf created
with \link{read_stacks_vcf}.
}
\details{
The table contains summary statistics: mean, median, min, max and 
diff (max-min), of genotype likelihood by locus and populations. To access 
the two tables, use $. The table that summarise by populations was created
using average nested: loci -> individuals -> populations.
}

