% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_haplotypes.R
\name{summary_haplotypes}
\alias{summary_haplotypes}
\title{Haplotypes file summary}
\usage{
summary_haplotypes(data, strata, whitelist.markers = NULL,
  blacklist.id = NULL, pop.levels = NULL, pop.labels = NULL, read.length,
  parallel.core = detectCores() - 1)
}
\arguments{
\item{data}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{strata}{A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}. 
The \code{STRATA} column can be any hierarchical grouping. 
To create a strata file see \code{\link[stackr]{individuals2strata}}.
If you have already run 
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data, 
the strata file is similar to a stacks `population map file`, make sure you 
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
Default: \code{strata = NULL}.}

\item{whitelist.markers}{(optional) A whitelist of loci with a column name
'LOCUS'. The whitelist is located in the global environment or in the 
directory (e.g. "whitelist.txt"). 
If a whitelist is used, the files written to the directory will have 
'.filtered' in the filename. 
Default: \code{whitelist.markers = NULL}.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the global environment 
or in the directory (with "blacklist.txt").
Default: \code{blacklist.id = NULL}.}

\item{pop.levels}{An optional character string with your populations ordered.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{An optional character string with new populations names.
Default: \code{pop.labels = NULL}.}

\item{read.length}{(number) The length in nucleotide of your reads 
(e.g. \code{read.length = 100}).}

\item{parallel.core}{(optional) The number of core for parallel
programming during Pi calculations. 
Default: \code{parallel.core = detectCores() - 1}.}
}
\value{
The function returns a list with:
\enumerate{
\item $summary
\item $paralogs.pop
\item $paralogs.loci
\item $consensus.pop
\item $consensus.loci
\item $fh.pi.individuals: the individual's info for FH and Pi
}
Also in the list 3 plots:
\enumerate{
\item $scatter.plot
\item $boxplot.pi
\item $boxplot.fh
}
use $ to access each #' objects in the list.
The function potentially write 3 files in the working directory:
blacklist of unique putative paralogs and unique consensus loci 
and a summary of the haplotypes file by population.
}
\description{
STACKS batch_x.haplotypes.tsv file summary.
The output of the function is a summary table for populations with:
\enumerate{
\item putative sequencing errors: loci with > 2 alleles and/or potential 
paralogs. Genotypes with more than 2 alleles are 
erased before estimating the subsequent statistics.
\item consensus loci
\item monomorphic loci
\item polymorphic loci
\item haplotypes statistics for the observed and expected homozygosity and
heterozygosity.
\item Wright’s inbreeding coefficient (Fis)
\item \strong{IBDG}: a proxy measure of the realized proportion of the genome
that is identical by descent
\item \strong{FH measure}: based on the excess in the observed number of homozygous
genotypes within an individual relative to the mean number of homozygous 
genotypes expected under random mating (Keller et al., 2011; Kardos et al., 2015).
\item \strong{Pi}: the nucleotide diversity measured here consider the 
consensus loci in the catalog (no variation between population sequences).
}
}
\examples{
\dontrun{
sum <- summary_haplotypes(
data = "batch_1.haplotypes.tsv", 
strata = "strata_brook_charr.tsv", 
pop.levels = c("SKY", "LIM", "TWE"), 
read.length = 90)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com} and 
Anne-Laure Ferchaud \email{annelaureferchaud@gmail.com}
}
\references{
Keller MC, Visscher PM, Goddard ME (2011)
Quantification of inbreeding due to distant ancestors and its detection
 using dense single nucleotide polymorphism data. Genetics, 189, 237–249.

Kardos M, Luikart G, Allendorf FW (2015)
Measuring individual inbreeding in the age of genomics: marker-based
measures are better than pedigrees. Heredity, 115, 63–72.

Nei M, Li WH (1979)
Mathematical model for studying genetic variation in terms of 
restriction endonucleases. 
Proceedings of the National Academy of Sciences of 
the United States of America, 76, 5269–5273.
}

