% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_hapstats}
\alias{summary_hapstats}
\title{Import and summarise the batch_x.hapstats.tsv file.}
\usage{
summary_hapstats(data, pop.num, pop.col.types, pop.integer.equi, pop.levels)
}
\arguments{
\item{data}{The 'batch_x.hapstats.tsv' created by STACKS.}

\item{pop.col.types}{Integer or Character used in STACKS populations module?}

\item{pop.integer.equi}{When Integer was used for your population id,
give the character equivalence}

\item{pop.levels}{A character string with your populations in order.}

\item{number.pop}{The number of populations analysed.}
}
\description{
Import and summarise the batch_x.hapstats.tsv file.
Necessary preparation for density distribution and box plot figures.
}

