% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_polymorphism_haplotypes}
\alias{summary_polymorphism_haplotypes}
\title{Haplotypes summary}
\usage{
summary_polymorphism_haplotypes(haplotypes.file, pop.id.start, pop.id.end,
  pop.levels)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{pop.levels}{An optional character string with your populations ordered.}
}
\value{
The function returns a list with the summary, the paralogs and
consensus loci by populations and unique loci (use $ to access each
components).
Write 3 files in the working directory:
blacklist of unique putative paralogs and unique consensus loci.
and a summary of the haplotypes file by population.
}
\description{
STACKS batch_x.haplotypes.tsv file summary.
Output summary table for populations with putative paralogs,
consensus, monomorphic and polymorphic loci.
}
\details{
haplo.summary <- haplotype.file.summary$summary
paralogs.pop <- haplotype.file.summary$paralogs.pop
paralogs.loci <- haplotype.file.summary$paralogs.loci
consensus.pop <- haplotype.file.summary$consensus.pop
consensus.loci <- haplotype.file.summary$consensus.loci
}

