% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_reads.R
\name{summary_reads}
\alias{summary_reads}
\title{Summarise the reads for indel and GC content and produce the read depth plot}
\usage{
summary_reads(
  fq.files,
  output.dir = NULL,
  read.depth.plot = TRUE,
  min.coverage.fig = 7L,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{fq.files}{(character, path) The path to the individual fastq file to check,
or the entire fq folder.}

\item{output.dir}{(path) Write the output in a specific directory.
Default: \code{output.dir = NULL}, uses the working directory.}

\item{read.depth.plot}{(logical) To get the interesting summaries on
the read content, the function is very close to similar to
\code{\link{read_depth_plot}}, so you can produce it for each sample with
minimal cost on time.
Default: \code{read.depth.plot = TRUE}.}

\item{min.coverage.fig}{(character, path). Minimum coverage used to draw the
color on the figure.
Default: \code{min.coverage.fig = 7L}.}

\item{parallel.core}{(integer) Enable parallel execution with the number of threads.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
The function returns the read depth groups plot and the read stats overall
and by read depth groups.
}
\description{
Still in dev, but work nicely.
Summarise the reads for indel and GC content and produce the read depth plot
}
\examples{
\dontrun{
require(vroom)
sum <- summary_reads(fq.files = "my_fq_folder")
}
}
