% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rxstacks.R
\name{summary_rxstacks}
\alias{summary_rxstacks}
\title{Summarize STACKS rxstacks files}
\usage{
summary_rxstacks(rxstacks.folder, strata = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = FALSE, ...)
}
\arguments{
\item{rxstacks.folder}{(character). The path to the rxstacks output files.
e.g. \code{rxstacks.folder = "07_rxstacks_cstacks_sstacks_populations"}}

\item{strata}{(optional) Path to a strata file to get
the rxstacks results by population. With default, the results is presented
overall samples.
The strata used here is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
Additionnally, if you include the corresponding \code{SQL_ID}
column header of your samples,
the summary of the rxstacks haplotypes and snps files
will include the user friendly sample id (easy to get if you've used the stackr pipeline).
The \code{STRATA} column can be any hierarchical grouping.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS}, \code{STRATA} and optionally \code{SQL_ID}).
Default: \code{strata = NULL}.}

\item{parallel.core}{(integer, optional) The number of core used for parallel
execution.
Default: \code{parallel::detectCores() - 1}.}

\item{verbose}{(logical, optional) Make the function a little more chatty during
execution.
Default: \code{verbose = FALSE}.}

\item{...}{Used internally in stackr.}
}
\value{
The function returns a summary (data frame) containing:
\enumerate{
\item INDIVIDUALS: the sample id
\item LOCUS_NUMBER: the number of locus
\item BLACKLIST_rxstacks: the number of locus blacklisted by ustacks
\item FOR_CATALOG: the number of locus available to generate the catalog
\item BLACKLIST_ARTIFACT: the number of artifact genotypes (> 2 alleles, see details)
\item FILTERED: the number of locus after artifacts are removed
\item HOMOZYGOSITY: the number of homozygous genotypes
\item HETEROZYGOSITY: the number of heterozygous genotypes
\item MEAN_NUMBER_SNP_LOCUS: the mean number of SNP/locus (excluding the artifact locus)
\item MAX_NUMBER_SNP_LOCUS: the max number of SNP/locus observed for this individual (excluding the artifact locus)
\item NUMBER_LOCUS_4SNP: the number of locus with 4 or more SNP/locus (excluding the artifact locus)
}
}
\description{
This function reads the output of STACKS
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}
and summarise the information.
}
\examples{
\dontrun{
rxstacks.sum <- stackr::summary_rxstacks(
rxstacks.folder = "07_rxstacks_cstacks_sstacks_populations")

# To get the number of snp/locus for a specific individual:
snp.info <- ustacks.summary \%>\%
dplyr::filter(INDIVIDUALS == "ID2") \%>\%
dplyr::select(INDIVIDUALS, SNP_LOCUS) \%>\%
tidyr::unnest(.)

#Similarly, for the blacklisted locus (artifactual):
blacklist <- ustacks.summary \%>\%
dplyr::filter(INDIVIDUALS == "ID2") \%>\%
dplyr::select(INDIVIDUALS, BLACKLIST_ARTIFACT) \%>\%
tidyr::unnest(.)
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}
}
