% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_stats_vcf_tidy}
\alias{summary_stats_vcf_tidy}
\title{Summary statistics of a tidy VCF by population and markers}
\usage{
summary_stats_vcf_tidy(data)
}
\arguments{
\item{data}{The tidy VCF file created with read_stacks_vcf.}
}
\description{
Summarise and prepare the tidy VCF.
Summary, by population and markers (SNP), of frequency of the REF
and the ALT alleles, the observed and the expected heterozygosity
and the inbreeding coefficient. The Global MAF of Loci,
with STACKS GBS/RAD loci = read or de novo haplotypes,
is included and repeated over SNP.
}

