% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tsv2bam.R
\name{summary_tsv2bam}
\alias{summary_tsv2bam}
\title{Summary of catalog log likelihood}
\usage{
summary_tsv2bam(tsv2bam.output = "06_ustacks_cstacks_sstacks",
  verbose = TRUE)
}
\arguments{
\item{tsv2bam.output}{(character, path). Path to tsv2bam output files.
Default: \code{P = "06_ustacks_cstacks_sstacks"}.}

\item{verbose}{(optional) Make the function a little more chatty during
execution.
Default: \code{verbose = TRUE}.}
}
\value{
The function returns a list with the number of individuals, the batch ID number,
a summary data frame and a plot containing:
\enumerate{
\item INDIVIDUALS: the sample id
\item ALL_LOCUS: the total number of locus for the individual (shown in subplot A)
\item LOCUS: the number of locus with a one-to-one relationship (shown in subplot B)
with the catalog
\item MATCH_PERCENT: the percentage of locus with a one-to-one relationship
with the catalog (shown in subplot C)
}
}
\description{
This function reads the log file of
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{tsv2bam}
module and produce a summary data frame and plots with loci distribution.
This function is run automatically with
\pkg{stackr} \code{\link{run_tsv2bam}} function.
}
\examples{
\dontrun{
# with all the files in default folders, very simple:
sum <- stackr::summary_tsv2bam()
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{stacks Version 2.0Beta2}
}
