% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ustacks.R
\name{summary_ustacks}
\alias{summary_ustacks}
\title{Summarize STACKS ustacks files}
\usage{
summary_ustacks(
  ustacks.folder,
  parallel.core = parallel::detectCores() - 1,
  filename = NULL,
  verbose = FALSE
)
}
\arguments{
\item{ustacks.folder}{(character). The path to the ustacks output files.}

\item{parallel.core}{(integer, optional) The number of core used for parallel
execution.
Default: \code{parallel::detectCores() - 1}.}

\item{filename}{(optional) The filename prefix for the file written
in the working directory. Default: \code{filename = NULL}.
A default name will be used.}

\item{verbose}{(logical, optional) Make the function a little more chatty during
execution.
Default: \code{verbose = FALSE}.}
}
\value{
The function returns a summary (data frame) containing:
\enumerate{
\item INDIVIDUALS: the sample id
\item LOCUS_NUMBER: the number of locus
\item BLACKLIST_USTACKS: the number of locus blacklisted by ustacks
\item FOR_CATALOG: the number of locus available to generate the catalog
\item BLACKLIST_ARTIFACT: the number of artifact genotypes (> 2 alleles, see details)
\item FILTERED: the number of locus after artifacts are removed
\item HOMOZYGOSITY: the number of homozygous genotypes
\item HETEROZYGOSITY: the number of heterozygous genotypes
\item MEAN_NUMBER_SNP_LOCUS: the mean number of SNP/locus (excluding the artifact locus)
\item MAX_NUMBER_SNP_LOCUS: the max number of SNP/locus observed for this individual (excluding the artifact locus)
\item NUMBER_LOCUS_4SNP: the number of locus with 4 or more SNP/locus (excluding the artifact locus)
}
}
\description{
This function reads the output of
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks} folder
and summarise the information.

The information shown can help to choose individuals to include/exclude from
the catalog, the next step in
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{https://github.com/thierrygosselin/stackr#stacks-modules-and-radseq-typical-workflow}{pipeline}
(see example).

This function is run automatically inside \code{\link{run_ustacks}}.
}
\examples{
\dontrun{
ustacks.summary <- stackr::summary_ustacks(
ustacks.folder = "06_ustacks_cstacks_sstacks", verbose = TRUE)

# To get the number of snp/locus for a specific individual:
snp.info <- ustacks.summary \%>\%
dplyr::filter(INDIVIDUALS == "ID2") \%>\%
dplyr::select(INDIVIDUALS, SNP_LOCUS) \%>\%
tidyr::unnest(.)

#Similarly, for the blacklisted locus (artifactual):
blacklist <- ustacks.summary \%>\%
dplyr::filter(INDIVIDUALS == "ID2") \%>\%
dplyr::select(INDIVIDUALS, BLACKLIST_ARTIFACT) \%>\%
tidyr::unnest(.)

# Decide individuals to include in catalog:

# Make the ustacks.summary dataframe population wise by including pop info.
# For this we need a strata file ... it's similar to a population map, but with headers.
# It's a 2 columns files with INDIVIDUALS and STRATA as column header.
strata <- readr::read_tsv("strata.octopus.tsv", col_types = "cc") \%>\%
dplyr::rename(POP_ID = STRATA)

# join the dataframes
individuals.catalog <- dplyr::left_join(ustacks.summary, strata, by = "INDIVIDUALS") \%>\%
dplyr::arrange(desc(FILTERED))

# look at the FILTERED column
hist(individuals.catalog$FILTERED)
boxplot(individuals.catalog$FILTERED)
mean(individuals.catalog$FILTERED)

# lets say we want to filter out individuals below 45000 markers
individuals.catalog.filtered <- individuals.catalog \%>\%
dplyr::filter(FILTERED > 45000)

# number of ind per pop
dplyr::select(individuals.catalog.filtered, INDIVIDUALS, POP_ID) \%>\%
dplyr::group_by(POP_ID) \%>\% dplyr::tally(.)

# choose 20\% of individuals per pop to represent the catalog

individuals.catalog.select <- individuals.catalog.filtered \%>\%
dplyr::group_by(POP_ID) \%>\%
dplyr::sample_frac(tbl = ., size = 0.2, replace = FALSE) \%>\%
dplyr::ungroup(.) \%>\%
dplyr::arrange(desc(FILTERED)) \%>\%
dplyr::distinct(INDIVIDUALS, POP_ID)

# Using desc (from large to lower number of markers) ensure that
# the catalog is populated with samples with the highest number of loci first.
# This speed up the process in cstacks!


# Create a file with individuals and pop to use in
# cstacks or run_cstacks

readr::write_tsv(
x = individuals.catalog.select,
path = "population.map.octopus.samples.catalog.tsv",
col_names = FALSE) # stacks doesn't want column header
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks}
\href{http://catchenlab.life.illinois.edu/stacks/comp/cstacks.php}{cstacks}
\code{\link{run_cstacks}}
}
