% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stacks_modification.R
\name{sumstats_prep}
\alias{sumstats_prep}
\title{Import and prepare for analysis a batch_x.sumstats.tsv file produced by STACKS}
\usage{
sumstats_prep(sumstats, skip.line, pop.num, pop.col.types, pop.integer.equi,
  pop.levels, filename)
}
\arguments{
\item{sumstats}{The STACKS batch_x.susmtats.tsv file.}

\item{skip.line}{The number of line before reading the data (you need to count the header).}

\item{pop.num}{The number of populations.}

\item{pop.col.types}{(string) \code{"integer"} or \code{character} used in STACKS populations module?}

\item{pop.integer.equi}{When Integer was used for your population id, give the character equivalence}

\item{pop.levels}{A character string with the hierarchy/order of your populations.}

\item{filename}{The name of the file written in the directory.}
}
\description{
Import STACKS sumstats.tsv and add useful info (frequency of REF allele, Global MAF) for subsequent use. to a tidy format.
}

