% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_stacks_vcf.R
\name{tidy_vcf_summary}
\alias{tidy_vcf_summary}
\title{Tidy VCF summary.}
\usage{
tidy_vcf_summary(data)
}
\arguments{
\item{data}{The tidy VCF file.}
}
\description{
Summarise and prepare for filter the tidy VCF. By population, frequency of REF and ALT alleles, observed and expected heterozygosity and inbreeding coefficient. Global MAF is also for each marker.
}

