% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitelist_blacklist.R
\name{whitelist_loci}
\alias{whitelist_loci}
\title{Whitelist loci
This function creates a whitelist of loci, used after applying a filter
to keep track of the loci kept by a filter.}
\usage{
whitelist_loci(data, filename, col.header)
}
\arguments{
\item{data}{The data frame after the filter. Object
or file (using ".tsv") of class sumstats.}

\item{filename}{The name of the file written in the directory.}

\item{col.header}{TRUE and the loci will have a column header 'LOCUS'.}
}
\description{
Whitelist loci
This function creates a whitelist of loci, used after applying a filter
to keep track of the loci kept by a filter.
}

