% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitelist_blacklist.R
\name{whitelist_loci_snp}
\alias{whitelist_loci_snp}
\title{Whitelist loci and snp.}
\usage{
whitelist_loci_snp(data, filename, col.header)
}
\arguments{
\item{data}{The data frame after the filter. Object
or file (using ".tsv") of class sumstats.}

\item{filename}{The name of the file written in the directory.}

\item{col.header}{TRUE and the loci will have a column header 'LOCUS'.}
}
\description{
This function creates a whitelist of loci and snp,
useful in the populations module of STACKS.
}

