% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitelist_blacklist.R
\name{whitelist_loci_vcf}
\alias{whitelist_loci_vcf}
\title{Whitelist loci for VCF tools}
\usage{
whitelist_loci_vcf(data, filename)
}
\arguments{
\item{data}{The data frame after the filter. Object
or file (using ".tsv") of class sumstats.}

\item{filename}{The name of the file written in the directory.}
}
\description{
This function creates a whitelist of loci for VCF tools.
With 2 columns (CHROM and ID).
}

