% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_genepop.R
\name{write_genepop}
\alias{write_genepop}
\title{Use the batch_x.haplotypes.tsv file to write a genpop file}
\usage{
write_genepop(haplotypes.file, whitelist.loci = NULL, blacklist.id = NULL,
  genepop.filename, genepop.header, pop.levels, pop.id.start, pop.id.end)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{whitelist.loci}{(optional) A whitelist of loci and
a column header 'LOCUS'.
The whitelist is in the directory (e.g. "whitelist.txt").}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the directory
 (e.g. "blacklist.txt").}

\item{genepop.filename}{The name of the file written to the directory.}

\item{genepop.header}{The first line of the Genepop file
(ex: genepop of 10 pop of Acipenser fulvescens filtered MAF 0.05).}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}
}
\description{
This function can first filter the haplotypes file
with a whitelist of loci
and a blacklist of individuals (optional).
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Rousset F (2008) genepop’007: a complete re-implementation
of the genepop software for Windows and Linux.
Molecular Ecology Resources, 8, 103-106.

Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}

