% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtypes.R
\name{write_gtypes}
\alias{write_gtypes}
\title{Tidy genomic data to \code{\link[strataG]{gtypes}}}
\usage{
write_gtypes(data)
}
\arguments{
\item{data}{A file in the working directory or object in the global environment 
in wide or long (tidy) formats. See details for more info.}
}
\value{
An object of the class \code{\link[strataG]{gtypes}} is returned.
}
\description{
Write a \code{\link[strataG]{gtypes}} object from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and might be of interest for users.
}
\details{
\strong{Details for Input data:}

To discriminate the long from the wide format, 
the function \pkg{stackr} \code{\link[stackr]{read_long_tidy_wide}} searches 
for "MARKERS" in column names (TRUE = long format).
The data frame is tab delimitted.
\strong{Wide format:}
The wide format cannot store metadata info.
The wide format starts with these 2 id columns: 
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping of individuals), 
the remaining columns are the markers in separate columns storing genotypes.

\strong{Long/Tidy format:}
The long format is considered to be a tidy data frame and can store metadata info. 
(e.g. from a VCF see \pkg{stackr} \code{\link{tidy_genomic_data}}). A minimum of 4 columns
are required in the long format: \code{INDIVIDUALS}, \code{POP_ID}, 
\code{MARKERS} and \code{GENOTYPE or GT}. The rest are considered metata info.

\strong{2 genotypes formats are available:}
6 characters no separator: e.g. \code{001002 of 111333} (for heterozygote individual).
6 characters WITH separator: e.g. \code{001/002 of 111/333} (for heterozygote individual).
The separator can be any of these: \code{"/", ":", "_", "-", "."}.

\emph{How to get a tidy data frame ?}
\pkg{stackr} \code{\link{tidy_genomic_data}} can transform 6 genomic data formats 
in a tidy data frame.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Eric Archer, Paula Adams and Brita Schneiders (2016). 
strataG: Summaries and Population Structure Analyses of
Genetic Data. R package version 1.0.5. https://CRAN.R-project.org/package=strataG
}
\seealso{
\code{strataG.devel} is available on github \url{https://github.com/EricArcher/}
}

