% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_sample.R
\name{write_normalize}
\alias{write_normalize}
\title{write_normalize}
\usage{
write_normalize(number.samples, subsample.random, fastq.files)
}
\description{
Write the normalize fastq file
}
\keyword{internal}
