% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Used internally in stackr to write a vcf file from a tidy 
data frame}
\usage{
write_vcf(data, pop.info = FALSE, filename = NULL)
}
\arguments{
\item{data}{A file in the working directory or object in the global environment 
in wide or long (tidy) formats. See details for more info.}

\item{pop.info}{(optional, logical) Should the population information be 
included in the FORMAT field (along the GT info for each samples ?). To make
the VCF population-ready use \code{pop.info = TRUE}. The populatio information
must be included in the \code{POP_ID} column of the tidy dataset.
Default: \code{pop.info = FALSE}.}

\item{filename}{(optional) The file name prefix for the vcf file 
written to the working directory. With default: \code{filename = NULL}, 
the date and time is appended to \code{stackr_vcf_file_}.}
}
\description{
Write a vcf file from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and might be of interest for users.
}
\details{
\strong{Input data:}
 
To discriminate the long from the wide format, 
the function \pkg{stackr} \code{\link[stackr]{read_long_tidy_wide}} searches 
for \code{MARKERS or LOCUS} in column names (TRUE = long format).
The data frame is tab delimitted.
\strong{Wide format:}
The wide format cannot store metadata info.
The wide format starts with these 2 id columns: 
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping of individuals), 
the remaining columns are the markers in separate columns storing genotypes.

\strong{Long/Tidy format:}
The long format is considered to be a tidy data frame and can store metadata info. 
(e.g. from a VCF see \pkg{stackr} \code{\link{tidy_genomic_data}}). A minimum of 4 columns
are required in the long format: \code{INDIVIDUALS}, \code{POP_ID}, 
\code{MARKERS or LOCUS} and \code{GENOTYPE or GT}. The rest are considered metata info.

\strong{2 genotypes formats are available:}
6 characters no separator: e.g. \code{001002 of 111333} (for heterozygote individual).
6 characters WITH separator: e.g. \code{001/002 of 111/333} (for heterozygote individual).
The separator can be any of these: \code{"/", ":", "_", "-", "."}.

\emph{How to get a tidy data frame ?}
\pkg{stackr} \code{\link{tidy_genomic_data}} can transform 6 genomic data formats 
in a tidy data frame.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}

