% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_monomorphic_markers.R
\name{discard_monomorphic_markers}
\alias{discard_monomorphic_markers}
\title{Discard monomorphic markers}
\usage{
discard_monomorphic_markers(data, verbose = FALSE)
}
\arguments{
\item{data}{A tidy genomic data set in the working directory tidy formats.
\emph{How to get a tidy data frame ?}
Look for \pkg{stackr} \code{\link{tidy_genomic_data}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty 
during execution. 
Default: \code{verbose = FALSE}.}
}
\value{
A list with the filtered input file and the blacklist of markers removed.
}
\description{
Discard monomorphic markers.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

