% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{erase_genotypes}
\alias{erase_genotypes}
\title{Erase genotypes in the batch_x.haplotypes.tsv that
didn't pass min coverage threshold filter.}
\usage{
erase_genotypes(haplotypes.file, blacklist.genotypes, filename)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{blacklist.genotypes}{A blacklist of loci and genotypes
containing at least 2 columns header 'LOCUS' and 'SAMPLES'.}

\item{filename}{The filename saved to the working directory.}
}
\description{
This function modify the 'batch_x.haplotypes.tsv' file
with a blacklist of loci and individuals.
}

