% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignment.R
\name{figure_assignment}
\alias{figure_assignment}
\title{Cleveland dot plot figure of assignement results.}
\usage{
figure_assignment(assignment.summary)
}
\arguments{
\item{assignment.summary}{The assignment summary file created
with assignment_genodive function.}
}
\description{
GGPLOT2 Cleveland dot plot figure of assignment results.
The figure will need some work in Adobe Illustrator or similar sofware.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Meirmans PG, Van Tienderen PH (2004) genotype and genodive:
two programs for the analysis of genetic diversity of asexual organisms.
Molecular Ecology Notes, 4, 792-794.
}

