% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_genotype_likelihood.R
\name{filter_genotype_likelihood}
\alias{filter_genotype_likelihood}
\title{Genotype likelihood filter}
\usage{
filter_genotype_likelihood(tidy.vcf.file, allele.min.depth.threshold,
  read.depth.max.threshold, gl.mean.threshold, gl.min.threshold,
  gl.diff.threshold, pop.threshold, percent, filename)
}
\arguments{
\item{tidy.vcf.file}{A data frame object or file (using ".tsv")
of class tidy VCF.}

\item{allele.min.depth.threshold}{Threshold number
of min depth of REF or ALT alleles.}

\item{read.depth.max.threshold}{Threshold number
of maximum read depth.}

\item{gl.mean.threshold}{Threshold number of mean genotype likelihood.}

\item{gl.min.threshold}{Threshold number of min genotype likelihood.}

\item{gl.diff.threshold}{Threshold number of diff genotype likelihood,
the difference between the max and min GL over a loci/read/haplotype.}

\item{pop.threshold}{A threshold number: proportion, percentage
or fixed number e.g. 0.70, 70 or 15.}

\item{percent}{Is the threshold a percentage ? TRUE or FALSE.}

\item{filename}{Name of the file written to the working directory.}
}
\description{
Filter markers based coverage and/or genotype likelihood
using a tidy VCF file.
}
\details{
The summary statistics are averaged
and nested SNP -> individuals -> population -> loci. e.g. the mean GL is the average
genotype likelihood for all individuals of pop x for loci x.
Here the filter focus on global trend accross pop for individual loci.
The gl.diff.threshold can spot big difference in GL for a loci, e.g. when
3 SNP have similar GL the difference will be close to 0, higher value, higher
the difference between the SNP inside the loci.
}

