% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_individual.R
\name{filter_individual}
\alias{filter_individual}
\title{Individual filter}
\usage{
filter_individual(data, population.map, pop.id.start, pop.id.end, pop.levels,
  pop.labels, ind.threshold, threshold.fixed, filename)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv")
of class sumstats or tidy VCF.}

\item{population.map}{With tidy VCF (optional). With sumstats (REQUIRED).
The population map or individuals listed in one column.
No headers.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.levels}{An character string with your populations ordered.}

\item{pop.labels}{An optional character string with new populations names.}

\item{ind.threshold}{The individual threshold, proportion, percentage or 
number e.g. 0.70, 70, 15.}

\item{threshold.fixed}{Is the threshold fixed ? TRUE or FALSE.}

\item{filename}{Name of the file written to the working directory.}
}
\description{
Filter individuals number, percentage or fix threshold
in a sumstats prepared file or a tidy VCF file.
}

