% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_individual.R
\name{filter_individual}
\alias{filter_individual}
\title{Individual filter}
\usage{
filter_individual(tidy.vcf, pop.id.start, pop.id.end, pop.levels, pop.labels,
  ind.threshold, percent, filename)
}
\arguments{
\item{tidy.vcf}{A tidy vcf object or file (using ".tsv").}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.levels}{An character string with your populations ordered.}

\item{pop.labels}{An optional character string with new populations names.}

\item{ind.threshold}{The individual threshold, proportion, percentage or 
number e.g. 0.70, 70, 15.}

\item{percent}{Is the threshold a percentage? TRUE or FALSE.
This argument is necessary to distinguish percentage from integer individual
threshold (e.g. 70 percent or 70 individuals).}

\item{filename}{Name of the file written to the working directory.}
}
\description{
Filter individuals number, percentage or fix threshold
from a tidy VCF file.
}

