% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_maf.R
\name{filter_maf}
\alias{filter_maf}
\title{MAF filter}
\usage{
filter_maf(data, local.maf.threshold, global.maf.threshold, pop.threshold,
  filename)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv")
of class sumstats or a tidy VCF summarised.}

\item{local.maf.threshold}{Number.}

\item{global.maf.threshold}{Number.}

\item{pop.threshold}{Fixed number of pop required to keep the locus.}

\item{filename}{Name of the file written to the working directory (optional).}
}
\description{
Minor Allele Frequency filter from a sumstats prepared file
or a tidy VCF file.
}
\details{
To help choose a threshold for the local and global MAF, look
at the function \link{diagnostic_maf}
}
\seealso{
\link{sumstats_prep}

\link{summary_stats_vcf_tidy}
}

