% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haplo2genind.R
\name{haplo2genind}
\alias{haplo2genind}
\title{Convert between batch_x.haplotypes.tsv and \code{genind} objects}
\usage{
haplo2genind(haplotypes.file, whitelist.loci = NULL, blacklist.id = NULL,
  pop.levels, pop.id.start, pop.id.end, strata = NULL, hierarchy = NULL)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{whitelist.loci}{(optional) A whitelist of loci and
a column header 'LOCUS'.
The whitelist is in the directory (e.g. "whitelist.txt").}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the directory
 (e.g. "blacklist.txt").}

\item{pop.levels}{(Optional) character string with your populations ordered.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{strata}{an optional data frame that defines population stratifications
for your samples. This is especially useful if you have a hierarchical or
factorial sampling design. See adegenet for details.}

\item{hierarchy}{a hierarchical formula that explicitely defines hierarchical
levels in your strata. See adegenet for details.}
}
\value{
an object of the class \code{genind}.
}
\description{
This function can first filter the haplotypes file
with a whitelist of loci
and a blacklist of individuals (optional). Then it will convert the file
to a adegenet \code{genind} object.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.

Jombart T (2008) adegenet: a R package for the multivariate
analysis of genetic markers. Bioinformatics, 24, 1403-1405.

Jombart T, Ahmed I (2011) adegenet 1.3-1:
new tools for the analysis of genome-wide SNP data.
Bioinformatics, 27, 3070-3071.
}
\seealso{
adegenet is available on CRAN and https://github.com/thibautjombart/
}

