% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_tools.R
\name{make_tfam}
\alias{make_tfam}
\title{Write a PLINK tfam file from a population map.}
\usage{
make_tfam(population.map, pop.id.start, pop.id.end, pop.levels, filename)
}
\arguments{
\item{population.map}{The population map or individuals listed in one column. No headers.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{filename}{The name of the file written in the directory.}
}
\description{
Make a tfam file for PLINK. Useful in ADMIXTURE for bootstrapping.
}

