% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blacklist_id.R
\name{missing_genotypes}
\alias{missing_genotypes}
\title{Missing genotypes information and blacklist of individuals based
on a threshold of missing genotype.}
\usage{
missing_genotypes(haplotypes.file, whitelist.loci = NULL, pop.id.start,
  pop.id.end, missing.geno.threshold)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{whitelist.loci}{(optional) A whitelist of loci with a column header
'LOCUS'. If the whitelist is written in the directory
\code{whitelist.loci = "whitelist.txt"}. If the whitelist is in
the global environment \code{whitelist.loci = whitelist.1000loci}}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{missing.geno.threshold}{(integer) Percentage of missing genotype
allowed per individuals. e.g. for a maximum of 30% of missing genotype
per individuals \code{missing.geno.threshold = 30}.}
}
\value{
a list with 4 data frames: $missing.genotypes.ind,
$missing.genotypes.pop, $blacklisted.id, $plot.missing.
}
\description{
Missing genotypes information summary per individuals and
per population. Create a blacklist of individuals based
on a threshold of missing genotype.
This function accept a whitelist of loci to create a blacklist
of individuals before or after filtering of loci.
Paralogs are automatically removed from STACKS haplotype file.
}
\details{
For the plot, to see the information with the popualtion in
different facet, use \code{+facet_wrap(~POP_ID, nrow=2,ncol=5)}
after the object of the plot, e.g. \code{fig <- missing$plot.missing}, to have
facet by pop \code{fig +facet_wrap(~POP_ID, nrow=2,ncol=5)}
where \code{nrow} and \code{ncol} in this exeample would spread
the 10 populations on 2 rows and 5 columns.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

