% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blacklist_id.R
\name{missing_genotypes}
\alias{missing_genotypes}
\title{Missing genotypes information and blacklist of individuals based 
on a threshold of missing genotype.}
\usage{
missing_genotypes(haplotypes.file, whitelist.loci = NULL, erase, pop.id.start,
  pop.id.end, pop.levels, pop.labels, missing.geno.threshold)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{whitelist.loci}{(optional) A whitelist of loci with a column header 
'LOCUS'. If the whitelist is written in the directory 
\code{whitelist.loci = "whitelist.txt"}. If the whitelist is in 
the global environment \code{whitelist.loci = whitelist.1000loci}}

\item{erase}{(character) Default when missing argument = \code{"loci"}. Loci with more than 2 alleles 
(paralogs and/or sequencing errors) will be removed. \code{"genotypes"}: genotypes 
with more than 2 alleles by individual (paralogs and/or sequencing errors) will be erased.
Keeping the loci for other individuals.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.levels}{A character string with your populations ordered.}

\item{pop.labels}{(optional character string) If you need to rename 
pop/sampling sites in \code{pop.levels} or combined sites/pop
into a different names, here is the place (e.g. 3 sites: 
c("ONT", "QUE", "SAS") and you want to combine "ONT" and "SAS" into the same 
grouping "CAN" use : c("CAN", "QUE", "CAN").}

\item{missing.geno.threshold}{(integer) Percentage of missing genotype 
allowed per individuals. e.g. for a maximum of 30% of missing genotype 
per individuals \code{missing.geno.threshold = 30}.}
}
\value{
a list with 4 data frames: $missing.genotypes.ind,
$missing.genotypes.pop, $blacklisted.id, $plot.missing.
}
\description{
Use this function to have a summary of missing genotypes 
per individuals and per population. Density distribution plot of misssing
information is produced. This function also create a blacklist
of individuals based on the desired threshold of missing genotype. 
When a whitelist of loci is provided, \code{_filtered} will be appended 
to the filename. Locus with more than 2 alleles by individual (paralogs and/or sequencing errors) can be deleted or 
genotypes with more than 2 alleles by individual can be erased.
}
\details{
For the plot, to see the information with the popualtion in 
different facet, use \code{+facet_wrap(~POP_ID, nrow=2,ncol=5)} 
after the object of the plot, e.g. \code{fig <- missing$plot.missing}, to have
facet by pop \code{fig +facet_wrap(~POP_ID, nrow=2,ncol=5)}
where \code{nrow} and \code{ncol} in this exeample would spread 
the 10 populations on 2 rows and 5 columns.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

