% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_ibm.R
\name{plink_ibm}
\alias{plink_ibm}
\title{Identity-by-missingness (IBM) analysis.}
\usage{
plink_ibm(vcf.file, whitelist = NULL, denovo, blacklist.id, strata,
  strata.select)
}
\arguments{
\item{vcf.file}{The VCF file created by STACKS.}

\item{whitelist}{(optional) A whitelist of markers to keep, 
for VCFTools 2 columns are required: 'CHROM' and 'POS'. Default = \code{NULL}.}

\item{denovo}{(logical) Default \code{TRUE}. For de novo analysis the vcf file 
created by stacks as a \code{un} in the \code{CHROM} column. We need to change that 
internally for the function to work in VCFTools and PLINK.}

\item{blacklist.id}{(optional) A blacklist with individual ID. NO 
column header. The blacklist is in the directory (e.g. "blacklist.id.txt").}

\item{strata}{(optional) File with column header 'INDIVIDUALS' and any other 
columns containing e.g. population id, sequencer id, lanes, etc. 
Any info that might impact missing data. 
Willl be used in \code{strata.select} argument.}

\item{strata.select}{Control the colour in ggplot2 plot. 
e.g. use \code{"POP_ID")} to see the strata column \code{"POP_ID"} 
or \code{"LANES"}, if you have a column in your strata file that starts 
with LANES (for sequencing lane...).}
}
\value{
The function returns a list with the identity-by-missingness results ($ibm) and 
a MDS plot (MultiDimensional Scaling) of ibm results ($plot)
In the working directory:
the modified vcf file, if de novo was selected, plink tfiles and ibm results.
}
\description{
Diagnose your missing genotypes pattern with PLINK's 
identity-by-missingness (IBM) analysis.
}
\examples{
\dontrun{
If the object for the function is 'ibm' then:
to access the figure
ibm.plot <- ibm$plot
The ibm results of plink in a dataframe with your strata:
ibm.results <- ibm$ibm
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Purcell S, Neale B, Todd-Brown K et al. (2007) 
PLINK: a tool set for whole-genome association and population-based linkage 
analyses. American Journal of Human Genetics, 81, 559–575.

Catchen JM, Amores A, Hohenlohe PA et al. (2011) 
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences. 
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013) 
Stacks: an analysis tool set for population genomics. 
Molecular Ecology, 22, 3124-3140.
}

