% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{population_filter}
\alias{population_filter}
\title{Population sumstats filter}
\usage{
population_filter(data, pop.threshold, percent)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv") of a tidy VCF.}

\item{population.map}{The population map or individuals listed in one column.
No headers.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{ind.threshold}{The individual threshold, proportion, percentage or
number e.g. 0.70, 70, 15.}

\item{threshold.fixed}{Is the threshold fixed ? TRUE or FALSE.}
}
\description{
Filter a tidy VCF file based
on individual number, percentage or fix threshold.
}

