% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_stacks_vcf.R
\name{read_stacks_vcf}
\alias{read_stacks_vcf}
\title{Read a VCF file produced by STACKS and transform in tidy format}
\usage{
read_stacks_vcf(vcf.file, pop.id.start, pop.id.end, pop.levels, whitelist,
  blacklist, filename)
}
\arguments{
\item{vcf.file}{The VCF file created by STACKS.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{whitelist}{An optional filter of loci can be applied to the vcf, using
a file in the working directory (e.g. "myfile.txt") or an object
in the global environment.}

\item{blacklist}{An optional filter of loci can be applied to the vcf, using
a file in the working directory (e.g. "myfile.txt") or an object
in the global environment.}

\item{filename}{The name of the file written in the directory.}
}
\description{
Import a VCF file created by STACKS and mofify to a tidy format.
}

