% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gstacks.R
\name{run_gstacks}
\alias{run_gstacks}
\title{Run STACKS new module called gstacks}
\usage{
run_gstacks(
  denovo = TRUE,
  P = "06_ustacks_2_gstacks",
  M = "02_project_info/population.map.tsv2bam.tsv",
  I = NULL,
  B = NULL,
  O = NULL,
  parallel.core = parallel::detectCores() - 1,
  model = "marukilow",
  var.alpha = 0.01,
  gt.alpha = 0.05,
  rm.pcr.duplicates = FALSE,
  rm.unpaired.reads = FALSE,
  ignore.pe.reads = TRUE,
  unpaired = TRUE,
  kmer.length = 31,
  max.debruijn.reads = 1000,
  min.kmer.cov = 2,
  write.alignments = FALSE,
  min.mapq = 10,
  max.clipped = 0.2,
  max.insert.len = 1000,
  details = FALSE,
  phasing.cooccurrences.thr.range = c(1, 2),
  phasing.dont.prune.hets = FALSE,
  h = FALSE
)
}
\arguments{
\item{denovo}{(logical) de novo vs reference-based mode.
Default: \code{denovo = TRUE}.}

\item{P}{(path, character) De novo mode.
Path to the directory containing STACKS files.
Default: \code{P = "06_ustacks_2_gstacks"}.
Inside the folder, you should have:
\itemize{
  \item \strong{the catalog files:} starting with \code{batch_} and ending with
  \code{.alleles.tsv.gz, .snps.tsv.gz, .tags.tsv.gz and .bam};
  \item \strong{5 files for each samples:} The sample name is the prefix for
  the files ending with:
\code{.alleles.tsv.gz, .models.tsv.gz, .snps.tsv.gz, .tags.tsv.gz and .bam}.
Those files are created in the
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks},
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks},
\href{http://catchenlab.life.illinois.edu/stacks/comp/cxstacks.php}{cxstacks},
\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{tsv2bam}
modules
}.}

\item{M}{(path, character) Path to a population map giving the list of samples.}

\item{I}{(character, path) Reference-based mode.
Input directory containing BAM files.
Default: \code{I = NULL}.}

\item{B}{(character, path) Reference-based mode. Path to input BAM files.
The input BAM file(s) must be sorted by coordinate.
With \code{B}, records must be assigned to samples using BAM "reads groups"
(gstacks uses the ID/identifier and SM/sample name fields). Read groups
must be consistent if repeated different files. With \code{I},
read groups are unneeded and ignored.
Please refer to the gstacks manual page for information. stacks also provide
information about how to
generate such a BAM file with Samtools or Sambamba, and examples.
Default: \code{B = NULL}.}

\item{O}{(character, path) Path to output directory. \code{stackr} by default
will output in the input directory.
Default: \code{O = NULL} (same as input).}

\item{parallel.core}{(integer) Enable parallel execution with the number of threads.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{model}{(character) The model to use to call variants and genotypes;
one of \code{"marukilow"}, \code{"marukihigh"}, or \code{"snp"}.
See ref for more details on algorithms.
Default: \code{model = "marukilow"}.}

\item{var.alpha}{(double) Alpha threshold for discovering SNPs.
Default: \code{var.alpha = 0.01}.}

\item{gt.alpha}{(double) Alpha threshold for calling genotypes.
Default: \code{gt.alpha = 0.05}.}

\item{rm.pcr.duplicates}{(logical) remove all but one set of read pairs of
the same sample that have the same insert length
(implies \code{rm.unpaired.reads = TRUE}.).
Default: \code{rm.pcr.duplicates = FALSE}.}

\item{rm.unpaired.reads}{(logical) Discard unpaired reads.
Default: \code{rm.unpaired.reads = FALSE}.}

\item{ignore.pe.reads}{(logical) With default the function will
ignore paired-end reads even if present in the input.
Default: \code{ignore.pe.reads = TRUE}.}

\item{unpaired}{(logical) Ignore read pairing (only for paired-end GBS;
treat READ2's as if they were READ1's).
Default: \code{unpaired = TRUE}.}

\item{kmer.length}{(integer) De novo mode.
kmer length for the de Bruijn graph. For expert.
Default: \code{kmer.length = 31}.}

\item{max.debruijn.reads}{(integer) Maximum number of reads to use in the
de Bruijn graph. For expert.
Default: \code{max.debruijn.reads = 1000}.}

\item{min.kmer.cov}{(integer) De novo mode.
Minimum coverage to consider a kmer. For expert.
Default: \code{min.kmer.cov = 2}.}

\item{write.alignments}{(logical) Save read alignments (heavy BAM files).
Default: \code{write.alignments = FALSE}.}

\item{min.mapq}{(double) Reference-based mode.
Minimum PHRED-scaled mapping quality to consider a read.
Default: \code{min.mapq = 10}.}

\item{max.clipped}{(double) Reference-based mode.
Maximum soft-clipping level, in fraction of read length.
Default: \code{max.clipped = 0.20}.}

\item{max.insert.len}{(integer) Reference-based mode.
Maximum allowed sequencing insert length
Default: \code{max.insert.len = 1000}.}

\item{details}{(logical) With default the function will write a more detailed
output.
Default: \code{details = FALSE}.}

\item{phasing.cooccurrences.thr.range}{(integer) range of edge coverage thresholds to
iterate over when building the graph of allele cooccurrences for
SNP phasing.
Default: \code{phasing.cooccurrences.thr.range = c(1,2)}.}

\item{phasing.dont.prune.hets}{(logical) Don't try to ignore dubious heterozygote
genotypes during phasing. By default, during phasing, dubious het are ignored.
Default: \code{phasing.dont.prune.hets = FALSE}.}

\item{h}{Display this help messsage.
Default: \code{h = FALSE}}
}
\value{
\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{tsv2bam}
returns a set of \code{.matches.bam} files.

The function \code{run_gstacks} returns a list with the number of individuals, the batch ID number,
a summary data frame and a plot containing:
\enumerate{
\item INDIVIDUALS: the sample id
\item ALL_LOCUS: the total number of locus for the individual (shown in subplot A)
\item LOCUS: the number of locus with a one-to-one relationship (shown in subplot B)
with the catalog
\item MATCH_PERCENT: the percentage of locus with a one-to-one relationship
with the catalog (shown in subplot C)

Addtionally, the function returns a batch_X.catalog.bam file that was generated
by merging all the individual BAM files in parallel.
}
}
\description{
Run \href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/stacks_v2.php}{gstacks}
module inside R!
}
\examples{
\dontrun{
# The simplest form of the function with De novo data:
bam.sum <- stackr::run_gstacks() # that's it !
# This will use, by default, the same population map used in run_tsv2bam
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.

Maruki T, Lynch M (2017)
Genotype Calling from Population-Genomic Sequencing Data. G3, 7, 1393-1404.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
}
