% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_populations.R
\name{run_populations}
\alias{run_populations}
\title{Run STACKS Version >= 2.0 populations module}
\usage{
run_populations(P = "06_ustacks_cstacks_sstacks", V = NULL,
  O = "07_populations", M, t = parallel::detectCores() - 1,
  batch_size = 10000, p = 1, r = 0.3, min_maf = NULL,
  min_mac = NULL, max_obs_het = NULL, lnl_lim = NULL,
  write_single_snp = FALSE, write_random_snp = FALSE, B = NULL,
  W = NULL, e = NULL, merge_sites = FALSE, merge_prune_lim = NULL,
  hwe = FALSE, fstats = FALSE, fst_correction = NULL,
  p_value_cutoff = 0.05, k = FALSE, smooth_fstats = FALSE,
  smooth_popstats = FALSE, sigma = 150000, bootstrap = FALSE,
  N = 100, bootstrap_pifis = FALSE, bootstrap_fst = FALSE,
  bootstrap_div = FALSE, bootstrap_phist = FALSE,
  bootstrap_wl = NULL, ordered_export = FALSE, fasta_samples = FALSE,
  fasta_samples_raw = FALSE, fasta_loci = FALSE, vcf = TRUE,
  genepop = FALSE, structure = FALSE, radpainter = FALSE,
  phase = FALSE, fastphase = FALSE, plink = FALSE, hzar = FALSE,
  phylip = FALSE, phylip_var = FALSE, phylip_var_all = FALSE,
  treemix = FALSE, no_hap_exports = FALSE, h = FALSE,
  verbose = FALSE, v = FALSE, log_fst_comp = FALSE)
}
\arguments{
\item{P}{(path, character) Path to the directory containing all the STACKS files.
Default: \code{P = "06_ustacks_cstacks_sstacks"}.}

\item{V}{(character) Path to an input VCF file. When this module is used to
filter an existing vcf file.
Default: \code{V = NULL}.}

\item{O}{(character) Path to a directory where to write the output files.
With default: \code{O = "07_populations"}, the function creates a folder inside
\code{07_populations}, with date and time appended to \code{populations_}.}

\item{M}{path to a population map file. The format is a tab-separated file,
with first column containing sample name and second column population id.
No heather (column name).
e.g. \code{M = "02_project_info/population.map.catalog.tsv"}}

\item{t}{(integer) enable parallel execution with num_threads threads.
Default: \code{t = parallel::detectCores() - 1}}

\item{batch_size}{(integer) The number of loci (de novo mode) or
chromosome (reference mode), to process in a batch.
Increase to speed analysis, uses more memory, decrease to save memory).
Default in de novo mode (loci/batch): \code{batch_size = 10000}.
Default in reference mode (chromosome/batch): \code{batch_size = 1}.}

\item{p}{(integer) Minimum number of populations a locus must be present in to process a locus.
Default: \code{p = 1}.}

\item{r}{(double) Minimum percentage of individuals in a population required to process a locus for that population.
Default: \code{r = 0.3}.}

\item{min_maf}{(double) Specify a minimum minor allele frequency required to process a nucleotide site at a locus (0 < min_maf < 0.5).
Default: \code{min_maf = NULL}. Recommendation: use my other R package RADIATOR to filter data based on MAF.}

\item{min_mac}{(integer) Specify a minimum minor allele count required to process a nucleotide site at a locus.
Default: \code{min_mac = NULL}.}

\item{max_obs_het}{(double) Specify a maximum observed heterozygosity required to process a nucleotide site at a locus.
Default: \code{max_obs_het = NULL}.  Recommendation: use my other R package RADIATOR to filter data based on heterozygosity.}

\item{lnl_lim}{(integer) Filter loci with log likelihood values below this threshold.
Default: \code{lnl_lim = NULL}.}

\item{write_single_snp}{(logical) Restrict data analysis to only the first SNP per locus (implies --no-haps).
Default: \code{write_single_snp = FALSE}.}

\item{write_random_snp}{(logical) Restrict data analysis to one random SNP per locus (implies --no-haps).
Default: \code{write_random_snp = FALSE}.}

\item{B}{(character) Path to a file containing Blacklisted markers to be excluded from the export.
Default: \code{B = NULL}.}

\item{W}{(character) Path to a file containing Whitelisted markers to include in the export.
Default: \code{W = NULL}.}

\item{e}{(character) Restriction enzyme name.
Default: \code{e = NULL}.}

\item{merge_sites}{(logical) Merge loci that were produced from the same restriction enzyme cutsite (requires reference-aligned data).
Default: \code{merge_sites = FALSE}.}

\item{merge_prune_lim}{(integer) When merging adjacent loci, if at least X% samples posses both loci prune the remaining samples out of the analysis.
Default: \code{merge_prune_lim = NULL}.}

\item{hwe}{(logical) Calculate divergence from Hardy-Weinberg equilibrium
using the exact test at the SNP level and Guo and Thompson MCMC algorithm at
the haplotype level.
Default: \code{hwe = FALSE}.}

\item{fstats}{(logical) Enable SNP and haplotype-based F statistics.
Default: \code{fstats = FALSE}.}

\item{fst_correction}{(character) Specify a correction to be applied to Fst values: 'p_value', 'bonferroni_win', or 'bonferroni_gen'.
Default: \code{fst_correction = NULL}.}

\item{p_value_cutoff}{(double) maximum p-value to keep an Fst measurement.
Also used as base for Bonferroni correction.
Default: \code{p_value_cutoff = 0.05}.}

\item{k}{(logical) Enable kernel-smoothed Pi, Fis, Fst, Fst', and Phi_st calculations.
Default: \code{k = FALSE}.}

\item{smooth_fstats}{(logical) Enable kernel-smoothed Fst, Fst', and Phi_st calculations.
Default: \code{smooth_fstats = FALSE}.}

\item{smooth_popstats}{(logical) Enable kernel-smoothed Pi and Fis calculations.
Default: \code{smooth_popstats = FALSE}.}

\item{sigma}{(integer) Standard deviation of the kernel smoothing weight distribution.
Default: \code{sigma = 150000} (150kb).}

\item{bootstrap}{(logical) Turn on boostrap resampling for all smoothed statistics.
Default: \code{bootstrap = FALSE}.}

\item{N}{(integer) Number of bootstrap resamplings to calculate.
Default: \code{N = 100}.}

\item{bootstrap_pifis}{(logical) Turn on boostrap resampling for smoothed SNP-based Pi and Fis calculations.
Default: \code{bootstrap_pifis = FALSE}.}

\item{bootstrap_fst}{(logical) Turn on boostrap resampling for smoothed Fst calculations based on pairwise population comparison of SNPs.
Default: \code{bootstrap_fst = FALSE}.}

\item{bootstrap_div}{(logical) Turn on boostrap resampling for smoothed haplotype diveristy and gene diversity calculations based on haplotypes.
Default: \code{bootstrap_div = FALSE}.}

\item{bootstrap_phist}{(logical) Turn on boostrap resampling for smoothed Phi_st calculations based on haplotypes.
Default: \code{bootstrap_phist = FALSE}.}

\item{bootstrap_wl}{(character) Path to a whitelist file. Only use bootstrap loci contained in this whitelist.
Default: \code{bootstrap_wl = NULL}.}

\item{ordered_export}{(logical) If data is reference aligned, exports will be ordered; only a single representative of each overlapping site.
Default: \code{ordered_export = FALSE}.}

\item{fasta_samples}{(logical) Output the sequences of the two haplotypes of each (diploid) sample, for each locus, in FASTA format.
Default: \code{fasta_samples = FALSE}.}

\item{fasta_samples_raw}{(logical) Output all haplotypes observed in each sample, for each locus, in FASTA format.
Default: \code{fasta_samples_raw = FALSE}.}

\item{fasta_loci}{(logical) Output consensus sequences of all loci, in FASTA format.
Default: \code{fasta_loci = FALSE}.}

\item{vcf}{(logical) Output SNPs in Variant Call Format (VCF).
Default: \code{vcf = TRUE}.}

\item{genepop}{(logical) Output results in GenePop format.
Default: \code{genepop = FALSE}.}

\item{structure}{(logical) Output results in Structure format.
Default: \code{structure = FALSE}.}

\item{radpainter}{(logical) Output results results in fineRADstructure/RADpainter format.
Default: \code{radpainter = FALSE}.}

\item{phase}{(logical) Output genotypes in PHASE format.
Default: \code{phase = FALSE}.}

\item{fastphase}{(logical) Output genotypes in fastPHASE format.
Default: \code{fastphase = FALSE}.}

\item{plink}{(logical) Output genotypes in PLINK format.
Default: \code{plink = FALSE}.}

\item{hzar}{(logical) Output genotypes in Hybrid Zone Analysis using R (HZAR) format.
Default: \code{hzar = FALSE}.}

\item{phylip}{(logical) Output nucleotides that are fixed-within, and variant among populations in Phylip format for phylogenetic tree construction.
Default: \code{phylip = FALSE}.}

\item{phylip_var}{(logical) Include variable sites in the phylip output encoded using IUPAC notation.
Default: \code{phylip_var = FALSE}.}

\item{phylip_var_all}{(logical) Include all sequence as well as variable sites in the phylip output encoded using IUPAC notation.
Default: \code{phylip_var_all = FALSE}.}

\item{treemix}{(logical) Output SNPs in a format useable for the TreeMix program (Pickrell and Pritchard).
Default: \code{treemix = FALSE}.}

\item{no_hap_exports}{(logical) Omit haplotype outputs.
Default: \code{no_hap_exports = FALSE}.}

\item{h}{Display this help messsage.
Default: \code{h = FALSE}}

\item{verbose}{turn on additional logging.
Default: \code{verbose = FALSE}}

\item{v}{print program version.
Default: \code{v = FALSE}}

\item{log_fst_comp}{(logical) Log components of Fst/Phi_st calculations to a file.
Default: \code{log_fst_comp = FALSE}}
}
\value{
\href{http://catchenlab.life.illinois.edu/stacks/comp/populations.php}{populations}
returns different output depending on arguments selected.
}
\description{
Run \href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/comp/populations.php}{populations}
module inside R!
}
\examples{
\dontrun{
pop <- stackr::run_populations(M = "population.map.tsv")
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.

Guo SW, Thompson EA (1992)
Performing the exact test of Hardy-Weinberg proportion for multiple alleles.
Biometrics, 48, 361-372.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/populations.php}{populations}
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS Version 2.0b}
}
