% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_populations.R
\name{run_populations}
\alias{run_populations}
\title{Run STACKS populations module}
\usage{
run_populations(P = "07_rxstacks_cstacks_sstacks_populations", V = NULL,
  O = NULL, M, t = parallel::detectCores() - 1, b = "guess", s = FALSE,
  p = 1, r = 0.3, min_maf = NULL, max_obs_het = NULL, m = NULL,
  lnl_lim = NULL, write_single_snp = FALSE, write_random_snp = FALSE,
  B = NULL, W = NULL, e = NULL, merge_sites = FALSE,
  merge_prune_lim = NULL, fstats = FALSE, fst_correction = NULL,
  p_value_cutoff = 0.05, k = FALSE, sigma = 150000, bootstrap = FALSE,
  N = 100, bootstrap_pifis = FALSE, bootstrap_fst = FALSE,
  bootstrap_div = FALSE, bootstrap_phist = FALSE, bootstrap_wl = NULL,
  ordered_export = FALSE, genomic = FALSE, fasta = FALSE,
  fasta_strict = FALSE, vcf = TRUE, vcf_haplotypes = TRUE,
  genepop = FALSE, structure = FALSE, phase = FALSE, fastphase = FALSE,
  beagle = FALSE, beagle_phased = FALSE, plink = FALSE, hzar = FALSE,
  phylip = FALSE, phylip_var = FALSE, phylip_var_all = FALSE,
  treemix = FALSE, h = FALSE, verbose = FALSE, v = FALSE,
  log_fst_comp = FALSE)
}
\arguments{
\item{P}{(character) Path to the directory containing STACKS files.
Default: \code{P = "07_rxstacks_cstacks_sstacks_populations"}.
Inside the folder, you should have:
\itemize{
  \item \strong{the catalog files:} starting with \code{batch_} and ending with
  \code{.alleles.tsv.gz, .snps.tsv.gz, .tags.tsv.gz};
  \item \strong{4 files for each samples:} The sample name is the prefix for
  the files ending with:
\code{.alleles.tsv.gz, .models.tsv.gz, .snps.tsv.gz, .tags.tsv.gz}.
Those files are created in the
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks},
\href{http://catchenlab.life.illinois.edu/stacks/comp/sstacks.php}{sstacks} and
\href{http://catchenlab.life.illinois.edu/stacks/comp/rxstacks.php}{rxstacks}
modules.
}}

\item{V}{(character) Path to an input VCF file. When this module is used to
filter an existing vcf file.
Default: \code{V = NULL}.}

\item{O}{(character) Path to a directory where to write the output files.
With default: \code{O = NULL}, the function creates a folder inside
\code{P} with date and time appended to \code{stackr_stacks_populations_}.}

\item{M}{path to a population map file. The format is a tab-separated file,
with first column containing sample name and second column population id.
No heather (column name).
e.g. \code{M = "07_rxstacks_cstacks_sstacks_populations/population.map.catalog.tsv"}}

\item{t}{(integer) enable parallel execution with num_threads threads.
Default: \code{t = parallel::detectCores() - 1}}

\item{b}{(integer) Database/batch ID of the input catalog to consider.
Advice: don't modify the default.
Default: \code{b = "guess"}.}

\item{s}{(logical) Output a file to import results into an SQL database.
Default: \code{s = FALSE}.}

\item{p}{(integer) Minimum number of populations a locus must be present in to process a locus.
Default: \code{p = 1}.}

\item{r}{(double) Minimum percentage of individuals in a population required to process a locus for that population.
Default: \code{r = 0.3}.}

\item{min_maf}{(double) Specify a minimum minor allele frequency required to process a nucleotide site at a locus (0 < min_maf < 0.5).
Default: \code{min_maf = NULL}. Recommendation: use my other R package RADIATOR to filter data based on MAF.}

\item{max_obs_het}{(double) Specify a maximum observed heterozygosity required to process a nucleotide site at a locus.
Default: \code{max_obs_het = NULL}.  Recommendation: use my other R package RADIATOR to filter data based on heterozygosity.}

\item{m}{(integer) Specify a minimum stack depth required for individuals at a locus.
Default: \code{m = NULL}.}

\item{lnl_lim}{(integer) Filter loci with log likelihood values below this threshold.
Default: \code{lnl_lim = NULL}.}

\item{write_single_snp}{(logical) Restrict data analysis to only the first SNP per locus.
Default: \code{write_single_snp = FALSE}.}

\item{write_random_snp}{(logical) Restrict data analysis to one random SNP per locus.
Default: \code{write_random_snp = FALSE}.}

\item{B}{(character) Path to a file containing Blacklisted markers to be excluded from the export.
Default: \code{B = NULL}.}

\item{W}{(character) Path to a file containing Whitelisted markers to include in the export.
Default: \code{W = NULL}.}

\item{e}{(character) Restriction enzyme name.
Default: \code{e = NULL}.}

\item{merge_sites}{(logical) Merge loci that were produced from the same restriction enzyme cutsite (requires reference-aligned data).
Default: \code{merge_sites = FALSE}.}

\item{merge_prune_lim}{(integer) When merging adjacent loci, if at least X% samples posses both loci prune the remaining samples out of the analysis.
Default: \code{merge_prune_lim = NULL}.}

\item{fstats}{(logical) Enable SNP and haplotype-based F statistics.
Default: \code{fstats = FALSE}.}

\item{fst_correction}{(character) Specify a correction to be applied to Fst values: 'p_value', 'bonferroni_win', or 'bonferroni_gen'.
Default: \code{fst_correction = NULL}.}

\item{p_value_cutoff}{(double) maximum p-value to keep an Fst measurement.
Also used as base for Bonferroni correction.
Default: \code{p_value_cutoff = 0.05}.}

\item{k}{(logical) Enable kernel-smoothed Pi, Fis, Fst, Fst', and Phi_st calculations.
Default: \code{k = FALSE}.}

\item{sigma}{(integer) Standard deviation of the kernel smoothing weight distribution.
Default: \code{sigma = 150000} (150kb).}

\item{bootstrap}{(logical) Turn on boostrap resampling for all smoothed statistics.
Default: \code{bootstrap = FALSE}.}

\item{N}{(integer) Number of bootstrap resamplings to calculate.
Default: \code{N = 100}.}

\item{bootstrap_pifis}{(logical) Turn on boostrap resampling for smoothed SNP-based Pi and Fis calculations.
Default: \code{bootstrap_pifis = FALSE}.}

\item{bootstrap_fst}{(logical) Turn on boostrap resampling for smoothed Fst calculations based on pairwise population comparison of SNPs.
Default: \code{bootstrap_fst = FALSE}.}

\item{bootstrap_div}{(logical) Turn on boostrap resampling for smoothed haplotype diveristy and gene diversity calculations based on haplotypes.
Default: \code{bootstrap_div = FALSE}.}

\item{bootstrap_phist}{(logical) Turn on boostrap resampling for smoothed Phi_st calculations based on haplotypes.
Default: \code{bootstrap_phist = FALSE}.}

\item{bootstrap_wl}{(character) Path to a whitelist file. Only use bootstrap loci contained in this whitelist.
Default: \code{bootstrap_wl = NULL}.}

\item{ordered_export}{(logical) If data is reference aligned, exports will be ordered; only a single representative of each overlapping site.
Default: \code{ordered_export = FALSE}.}

\item{genomic}{(logical) Output each nucleotide position (fixed or polymorphic) in all population members to a file (requires restriction enzyme name \code{e}).
Default: \code{genomic = FALSE}.}

\item{fasta}{(logical) Output full sequence for each unique haplotype, from each sample locus in FASTA format, regardless of plausibility.
Default: \code{fasta = FALSE}.}

\item{fasta_strict}{(logical) Output full sequence for each haplotype, from each sample locus in FASTA format, only for biologically plausible loci.
Default: \code{fasta_strict = FALSE}.}

\item{vcf}{(logical) Output SNPs in Variant Call Format (VCF).
Default: \code{vcf = TRUE}.}

\item{vcf_haplotypes}{(logical) Output haplotypes in Variant Call Format (VCF).
Default: \code{vcf_haplotypes = TRUE}.}

\item{genepop}{(logical) Output results in GenePop format.
Default: \code{genepop = FALSE}.}

\item{structure}{(logical) Output results in Structure format.
Default: \code{structure = FALSE}.}

\item{phase}{(logical) Output genotypes in PHASE format.
Default: \code{phase = FALSE}.}

\item{fastphase}{(logical) Output genotypes in fastPHASE format.
Default: \code{fastphase = FALSE}.}

\item{beagle}{(logical) Output genotypes in Beagle format.
Default: \code{beagle = FALSE}.}

\item{beagle_phased}{(logical) Output haplotypes in Beagle format.
Default: \code{beagle_phased = FALSE}.}

\item{plink}{(logical) Output genotypes in PLINK format.
Default: \code{plink = FALSE}.}

\item{hzar}{(logical) Output genotypes in Hybrid Zone Analysis using R (HZAR) format.
Default: \code{hzar = FALSE}.}

\item{phylip}{(logical) Output nucleotides that are fixed-within, and variant among populations in Phylip format for phylogenetic tree construction.
Default: \code{phylip = FALSE}.}

\item{phylip_var}{(logical) Include variable sites in the phylip output encoded using IUPAC notation.
Default: \code{phylip_var = FALSE}.}

\item{phylip_var_all}{(logical) Include all sequence as well as variable sites in the phylip output encoded using IUPAC notation.
Default: \code{phylip_var_all = FALSE}.}

\item{treemix}{(logical) Output SNPs in a format useable for the TreeMix program (Pickrell and Pritchard).
Default: \code{treemix = FALSE}.}

\item{h}{Display this help messsage.
Default: \code{h = FALSE}}

\item{verbose}{turn on additional logging.
Default: \code{verbose = FALSE}}

\item{v}{print program version.
Default: \code{v = FALSE}}

\item{log_fst_comp}{(logical) Log components of Fst/Phi_st calculations to a file.
Default: \code{log_fst_comp = FALSE}}
}
\value{
\href{http://catchenlab.life.illinois.edu/stacks/comp/populations.php}{populations}
returns different output depending on arguments selected.
}
\description{
Run \href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
\href{http://catchenlab.life.illinois.edu/stacks/comp/populations.php}{populations}
module inside R!
}
\examples{
\dontrun{
to do
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/populations.php}{populations}
}
