% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking_weights_crps.R
\name{stack_crps}
\alias{stack_crps}
\title{Model weighting by CRPS}
\usage{
stack_crps(data, lambda = NULL, gamma = NULL, dirichlet_alpha = 1.001)
}
\arguments{
\item{data}{a data.frame with the following entries: 
\itemize{
  \item model (the model used to generate the correspondig predictions)
  \item geography (the region for which predictions are generated)
  \item date (the date of corresponding )
}}

\item{lambda}{weigh timepoints differently}

\item{gamma}{weigh countries differently}

\item{dirichlet_alpha}{prior for the weights}
}
\value{
returns a vector with the model weights
}
\description{
Model weights for stacking by CRPS. MISSING.
}
\examples{

stack_crps(stackr::sample_prepared_data)

#Missing
}
\references{
Missing
}
