% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopensemble-package.R
\docType{package}
\name{lopensemble-package}
\alias{lopensemble-package}
\alias{lopensemble}
\title{The 'lopensemble' package.}
\description{
The lopensemble package provides an easy way to stack models based on
the Continuous Ranked Probability Score (CRPS) based on predictive samples.
Given some training data with true observed values as well as predictions
generated from different models, lopensemble finds the optimal weights to form an
ensemble from these models. Using these weights, lopensemble can then provide
samples from the optimal model mixture by drawing from the predictice samples
of those models in the correct proportion. This gives a true mixture model
solely based on predictive samples and is in this regard superior to other
ensembling techniques like Bayesian Model Averaging.

Currently stacking using CRPS is supported. Weights are generated using
the \link{crps_weights} function. With these weights and predictive
samples, \link{mixture_from_samples} can be used to obtain predictive
samples from the optimal mixture model.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package
  version 2.19.3. https://mc-stan.org

Using Stacking to Average Bayesian Predictive Distributions,
Yuling Yao , Aki Vehtari, Daniel Simpson, and Andrew Gelman, 2018,
Bayesian Analysis 13, Number 3, pp. 917–1003

Strictly Proper Scoring Rules, Prediction,and Estimation,
Tilmann Gneiting and Adrian E. Raftery, 2007, Journal of the American
Statistical Association, Volume 102, 2007 - Issue 477
}
\author{
\strong{Maintainer}: Nikos Bosse \email{nikosbosse@gmail.com} (\href{https://orcid.org/0000-0002-7750-5280}{ORCID})

Authors:
\itemize{
  \item Yuling Yao \email{yy2619@columbia.edu}
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID})
}

}
\keyword{internal}
