% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackR-package.R
\docType{package}
\name{stackR-package}
\alias{stackR}
\alias{stackR-package}
\title{stackR: Functions to Explore, Analyze, and Visualize Open Data from stackexchange.com}
\description{
Functions to explore, analyze, and visualize open data about
    the question answering platform run by
    [stackexchange.com](https://stackexchange.com).
}
\author{
\strong{Maintainer}: Jim Tyhurst \email{jim@tyhurst.com}

}
\keyword{internal}
