% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackr_imputations_module.R
\name{stackr_imputations_module}
\alias{stackr_imputations_module}
\title{Map-independent imputation of missing genotype}
\usage{
stackr_imputations_module(data, imputation.method = NULL,
  impute = "genotype", imputations.group = "populations", num.tree = 100,
  iteration.rf = 10, split.number = 100, verbose = FALSE,
  parallel.core = parallel::detectCores() - 1, filename = NULL, ...)
}
\arguments{
\item{data}{A file in the working directory or object in the global environment 
in wide or long (tidy) formats. To import, the function uses 
\href{https://github.com/thierrygosselin/stackr}{stackr} 
\code{\link[stackr]{tidy_wide}}. See details for more info.}

\item{imputation.method}{(character, optional) 
Methods available for map-independent imputations of missing genotype: 
(1) \code{"max"} to use the most frequent category for imputations.
(2) \code{"rf"} using Random Forest algorithm. 
Default: no imputation \code{imputation.method = NULL}.}

\item{impute}{(character, optional) Imputation on missing genotype 
\code{impute = "genotype"} or alleles \code{impute = "allele"}.
Default: \code{"genotype"}.}

\item{imputations.group}{(character, optional) \code{"global"} or \code{"populations"}.
Should the imputations be computed globally or by populations. If you choose
global, turn the verbose to \code{TRUE}, to see progress.
Default = \code{"populations"}.}

\item{num.tree}{(integer, optional) The number of trees to grow in Random Forest. 
Default: \code{num.tree = 100}.}

\item{iteration.rf}{(integer, optional) The number of iterations of missing data algorithm
in Random Forest. 
Default: \code{iteration.rf = 10}.}

\item{split.number}{(integer, optional) Non-negative integer value used to specify
random splitting in Random Forest. 
Default: \code{split.number = 100}.}

\item{verbose}{(logical, optional) Should trace output be enabled on each iteration
in Random Forest ? 
Default: \code{verbose = FALSE}.}

\item{parallel.core}{(optional) The number of core for OpenMP shared-memory parallel
programming of Random Forest imputations. For more info on how to install the
OpenMP version see \code{\link[randomForestSRC]{randomForestSRC-package}}.
If not selected \code{detectCores()-1} is used as default.}

\item{filename}{(optional) The file name for the \strong{imputed} 
tidy data frame written to the working directory.
With missing argument or default: \code{filename = NULL}, the imputed 
tidy data is in the global environment only 
(i.e. not written in the working directory).}

\item{...}{other parameters passed to the function.}
}
\value{
The output in your global environment is the imputed tidy data frame.
If \code{filename} is provided, the imputed tidy data frame is also 
written to the working directory.
}
\description{
Map-independent imputation of missing genotype using Random Forest 
or the most frequent category. Impute genotypes or alleles. Used internally 
in \href{https://github.com/thierrygosselin/assigner}{assigner} and
\href{https://github.com/thierrygosselin/stackr}{stackr} and 
might be of interest for users. 

The goal of \code{missing_visualization} is to provide a simple solution for a complicated
problem: missing observations in genotype datasets. Before running this function
to populate the original dataset with synthetic data you should try 
\code{\link[stackr]{missing_visualization}} to detect patterns of missingness.
Follow the vignette for more info.
}
\examples{
\dontrun{
# The simplest form:
wolf.imputed.data <- stackr_imputations_module(data = "wolf.tidy.dataset.tsv")
This will impute missing genotype by populations using random forest.
The number of tree, iterations.rf, split.number and parallel.core used 
will be the default.
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Ishwaran H. and Kogalur U.B. (2015). Random Forests for Survival,
 Regression and Classification (RF-SRC), R package version 1.6.1.

Ishwaran H. and Kogalur U.B. (2007). Random survival forests
for R. R News 7(2), 25-31.

Ishwaran H., Kogalur U.B., Blackstone E.H. and Lauer M.S. (2008).
Random survival forests. Ann. Appl. Statist. 2(3), 841-860.
}

