% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_reads.R
\name{stackr_normalize}
\alias{stackr_normalize}
\title{stackr_normalize}
\usage{
stackr_normalize(
  fq.to.normalize,
  count.reads = FALSE,
  number.replicates,
  sample.reads,
  parallel.core = parallel::detectCores() - 1
)
}
\description{
main normalize function
}
\keyword{internal}
