% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_samples.R
\name{stackr_normalize}
\alias{stackr_normalize}
\title{stackr_normalize}
\usage{
stackr_normalize(x, fastq.files, fastq.files.short, project.info,
  number.replicates, sample.reads, parallel.core = parallel::detectCores() -
  1)
}
\description{
main normalize function
}
\keyword{internal}
