% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_haplotypes.R
\name{summary_haplotypes}
\alias{summary_haplotypes}
\title{Haplotypes file summary}
\usage{
summary_haplotypes(data, strata, read.length, whitelist.markers = NULL,
  blacklist.id = NULL, keep.consensus = TRUE, pop.levels = NULL,
  pop.labels = NULL, artifact.only = TRUE,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{strata}{A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks `population map file`, make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).}

\item{read.length}{(number) The length in nucleotide of your reads
(e.g. \code{read.length = 100}).}

\item{whitelist.markers}{(optional) A whitelist of loci with a column name
'LOCUS'. The whitelist is located in the global environment or in the
directory (e.g. "whitelist.txt").
If a whitelist is used, the files written to the directory will have
'.filtered' in the filename.
Default: \code{whitelist.markers = NULL}.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the global environment
or in the directory (with "blacklist.txt").
Default: \code{blacklist.id = NULL}.}

\item{keep.consensus}{(optional) Using whitelist of markers can automatically
remove consensus markers from the nucleotide diversity (Pi) calculations.
If you know what you are doing, play with this argument
to keep consensus markers.
Default: \code{keep.consensus = TRUE}.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.levels = c("NEW", "NEW", "ALB")}.
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated,
there is also the \code{strata} argument that can do the same thing,
see below.
White spaces in population names are replaced by underscore.}

\item{artifact.only}{(optional, logical) With default \code{artifact.only = TRUE},
the function will stop after generating blacklists of artifacts and consensus loci
(no pi, no FH measures). This is faster and useful to run at the beginning of the pipeline.
Run again at the end, with blacklists and whitelists to get reliable estimates of
gene diverisity, IBDG and FH.}

\item{parallel.core}{(optional) The number of core for parallel
programming during Pi calculations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
The function returns a list with:
\enumerate{
\item $consensus.pop # if consensus reads are found
\item $consensus.loci # if consensus reads are found
\item $artifacts.pop # if artifacts are found
\item $artifacts.loci # if artifacts are found
\item $individual.summary: the individual's info for FH and Pi
\item $summary: the summary statistics per populations and averaged over all pops.
}
Also in the list 3 plots (also written in the folder):
\enumerate{
\item $scatter.plot.Pi.Fh.ind: showing Pi and FH per individuals
\item $scatter.plot.Pi.Fh.pop: showing Pi and FH per pop
\item $boxplot.pi: showing the boxplot of Pi per pop
\item $boxplot.fh: showing the boxplot of FH per pop
}
use $ to access each #' objects in the list.
The function potentially write 4 files in the working directory:
blacklist of unique loci with assembly artifacts and/or sequencing errors
(per individuals and globally), a blacklist of unique consensus loci
and a summary of the haplotype file by population.
}
\description{
STACKS batch_x.haplotypes.tsv file summary.
The output of the function is a summary table for populations with:
\enumerate{
\item assembly artifacts and/or sequencing errors: loci with > 2 alleles.
Genotypes with more than 2 alleles are erased before estimating the subsequent statistics.
\item consensus reads (reads with no variation/polymorphism throughout the dataset).
\item monomorphic loci (at the population and overall level)
\item polymorphic loci (at the population and overall level)
\item haplotypes statistics for the observed and expected homozygosity and
heterozygosity (HOM_O, HOM_E, HET_O, HET_E).
\item Gene Diversity within populations (Hs) and averaged over all populations.
Not to confuse with Expected Heterozygosity(HET_E).
Here, Hs statistic includes a correction for sampling bias stemming from
sampling a limited number of individuals per population (Nei, 1987).
\item Wright’s inbreeding coefficient (Fis)
\item Nei's inbreeding coefficient (Gis) that include a correction for sampling bias.
\item \strong{IBDG}: a proxy measure of the realized proportion of the genome
that is identical by descent
\item \strong{FH measure}: an individual-base estimate that is
based on the excess in the observed number of homozygous
genotypes within an individual relative to the mean number of homozygous
genotypes expected under random mating (Keller et al., 2011; Kardos et al., 2015).
\item \strong{Pi}: the nucleotide diversity measured here consider the
consensus reads in the catalog (no variation between population sequences).
}
}
\examples{
\dontrun{
# The simplest way to run the function:
sum <- summary_haplotypes(
data = "batch_1.haplotypes.tsv",
strata = "strata_brook_charr.tsv",
read.length = 90)
}
}
\references{
Keller MC, Visscher PM, Goddard ME (2011)
Quantification of inbreeding due to distant ancestors and its detection
 using dense single nucleotide polymorphism data. Genetics, 189, 237–249.

Kardos M, Luikart G, Allendorf FW (2015)
Measuring individual inbreeding in the age of genomics: marker-based
measures are better than pedigrees. Heredity, 115, 63–72.

Nei M, Li WH (1979)
Mathematical model for studying genetic variation in terms of
restriction endonucleases.
Proceedings of the National Academy of Sciences of
the United States of America, 76, 5269–5273.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com} and
Anne-Laure Ferchaud \email{annelaureferchaud@gmail.com}
}
