% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ustacks.R
\name{summary_ustacks}
\alias{summary_ustacks}
\title{Summarize STACKS ustacks files}
\usage{
summary_ustacks(ustacks.folder, parallel.core = parallel::detectCores() - 1,
  verbose = FALSE)
}
\arguments{
\item{ustacks.folder}{(logical). Default: \code{mismatch.testing = FALSE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}

\item{verbose}{(optional) Make the function a little more chatty during
execution.
Default: \code{verbose = FALSE}.}
}
\value{
The function returns a summary (data frame) containing:
\enumerate{
\item INDIVIDUALS: the sample id
\item LOCUS_NUMBER: the number of locus
\item BLACKLIST_USTACKS: the number of locus blacklisted by ustacks
\item FOR_CATALOG: the number of locus available to generate the catalog
\item BLACKLIST_ARTIFACT: the number of artifact genotypes (> 2 alleles, see details)
\item FILTERED: the number of locus after artifacts are removed
\item HOMOZYGOSITY: the number of homozygous genotypes
\item HETEROZYGOSITY: the number of heterozygous genotypes
\item MEAN_NUMBER_SNP_LOCUS: the mean number of SNP/locus (excluding the artifact locus)
\item MAX_NUMBER_SNP_LOCUS: the max number of SNP/locus observed for this individual (excluding the artifact locus)
\item NUMBER_LOCUS_4SNP: the number of locus with 4 or more SNP/locus (excluding the artifact locus)
}
}
\description{
This function reads inside the output of STACKS ustasks folder
and summarise the information.
}
\examples{
\dontrun{
ustacks.summary <- stackr::summary_ustacks(
ustacks.folder = "06_ustacks_cstacks_sstacks")

# To get the number of snp/locus for a specific individual:
snp.info <- ustacks.summary \%>\%
dplyr::filter(INDIVIDUALS == "ID2") \%>\%
dplyr::select(INDIVIDUALS, SNP_LOCUS) \%>\%
tidyr::unnest(.)

#Similarly, for the blacklisted locus (artifactual):
blacklist <- ustacks.summary \%>\%
dplyr::filter(INDIVIDUALS == "ID2") \%>\%
dplyr::select(INDIVIDUALS, BLACKLIST_ARTIFACT) \%>\%
tidyr::unnest(.)
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.
}
\seealso{
\href{http://catchenlab.life.illinois.edu/stacks/comp/ustacks.php}{ustacks}
}
