% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stacks_modification.R
\name{sumstats_prep}
\alias{sumstats_prep}
\title{Import and prepare for analysis a batch_x.sumstats.tsv file produced by STACKS.}
\usage{
sumstats_prep(sumstats, skip.line, pop.num, pop.col.types, pop.integer.equi,
  pop.levels)
}
\arguments{
\item{sumstats}{The STACKS batch_x.susmtats.tsv file.}

\item{skip.line}{The number of line without the header to start reading the data}

\item{pop.num}{The number of populations.}

\item{pop.col.types}{Integer or Character used in STACKS populations module?}

\item{pop.integer.equi}{When Integer was used for your population id, give the character equivalence}

\item{pop.levels}{A character string with your populations in order.}
}
\description{
Import STACKS sumstats.tsv and add useful info (frequency of REF allele, Global MAF) for subsequent use. to a tidy format.
}

