% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_imputation.R
\name{vcf_imputation}
\alias{vcf_imputation}
\title{VCF data imputation}
\usage{
vcf_imputation(vcf.file, whitelist.loci = NULL, snp.LD = NULL,
  common.markers = FALSE, blacklist.id = NULL, blacklist.genotypes = NULL,
  pop.id.start, pop.id.end, filename = "batch_1.vcf.imputed.vcf",
  imputations = "rf", imputations.group = "populations", num.tree = 100,
  iteration.rf = 10, split.number = 100, verbose = FALSE,
  parallel.core = 2)
}
\arguments{
\item{vcf.file}{The VCF file created by STACKS.}

\item{whitelist.loci}{(optional) A whitelist of loci and 
a column header 'LOCUS'.
The whitelist is in the directory (e.g. "whitelist.txt").}

\item{snp.LD}{(optional) Minimize linkage disequilibrium (LD) by choosing 
among these 3 options: \code{"random"} selection, \code{"first"} or 
\code{"last"} SNP on the same read/haplotype. Default = \code{NULL}.}

\item{common.markers}{(optional) Logical. Default = \code{FALSE}. 
With \code{TRUE}, will keep markers present in all the populations.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is in the directory
 (e.g. "blacklist.txt").}

\item{blacklist.genotypes}{A blacklist of genotypes 
containing 3 columns header 'LOCUS', 'POS' and 'INDIVIDUALS'.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{filename}{The name of the file written to the directory.
Use the extension ".vcf" at the end. Default \code{batch_1.vcf.imputed.vcf}.}

\item{imputations}{For map-independent imputations available options are:
(1) \code{"max"} to use the most frequent category for imputations.
 (2) \code{"rf"} using Random Forest algorithm. Default = \code{"rf"}.}

\item{imputations.group}{\code{"global"} or \code{"populations"}.
Should the imputations be computed globally or by populations. If you choose
global, turn the verbose to \code{TRUE}, to see progress.
Default = \code{"populations"}.}

\item{num.tree}{The number of trees to grow in Random Forest. Default is 100.}

\item{iteration.rf}{The number of iterations of missing data algorithm 
in Random Forest. Default is 10.}

\item{split.number}{Non-negative integer value used to specify 
random splitting in Random Forest. Default is 100.}

\item{verbose}{Logical. Should trace output be enabled on each iteration 
in Random Forest ? Default is \code{FALSE}.}

\item{parallel.core}{(optional) The number of core for OpenMP 
shared-memory parallel programming of Random Forest imputations. 
For more info on how to install the
OpenMP version see \code{\link[randomForestSRC]{randomForestSRC-package}}.
If not selected \code{detectCores()-1} is used as default.}
}
\value{
When no imputation is selected ....
}
\description{
Map-independent imputation of a VCF using Random Forest or the 
most frequent category.
}
\details{
The imputations using Random Forest requires more time to compute 
and can take several minutes and hours depending on the size of 
the dataset and polymorphism of the species used. e.g. with a 
low polymorphic taxa, and a data set containing 30\% missing data, 
5 000 haplotypes loci and 500 individuals will require 15 min.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011) 
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences. 
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013) 
Stacks: an analysis tool set for population genomics. 
Molecular Ecology, 22, 3124-3140.

Ishwaran H. and Kogalur U.B. (2015). Random Forests for Survival,
 Regression and Classification (RF-SRC), R package version 1.6.1.

Ishwaran H. and Kogalur U.B. (2007). Random survival forests
for R. R News 7(2), 25-31.

Ishwaran H., Kogalur U.B., Blackstone E.H. and Lauer M.S. (2008).
Random survival forests. Ann. Appl. Statist. 2(3), 841--860.
}
\seealso{
\code{randomForestSRC} is available on 
CRAN \url{http://cran.r-project.org/web/packages/randomForestSRC/} 
and github \url{https://github.com/ehrlinger/randomForestSRC}
}

