% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitelist_blacklist.R
\name{whitelist_loci_vcf}
\alias{whitelist_loci_vcf}
\title{Whitelist loci for VCF tools
This function creates a whitelist of loci, used after applying a filter
to keep track of the loci kept by a filter.}
\usage{
whitelist_loci_vcf(data, filename)
}
\arguments{
\item{data}{The data frame after the filter. Object
or file (using ".tsv") of class sumstats.}

\item{filename}{The name of the file written in the directory.}
}
\description{
Whitelist loci for VCF tools
This function creates a whitelist of loci, used after applying a filter
to keep track of the loci kept by a filter.
}

