% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_genepop.R
\name{write_genepop}
\alias{write_genepop}
\title{Used internally in stackr to write a genepop file from a tidy data frame}
\usage{
write_genepop(data, sep = NULL, pop.levels = NULL,
  genepop.header = "my firt genepop", markers.line.format = "line",
  filename = "stackr_genepop.gen", ...)
}
\arguments{
\item{data}{A file or object in the global environment containing at least 
these 4 columns: 
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping of individuals.), 
\code{MARKERS} and \code{GENOTYPES} in a tidy format. During import, 
only those columns names will be kept.}

\item{sep}{(optional) A character string separating alleles. 
Default: \code{sep = NULL}.}

\item{pop.levels}{(optional, string) A character string with your populations ordered.
Default: \code{pop.levels = NULL}.}

\item{genepop.header}{The first line of the Genepop file.
Default: \code{genepop.header = "my firt genepop"}.}

\item{markers.line.format}{(optional, character) You can write the markers 
on a single line separated by commas \code{markers.line.format = "line"}, 
or have markers on a separate line, i.e. in one column, 
\code{markers.line.format = "column"} 
(not very useful with thousands of markers). 
Default: \code{markers.line.format = "line"}.}

\item{filename}{The name of the file written to the working directory.
Use the extension ".gen" at the end. 
Default: \code{filename = "stackr_genepop.gen"}.}

\item{...}{other parameters passed to the function.}
}
\value{
A genepop file is saved to the working directory.
}
\description{
Write a genepop file from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\details{
Details for the sep argument
This character is directly used in regular expressions using strigi. 
Some characters need to be preceeded by double backslashes \code{\\}. 
For instance, "/" works but "|" must be coded as "\\|".
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

