% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_genepop.R
\name{write_genepop}
\alias{write_genepop}
\title{write_genepop.}
\usage{
write_genepop(haplotypes.file, whitelist.loci = NULL, blacklist.id = NULL,
  genepop.filename, genepop.header, pop.levels, pop.id.start, pop.id.end)
}
\arguments{
\item{haplotypes.file}{The 'batch_x.haplotypes.tsv' created by STACKS.}

\item{whitelist.loci}{A whitelist of loci and a column header 'LOCUS'.}

\item{blacklist.id}{A blacklist with individual ID and
a column header 'SAMPLES'.}

\item{genepop.filename}{The name of the file written to the directory.}

\item{genepop.header}{The first line of the Genepop file
(ex: genepop of 10 pop of Acipenser fulvescens filtered MAF 0.05).}

\item{pop.levels}{An optional character string with your populations ordered.}

\item{pop.id.start}{The start of your population id
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id
in the name of your individual sample.}
}
\description{
This function can first filter the haplotypes file with a whitelist of loci
and a blacklist of individuals (optional).
}

