% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtypes.R
\name{write_gtypes}
\alias{write_gtypes}
\title{Write a \code{\link[strataG]{gtypes}} object from a tidy data frame}
\usage{
write_gtypes(data)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{stackr} \code{\link{tidy_genomic_data}}.}
}
\value{
An object of the class \code{\link[strataG]{gtypes}} is returned.
}
\description{
Write a \code{\link[strataG]{gtypes}} object from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and might be of interest for users.
}
\references{
Eric Archer, Paula Adams and Brita Schneiders (2016). 
strataG: Summaries and Population Structure Analyses of
Genetic Data. R package version 1.0.5. https://CRAN.R-project.org/package=strataG
}
\seealso{
\code{strataG.devel} is available on github \url{https://github.com/EricArcher/}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
