% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_samples.R
\name{write_normalize}
\alias{write_normalize}
\title{write_normalize}
\usage{
write_normalize(number.replicates, subsample.random, fastq.files, sample.reads)
}
\description{
Write the normalize fastq file
}
\keyword{internal}
