% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_reads.R
\name{write_normalize}
\alias{write_normalize}
\title{write_normalize}
\usage{
write_normalize(
  number.replicates,
  subsample.random,
  fq.to.normalize,
  sample.reads
)
}
\description{
Write the normalize fastq file
}
\keyword{internal}
