% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Write a vcf file from a tidy data frame}
\usage{
write_vcf(data, pop.info = FALSE, filename = NULL)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{stackr} \code{\link{tidy_genomic_data}}.}

\item{pop.info}{(optional, logical) Should the population information be 
included in the FORMAT field (along the GT info for each samples ?). To make
the VCF population-ready use \code{pop.info = TRUE}. The populatio information
must be included in the \code{POP_ID} column of the tidy dataset.
Default: \code{pop.info = FALSE}.}

\item{filename}{(optional) The file name prefix for the vcf file 
written to the working directory. With default: \code{filename = NULL}, 
the date and time is appended to \code{stackr_vcf_file_}.}
}
\description{
Write a vcf file from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/stackr}{stackr} 
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}

